% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_associate_configuration_items_to_application}
\alias{applicationdiscoveryservice_associate_configuration_items_to_application}
\title{Associates one or more configuration items with an application}
\usage{
applicationdiscoveryservice_associate_configuration_items_to_application(
  applicationConfigurationId, configurationIds)
}
\arguments{
\item{applicationConfigurationId}{[required] The configuration ID of an application with which items are to be
associated.}

\item{configurationIds}{[required] The ID of each configuration item to be associated with an application.}
}
\value{
An empty list.
}
\description{
Associates one or more configuration items with an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_configuration_items_to_application(
  applicationConfigurationId = "string",
  configurationIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
