% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_delete_tags}
\alias{applicationdiscoveryservice_delete_tags}
\title{Deletes the association between configuration items and one or more tags}
\usage{
applicationdiscoveryservice_delete_tags(configurationIds, tags)
}
\arguments{
\item{configurationIds}{[required] A list of configuration items with tags that you want to delete.}

\item{tags}{Tags that you want to delete from one or more configuration items.
Specify the tags that you want to delete in a \emph{key}-\emph{value} format. For
example:

\verb{\{"key": "serverType", "value": "webServer"\}}}
}
\value{
An empty list.
}
\description{
Deletes the association between configuration items and one or more
tags. This API accepts a list of multiple configuration items.
}
\section{Request syntax}{
\preformatted{svc$delete_tags(
  configurationIds = list(
    "string"
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
