% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_describe_continuous_exports}
\alias{applicationdiscoveryservice_describe_continuous_exports}
\title{Lists exports as specified by ID}
\usage{
applicationdiscoveryservice_describe_continuous_exports(exportIds,
  maxResults, nextToken)
}
\arguments{
\item{exportIds}{The unique IDs assigned to the exports.}

\item{maxResults}{A number between 1 and 100 specifying the maximum number of continuous
export descriptions returned.}

\item{nextToken}{The token from the previous call to
\code{\link[=applicationdiscoveryservice_describe_export_tasks]{describe_export_tasks}}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  descriptions = list(
    list(
      exportId = "string",
      status = "START_IN_PROGRESS"|"START_FAILED"|"ACTIVE"|"ERROR"|"STOP_IN_PROGRESS"|"STOP_FAILED"|"INACTIVE",
      statusDetail = "string",
      s3Bucket = "string",
      startTime = as.POSIXct(
        "2015-01-01"
      ),
      stopTime = as.POSIXct(
        "2015-01-01"
      ),
      dataSource = "AGENT",
      schemaStorageConfig = list(
        "string"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists exports as specified by ID. All continuous exports associated with
your user account can be listed if you call
\code{\link[=applicationdiscoveryservice_describe_continuous_exports]{describe_continuous_exports}}
as is without passing any parameters.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_continuous_exports(
  exportIds = list(
    "string"
  ),
  maxResults = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
