% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_describe_export_configurations}
\alias{applicationdiscoveryservice_describe_export_configurations}
\title{DescribeExportConfigurations is deprecated}
\usage{
applicationdiscoveryservice_describe_export_configurations(exportIds,
  maxResults, nextToken)
}
\arguments{
\item{exportIds}{A list of continuous export IDs to search for.}

\item{maxResults}{A number between 1 and 100 specifying the maximum number of continuous
export descriptions returned.}

\item{nextToken}{The token from the previous call to describe-export-tasks.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  exportsInfo = list(
    list(
      exportId = "string",
      exportStatus = "FAILED"|"SUCCEEDED"|"IN_PROGRESS",
      statusMessage = "string",
      configurationsDownloadUrl = "string",
      exportRequestTime = as.POSIXct(
        "2015-01-01"
      ),
      isTruncated = TRUE|FALSE,
      requestedStartTime = as.POSIXct(
        "2015-01-01"
      ),
      requestedEndTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
\code{\link[=applicationdiscoveryservice_describe_export_configurations]{describe_export_configurations}}
is deprecated. Use
\code{\link[=applicationdiscoveryservice_describe_import_tasks]{describe_import_tasks}},
instead.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_export_configurations(
  exportIds = list(
    "string"
  ),
  maxResults = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
