% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_describe_export_tasks}
\alias{applicationdiscoveryservice_describe_export_tasks}
\title{Retrieve status of one or more export tasks}
\usage{
applicationdiscoveryservice_describe_export_tasks(exportIds, filters,
  maxResults, nextToken)
}
\arguments{
\item{exportIds}{One or more unique identifiers used to query the status of an export
request.}

\item{filters}{One or more filters.
\itemize{
\item \code{AgentId} - ID of the agent whose collected data will be exported
}}

\item{maxResults}{The maximum number of volume results returned by
\code{\link[=applicationdiscoveryservice_describe_export_tasks]{describe_export_tasks}}
in paginated output. When this parameter is used,
\code{\link[=applicationdiscoveryservice_describe_export_tasks]{describe_export_tasks}}
only returns \code{maxResults} results in a single page along with a
\code{nextToken} response element.}

\item{nextToken}{The \code{nextToken} value returned from a previous paginated
\code{\link[=applicationdiscoveryservice_describe_export_tasks]{describe_export_tasks}}
request where \code{maxResults} was used and the results exceeded the value
of that parameter. Pagination continues from the end of the previous
results that returned the \code{nextToken} value. This value is null when
there are no more results to return.}
}
\value{
A list with the following syntax:\preformatted{list(
  exportsInfo = list(
    list(
      exportId = "string",
      exportStatus = "FAILED"|"SUCCEEDED"|"IN_PROGRESS",
      statusMessage = "string",
      configurationsDownloadUrl = "string",
      exportRequestTime = as.POSIXct(
        "2015-01-01"
      ),
      isTruncated = TRUE|FALSE,
      requestedStartTime = as.POSIXct(
        "2015-01-01"
      ),
      requestedEndTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Retrieve status of one or more export tasks. You can retrieve the status
of up to 100 export tasks.
}
\section{Request syntax}{
\preformatted{svc$describe_export_tasks(
  exportIds = list(
    "string"
  ),
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      ),
      condition = "string"
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
