% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_describe_import_tasks}
\alias{applicationdiscoveryservice_describe_import_tasks}
\title{Returns an array of import tasks for your account, including status
information, times, IDs, the Amazon S3 Object URL for the import file,
and more}
\usage{
applicationdiscoveryservice_describe_import_tasks(filters, maxResults,
  nextToken)
}
\arguments{
\item{filters}{An array of name-value pairs that you provide to filter the results for
the \code{DescribeImportTask} request to a specific subset of results.
Currently, wildcard values aren't supported for filters.}

\item{maxResults}{The maximum number of results that you want this request to return, up
to 100.}

\item{nextToken}{The token to request a specific page of results.}
}
\value{
A list with the following syntax:\preformatted{list(
  nextToken = "string",
  tasks = list(
    list(
      importTaskId = "string",
      clientRequestToken = "string",
      name = "string",
      importUrl = "string",
      status = "IMPORT_IN_PROGRESS"|"IMPORT_COMPLETE"|"IMPORT_COMPLETE_WITH_ERRORS"|"IMPORT_FAILED"|"IMPORT_FAILED_SERVER_LIMIT_EXCEEDED"|"IMPORT_FAILED_RECORD_LIMIT_EXCEEDED"|"DELETE_IN_PROGRESS"|"DELETE_COMPLETE"|"DELETE_FAILED"|"DELETE_FAILED_LIMIT_EXCEEDED"|"INTERNAL_ERROR",
      importRequestTime = as.POSIXct(
        "2015-01-01"
      ),
      importCompletionTime = as.POSIXct(
        "2015-01-01"
      ),
      importDeletedTime = as.POSIXct(
        "2015-01-01"
      ),
      serverImportSuccess = 123,
      serverImportFailure = 123,
      applicationImportSuccess = 123,
      applicationImportFailure = 123,
      errorsAndFailedEntriesZip = "string"
    )
  )
)
}
}
\description{
Returns an array of import tasks for your account, including status
information, times, IDs, the Amazon S3 Object URL for the import file,
and more.
}
\section{Request syntax}{
\preformatted{svc$describe_import_tasks(
  filters = list(
    list(
      name = "IMPORT_TASK_ID"|"STATUS"|"NAME",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
