% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_describe_tags}
\alias{applicationdiscoveryservice_describe_tags}
\title{Retrieves a list of configuration items that have tags as specified by
the key-value pairs, name and value, passed to the optional parameter
filters}
\usage{
applicationdiscoveryservice_describe_tags(filters, maxResults,
  nextToken)
}
\arguments{
\item{filters}{You can filter the list using a \emph{key}-\emph{value} format. You can separate
these items by using logical operators. Allowed filters include
\code{tagKey}, \code{tagValue}, and \code{configurationId}.}

\item{maxResults}{The total number of items to return in a single page of output. The
maximum value is 100.}

\item{nextToken}{A token to start the list. Use this token to get the next set of
results.}
}
\value{
A list with the following syntax:\preformatted{list(
  tags = list(
    list(
      configurationType = "SERVER"|"PROCESS"|"CONNECTION"|"APPLICATION",
      configurationId = "string",
      key = "string",
      value = "string",
      timeOfCreation = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Retrieves a list of configuration items that have tags as specified by
the key-value pairs, name and value, passed to the optional parameter
\code{filters}.

There are three valid tag filter names:
\itemize{
\item tagKey
\item tagValue
\item configurationId
}

Also, all configuration items associated with your user account that
have tags can be listed if you call
\code{\link[=applicationdiscoveryservice_describe_tags]{describe_tags}} as is
without passing any parameters.
}
\section{Request syntax}{
\preformatted{svc$describe_tags(
  filters = list(
    list(
      name = "string",
      values = list(
        "string"
      )
    )
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
