% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_export_configurations}
\alias{applicationdiscoveryservice_export_configurations}
\title{Deprecated}
\usage{
applicationdiscoveryservice_export_configurations()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  exportId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deprecated. Use
\code{\link[=applicationdiscoveryservice_start_export_task]{start_export_task}}
instead.

Exports all discovered configuration data to an Amazon S3 bucket or an
application that enables you to view and evaluate the data. Data
includes tags and tag associations, processes, connections, servers, and
system performance. This API returns an export ID that you can query
using the \emph{DescribeExportConfigurations} API. The system imposes a limit
of two configuration exports in six hours.
}
\keyword{internal}
