% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_get_discovery_summary}
\alias{applicationdiscoveryservice_get_discovery_summary}
\title{Retrieves a short summary of discovered assets}
\usage{
applicationdiscoveryservice_get_discovery_summary()
}
\description{
Retrieves a short summary of discovered assets.

This API operation takes no request parameters and is called as is at
the command prompt as shown in the example.
}
\section{Request syntax}{
\preformatted{svc$get_discovery_summary()
}
}

\keyword{internal}
