% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_start_data_collection_by_agent_ids}
\alias{applicationdiscoveryservice_start_data_collection_by_agent_ids}
\title{Instructs the specified agents or connectors to start collecting data}
\usage{
applicationdiscoveryservice_start_data_collection_by_agent_ids(agentIds)
}
\arguments{
\item{agentIds}{[required] The IDs of the agents or connectors from which to start collecting data.
If you send a request to an agent/connector ID that you do not have
permission to contact, according to your AWS account, the service does
not throw an exception. Instead, it returns the error in the
\emph{Description} field. If you send a request to multiple agents/connectors
and you do not have permission to contact some of those
agents/connectors, the system does not throw an exception. Instead, the
system shows \code{Failed} in the \emph{Description} field.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  agentsConfigurationStatus = list(
    list(
      agentId = "string",
      operationSucceeded = TRUE|FALSE,
      description = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Instructs the specified agents or connectors to start collecting data.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_data_collection_by_agent_ids(
  agentIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
