% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_start_import_task}
\alias{applicationdiscoveryservice_start_import_task}
\title{Starts an import task, which allows you to import details of your
on-premises environment directly into AWS Migration Hub without having
to use the Application Discovery Service (ADS) tools such as the
Discovery Connector or Discovery Agent}
\usage{
applicationdiscoveryservice_start_import_task(clientRequestToken, name,
  importUrl)
}
\arguments{
\item{clientRequestToken}{Optional. A unique token that you can provide to prevent the same import
request from occurring more than once. If you don\'t provide a token, a
token is automatically generated.

Sending more than one \code{StartImportTask} request with the same client
request token will return information about the original import task
with that client request token.}

\item{name}{[required] A descriptive name for this request. You can use this name to filter
future requests related to this import task, such as identifying
applications and servers that were included in this import task. We
recommend that you use a meaningful name for each import task.}

\item{importUrl}{[required] The URL for your import file that you\'ve uploaded to Amazon S3.

If you\'re using the AWS CLI, this URL is structured as follows:
\verb{s3://BucketName/ImportFileName.CSV}}
}
\description{
Starts an import task, which allows you to import details of your
on-premises environment directly into AWS Migration Hub without having
to use the Application Discovery Service (ADS) tools such as the
Discovery Connector or Discovery Agent. This gives you the option to
perform migration assessment and planning directly from your imported
data, including the ability to group your devices as applications and
track their migration status.
}
\details{
To start an import request, do this:
\enumerate{
\item Download the specially formatted comma separated value (CSV) import
template, which you can find here:
\url{https://s3-us-west-2.amazonaws.com/templates-7cffcf56-bd96-4b1c-b45b-a5b42f282e46/import_template.csv}.
\item Fill out the template with your server and application data.
\item Upload your import file to an Amazon S3 bucket, and make a note of
it\'s Object URL. Your import file must be in the CSV format.
\item Use the console or the \code{StartImportTask} command with the AWS CLI or
one of the AWS SDKs to import the records from your file.
}

For more information, including step-by-step procedures, see \href{https://docs.aws.amazon.com/application-discovery/latest/userguide/discovery-import.html}{Migration Hub Import}
in the \emph{AWS Application Discovery Service User Guide}.

There are limits to the number of import tasks you can create (and
delete) in an AWS account. For more information, see \href{https://docs.aws.amazon.com/application-discovery/latest/userguide/ads_service_limits.html}{AWS Application Discovery Service Limits}
in the \emph{AWS Application Discovery Service User Guide}.
}
\section{Request syntax}{
\preformatted{svc$start_import_task(
  clientRequestToken = "string",
  name = "string",
  importUrl = "string"
)
}
}

\keyword{internal}
