% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_stop_continuous_export}
\alias{applicationdiscoveryservice_stop_continuous_export}
\title{Stop the continuous flow of agent's discovered data into Amazon Athena}
\usage{
applicationdiscoveryservice_stop_continuous_export(exportId)
}
\arguments{
\item{exportId}{[required] The unique ID assigned to this export.}
}
\description{
Stop the continuous flow of agent's discovered data into Amazon Athena.
}
\section{Request syntax}{
\preformatted{svc$stop_continuous_export(
  exportId = "string"
)
}
}

\keyword{internal}
