% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_update_application}
\alias{applicationdiscoveryservice_update_application}
\title{Updates metadata about an application}
\usage{
applicationdiscoveryservice_update_application(configurationId, name,
  description)
}
\arguments{
\item{configurationId}{[required] Configuration ID of the application to be updated.}

\item{name}{New name of the application to be updated.}

\item{description}{New description of the application to be updated.}
}
\description{
Updates metadata about an application.
}
\section{Request syntax}{
\preformatted{svc$update_application(
  configurationId = "string",
  name = "string",
  description = "string"
)
}
}

\keyword{internal}
