% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_create_event_subscription}
\alias{databasemigrationservice_create_event_subscription}
\title{Creates an AWS DMS event notification subscription}
\usage{
databasemigrationservice_create_event_subscription(SubscriptionName,
  SnsTopicArn, SourceType, EventCategories, SourceIds, Enabled, Tags)
}
\arguments{
\item{SubscriptionName}{[required] The name of the AWS DMS event notification subscription. This name must
be less than 255 characters.}

\item{SnsTopicArn}{[required] The Amazon Resource Name (ARN) of the Amazon SNS topic created for event
notification. The ARN is created by Amazon SNS when you create a topic
and subscribe to it.}

\item{SourceType}{The type of AWS DMS resource that generates the events. For example, if
you want to be notified of events generated by a replication instance,
you set this parameter to \code{replication-instance}. If this value isn\'t
specified, all events are returned.

Valid values: \code{replication-instance} \\| \code{replication-task}}

\item{EventCategories}{A list of event categories for a source type that you want to subscribe
to. For more information, see \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Events.html}{Working with Events and Notifications}
in the \emph{AWS Database Migration Service User Guide.}}

\item{SourceIds}{A list of identifiers for which AWS DMS provides notification events.

If you don\'t specify a value, notifications are provided for all
sources.

If you specify multiple values, they must be of the same type. For
example, if you specify a database instance ID, then all of the other
values must be database instance IDs.}

\item{Enabled}{A Boolean value; set to \code{true} to activate the subscription, or set to
\code{false} to create the subscription but not activate it.}

\item{Tags}{One or more tags to be assigned to the event subscription.}
}
\description{
Creates an AWS DMS event notification subscription.
}
\details{
You can specify the type of source (\code{SourceType}) you want to be
notified of, provide a list of AWS DMS source IDs (\code{SourceIds}) that
triggers the events, and provide a list of event categories
(\code{EventCategories}) for events you want to be notified of. If you
specify both the \code{SourceType} and \code{SourceIds}, such as
\code{SourceType = replication-instance} and
\code{SourceIdentifier = my-replinstance}, you will be notified of all the
replication instance events for the specified source. If you specify a
\code{SourceType} but don\'t specify a \code{SourceIdentifier}, you receive notice
of the events for that source type for all your AWS DMS sources. If you
don\'t specify either \code{SourceType} nor \code{SourceIdentifier}, you will be
notified of events generated from all AWS DMS sources belonging to your
customer account.

For more information about AWS DMS events, see \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Events.html}{Working with Events and Notifications}
in the \emph{AWS Database Migration Service User Guide.}
}
\section{Request syntax}{
\preformatted{svc$create_event_subscription(
  SubscriptionName = "string",
  SnsTopicArn = "string",
  SourceType = "string",
  EventCategories = list(
    "string"
  ),
  SourceIds = list(
    "string"
  ),
  Enabled = TRUE|FALSE,
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
