% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_create_replication_task}
\alias{databasemigrationservice_create_replication_task}
\title{Creates a replication task using the specified parameters}
\usage{
databasemigrationservice_create_replication_task(
  ReplicationTaskIdentifier, SourceEndpointArn, TargetEndpointArn,
  ReplicationInstanceArn, MigrationType, TableMappings,
  ReplicationTaskSettings, CdcStartTime, CdcStartPosition,
  CdcStopPosition, Tags, TaskData, ResourceIdentifier)
}
\arguments{
\item{ReplicationTaskIdentifier}{[required] An identifier for the replication task.

Constraints:
\itemize{
\item Must contain 1-255 alphanumeric characters or hyphens.
\item First character must be a letter.
\item Cannot end with a hyphen or contain two consecutive hyphens.
}}

\item{SourceEndpointArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies the source
endpoint.}

\item{TargetEndpointArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies the target
endpoint.}

\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of a replication instance.}

\item{MigrationType}{[required] The migration type. Valid values: \code{full-load} | \code{cdc} |
\code{full-load-and-cdc}}

\item{TableMappings}{[required] The table mappings for the task, in JSON format. For more information,
see \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Tasks.CustomizingTasks.TableMapping.html}{Using Table Mapping to Specify Task Settings}
in the \emph{AWS Database Migration Service User Guide.}}

\item{ReplicationTaskSettings}{Overall settings for the task, in JSON format. For more information, see
\href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Tasks.CustomizingTasks.TaskSettings.html}{Specifying Task Settings for AWS Database Migration Service Tasks}
in the \emph{AWS Database Migration User Guide.}}

\item{CdcStartTime}{Indicates the start time for a change data capture (CDC) operation. Use
either CdcStartTime or CdcStartPosition to specify when you want a CDC
operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time “2018-03-08T12:12:12”}

\item{CdcStartPosition}{Indicates when you want a change data capture (CDC) operation to start.
Use either CdcStartPosition or CdcStartTime to specify when you want a
CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position “2018-03-08T12:12:12”

Checkpoint Example: --cdc-start-position
"checkpoint:V1#27#mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:1876#0#0#*#0#93"

LSN Example: --cdc-start-position “mysql-bin-changelog.000024:373”

When you use this task setting with a source PostgreSQL database, a
logical replication slot should already be created and associated with
the source endpoint. You can verify this by setting the \code{slotName} extra
connection attribute to the name of this logical replication slot. For
more information, see \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Source.PostgreSQL.html#CHAP_Source.PostgreSQL.ConnectionAttrib}{Extra Connection Attributes When Using PostgreSQL as a Source for AWS DMS}.}

\item{CdcStopPosition}{Indicates when you want a change data capture (CDC) operation to stop.
The value can be either server time or commit time.

Server time example: --cdc-stop-position
“server_time:2018-02-09T12:12:12”

Commit time example: --cdc-stop-position “commit_time:
2018-02-09T12:12:12 “}

\item{Tags}{One or more tags to be assigned to the replication task.}

\item{TaskData}{Supplemental information that the task requires to migrate the data for
certain source and target endpoints. For more information, see
\href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Tasks.TaskData.html}{Specifying Supplemental Data for Task Settings}
in the \emph{AWS Database Migration Service User Guide.}}

\item{ResourceIdentifier}{A friendly name for the resource identifier at the end of the
\code{EndpointArn} response parameter that is returned in the created
\code{Endpoint} object. The value for this parameter can have up to 31
characters. It can contain only ASCII letters, digits, and hyphen ('-').
Also, it can't end with a hyphen or contain two consecutive hyphens, and
can only begin with a letter, such as \code{Example-App-ARN1}. For example,
this value might result in the \code{EndpointArn} value
\code{arn:aws:dms:eu-west-1:012345678901:rep:Example-App-ARN1}. If you don't
specify a \code{ResourceIdentifier} value, AWS DMS generates a default
identifier value for the end of \code{EndpointArn}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ReplicationTask = list(
    ReplicationTaskIdentifier = "string",
    SourceEndpointArn = "string",
    TargetEndpointArn = "string",
    ReplicationInstanceArn = "string",
    MigrationType = "full-load"|"cdc"|"full-load-and-cdc",
    TableMappings = "string",
    ReplicationTaskSettings = "string",
    Status = "string",
    LastFailureMessage = "string",
    StopReason = "string",
    ReplicationTaskCreationDate = as.POSIXct(
      "2015-01-01"
    ),
    ReplicationTaskStartDate = as.POSIXct(
      "2015-01-01"
    ),
    CdcStartPosition = "string",
    CdcStopPosition = "string",
    RecoveryCheckpoint = "string",
    ReplicationTaskArn = "string",
    ReplicationTaskStats = list(
      FullLoadProgressPercent = 123,
      ElapsedTimeMillis = 123,
      TablesLoaded = 123,
      TablesLoading = 123,
      TablesQueued = 123,
      TablesErrored = 123,
      FreshStartDate = as.POSIXct(
        "2015-01-01"
      ),
      StartDate = as.POSIXct(
        "2015-01-01"
      ),
      StopDate = as.POSIXct(
        "2015-01-01"
      ),
      FullLoadStartDate = as.POSIXct(
        "2015-01-01"
      ),
      FullLoadFinishDate = as.POSIXct(
        "2015-01-01"
      )
    ),
    TaskData = "string",
    TargetReplicationInstanceArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a replication task using the specified parameters.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_replication_task(
  ReplicationTaskIdentifier = "string",
  SourceEndpointArn = "string",
  TargetEndpointArn = "string",
  ReplicationInstanceArn = "string",
  MigrationType = "full-load"|"cdc"|"full-load-and-cdc",
  TableMappings = "string",
  ReplicationTaskSettings = "string",
  CdcStartTime = as.POSIXct(
    "2015-01-01"
  ),
  CdcStartPosition = "string",
  CdcStopPosition = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  TaskData = "string",
  ResourceIdentifier = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Creates a replication task using the specified parameters.
svc$create_replication_task(
  CdcStartTime = "2016-12-14T18:25:43Z",
  MigrationType = "full-load",
  ReplicationInstanceArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJQ",
  ReplicationTaskIdentifier = "task1",
  ReplicationTaskSettings = "",
  SourceEndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ZW5UAN6P4E77EC7YWHK4RZZ3BE",
  TableMappings = "file://mappingfile.json",
  Tags = list(
    list(
      Key = "Acount",
      Value = "24352226"
    )
  ),
  TargetEndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ7E"
)
}

}
\keyword{internal}
