% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_certificate}
\alias{databasemigrationservice_delete_certificate}
\title{Deletes the specified certificate}
\usage{
databasemigrationservice_delete_certificate(CertificateArn)
}
\arguments{
\item{CertificateArn}{[required] The Amazon Resource Name (ARN) of the deleted certificate.}
}
\description{
Deletes the specified certificate.
}
\section{Request syntax}{
\preformatted{svc$delete_certificate(
  CertificateArn = "string"
)
}
}

\examples{
# Deletes the specified certificate.
\donttest{svc$delete_certificate(
  CertificateArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUSM457DE6XFJCJQ"
)}

}
\keyword{internal}
