% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_endpoint}
\alias{databasemigrationservice_delete_endpoint}
\title{Deletes the specified endpoint}
\usage{
databasemigrationservice_delete_endpoint(EndpointArn)
}
\arguments{
\item{EndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the
endpoint.}
}
\value{
A list with the following syntax:\preformatted{list(
  Endpoint = list(
    EndpointIdentifier = "string",
    EndpointType = "source"|"target",
    EngineName = "string",
    EngineDisplayName = "string",
    Username = "string",
    ServerName = "string",
    Port = 123,
    DatabaseName = "string",
    ExtraConnectionAttributes = "string",
    Status = "string",
    KmsKeyId = "string",
    EndpointArn = "string",
    CertificateArn = "string",
    SslMode = "none"|"require"|"verify-ca"|"verify-full",
    ServiceAccessRoleArn = "string",
    ExternalTableDefinition = "string",
    ExternalId = "string",
    DynamoDbSettings = list(
      ServiceAccessRoleArn = "string"
    ),
    S3Settings = list(
      ServiceAccessRoleArn = "string",
      ExternalTableDefinition = "string",
      CsvRowDelimiter = "string",
      CsvDelimiter = "string",
      BucketFolder = "string",
      BucketName = "string",
      CompressionType = "none"|"gzip",
      EncryptionMode = "sse-s3"|"sse-kms",
      ServerSideEncryptionKmsKeyId = "string",
      DataFormat = "csv"|"parquet",
      EncodingType = "plain"|"plain-dictionary"|"rle-dictionary",
      DictPageSizeLimit = 123,
      RowGroupLength = 123,
      DataPageSize = 123,
      ParquetVersion = "parquet-1-0"|"parquet-2-0",
      EnableStatistics = TRUE|FALSE,
      IncludeOpForFullLoad = TRUE|FALSE,
      CdcInsertsOnly = TRUE|FALSE,
      TimestampColumnName = "string",
      ParquetTimestampInMillisecond = TRUE|FALSE,
      CdcInsertsAndUpdates = TRUE|FALSE,
      DatePartitionEnabled = TRUE|FALSE,
      DatePartitionSequence = "YYYYMMDD"|"YYYYMMDDHH"|"YYYYMM"|"MMYYYYDD"|"DDMMYYYY",
      DatePartitionDelimiter = "SLASH"|"UNDERSCORE"|"DASH"|"NONE",
      UseCsvNoSupValue = TRUE|FALSE,
      CsvNoSupValue = "string",
      PreserveTransactions = TRUE|FALSE,
      CdcPath = "string"
    ),
    DmsTransferSettings = list(
      ServiceAccessRoleArn = "string",
      BucketName = "string"
    ),
    MongoDbSettings = list(
      Username = "string",
      Password = "string",
      ServerName = "string",
      Port = 123,
      DatabaseName = "string",
      AuthType = "no"|"password",
      AuthMechanism = "default"|"mongodb_cr"|"scram_sha_1",
      NestingLevel = "none"|"one",
      ExtractDocId = "string",
      DocsToInvestigate = "string",
      AuthSource = "string",
      KmsKeyId = "string",
      SecretsManagerAccessRoleArn = "string",
      SecretsManagerSecretId = "string"
    ),
    KinesisSettings = list(
      StreamArn = "string",
      MessageFormat = "json"|"json-unformatted",
      ServiceAccessRoleArn = "string",
      IncludeTransactionDetails = TRUE|FALSE,
      IncludePartitionValue = TRUE|FALSE,
      PartitionIncludeSchemaTable = TRUE|FALSE,
      IncludeTableAlterOperations = TRUE|FALSE,
      IncludeControlDetails = TRUE|FALSE,
      IncludeNullAndEmpty = TRUE|FALSE
    ),
    KafkaSettings = list(
      Broker = "string",
      Topic = "string",
      MessageFormat = "json"|"json-unformatted",
      IncludeTransactionDetails = TRUE|FALSE,
      IncludePartitionValue = TRUE|FALSE,
      PartitionIncludeSchemaTable = TRUE|FALSE,
      IncludeTableAlterOperations = TRUE|FALSE,
      IncludeControlDetails = TRUE|FALSE,
      MessageMaxBytes = 123,
      IncludeNullAndEmpty = TRUE|FALSE
    ),
    ElasticsearchSettings = list(
      ServiceAccessRoleArn = "string",
      EndpointUri = "string",
      FullLoadErrorPercentage = 123,
      ErrorRetryDuration = 123
    ),
    NeptuneSettings = list(
      ServiceAccessRoleArn = "string",
      S3BucketName = "string",
      S3BucketFolder = "string",
      ErrorRetryDuration = 123,
      MaxFileSize = 123,
      MaxRetryCount = 123,
      IamAuthEnabled = TRUE|FALSE
    ),
    RedshiftSettings = list(
      AcceptAnyDate = TRUE|FALSE,
      AfterConnectScript = "string",
      BucketFolder = "string",
      BucketName = "string",
      CaseSensitiveNames = TRUE|FALSE,
      CompUpdate = TRUE|FALSE,
      ConnectionTimeout = 123,
      DatabaseName = "string",
      DateFormat = "string",
      EmptyAsNull = TRUE|FALSE,
      EncryptionMode = "sse-s3"|"sse-kms",
      ExplicitIds = TRUE|FALSE,
      FileTransferUploadStreams = 123,
      LoadTimeout = 123,
      MaxFileSize = 123,
      Password = "string",
      Port = 123,
      RemoveQuotes = TRUE|FALSE,
      ReplaceInvalidChars = "string",
      ReplaceChars = "string",
      ServerName = "string",
      ServiceAccessRoleArn = "string",
      ServerSideEncryptionKmsKeyId = "string",
      TimeFormat = "string",
      TrimBlanks = TRUE|FALSE,
      TruncateColumns = TRUE|FALSE,
      Username = "string",
      WriteBufferSize = 123,
      SecretsManagerAccessRoleArn = "string",
      SecretsManagerSecretId = "string"
    ),
    PostgreSQLSettings = list(
      AfterConnectScript = "string",
      CaptureDdls = TRUE|FALSE,
      MaxFileSize = 123,
      DatabaseName = "string",
      DdlArtifactsSchema = "string",
      ExecuteTimeout = 123,
      FailTasksOnLobTruncation = TRUE|FALSE,
      Password = "string",
      Port = 123,
      ServerName = "string",
      Username = "string",
      SlotName = "string",
      SecretsManagerAccessRoleArn = "string",
      SecretsManagerSecretId = "string"
    ),
    MySQLSettings = list(
      AfterConnectScript = "string",
      DatabaseName = "string",
      EventsPollInterval = 123,
      TargetDbType = "specific-database"|"multiple-databases",
      MaxFileSize = 123,
      ParallelLoadThreads = 123,
      Password = "string",
      Port = 123,
      ServerName = "string",
      ServerTimezone = "string",
      Username = "string",
      SecretsManagerAccessRoleArn = "string",
      SecretsManagerSecretId = "string"
    ),
    OracleSettings = list(
      AddSupplementalLogging = TRUE|FALSE,
      ArchivedLogDestId = 123,
      AdditionalArchivedLogDestId = 123,
      AllowSelectNestedTables = TRUE|FALSE,
      ParallelAsmReadThreads = 123,
      ReadAheadBlocks = 123,
      AccessAlternateDirectly = TRUE|FALSE,
      UseAlternateFolderForOnline = TRUE|FALSE,
      OraclePathPrefix = "string",
      UsePathPrefix = "string",
      ReplacePathPrefix = TRUE|FALSE,
      EnableHomogenousTablespace = TRUE|FALSE,
      DirectPathNoLog = TRUE|FALSE,
      ArchivedLogsOnly = TRUE|FALSE,
      AsmPassword = "string",
      AsmServer = "string",
      AsmUser = "string",
      CharLengthSemantics = "default"|"char"|"byte",
      DatabaseName = "string",
      DirectPathParallelLoad = TRUE|FALSE,
      FailTasksOnLobTruncation = TRUE|FALSE,
      NumberDatatypeScale = 123,
      Password = "string",
      Port = 123,
      ReadTableSpaceName = TRUE|FALSE,
      RetryInterval = 123,
      SecurityDbEncryption = "string",
      SecurityDbEncryptionName = "string",
      ServerName = "string",
      Username = "string",
      SecretsManagerAccessRoleArn = "string",
      SecretsManagerSecretId = "string",
      SecretsManagerOracleAsmAccessRoleArn = "string",
      SecretsManagerOracleAsmSecretId = "string"
    ),
    SybaseSettings = list(
      DatabaseName = "string",
      Password = "string",
      Port = 123,
      ServerName = "string",
      Username = "string",
      SecretsManagerAccessRoleArn = "string",
      SecretsManagerSecretId = "string"
    ),
    MicrosoftSQLServerSettings = list(
      Port = 123,
      BcpPacketSize = 123,
      DatabaseName = "string",
      ControlTablesFileGroup = "string",
      Password = "string",
      ReadBackupOnly = TRUE|FALSE,
      SafeguardPolicy = "rely-on-sql-server-replication-agent"|"exclusive-automatic-truncation"|"shared-automatic-truncation",
      ServerName = "string",
      Username = "string",
      UseBcpFullLoad = TRUE|FALSE,
      SecretsManagerAccessRoleArn = "string",
      SecretsManagerSecretId = "string"
    ),
    IBMDb2Settings = list(
      DatabaseName = "string",
      Password = "string",
      Port = 123,
      ServerName = "string",
      SetDataCaptureChanges = TRUE|FALSE,
      CurrentLsn = "string",
      MaxKBytesPerRead = 123,
      Username = "string",
      SecretsManagerAccessRoleArn = "string",
      SecretsManagerSecretId = "string"
    ),
    DocDbSettings = list(
      Username = "string",
      Password = "string",
      ServerName = "string",
      Port = 123,
      DatabaseName = "string",
      NestingLevel = "none"|"one",
      ExtractDocId = TRUE|FALSE,
      DocsToInvestigate = 123,
      KmsKeyId = "string",
      SecretsManagerAccessRoleArn = "string",
      SecretsManagerSecretId = "string"
    )
  )
)
}
}
\description{
Deletes the specified endpoint.

All tasks associated with the endpoint must be deleted before you can
delete the endpoint.
}
\section{Request syntax}{
\preformatted{svc$delete_endpoint(
  EndpointArn = "string"
)
}
}

\examples{
\dontrun{
# Deletes the specified endpoint. All tasks associated with the endpoint
# must be deleted before you can delete the endpoint.
# 
svc$delete_endpoint(
  EndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:RAAR3R22XSH46S3PWLC3NJAWKM"
)
}

}
\keyword{internal}
