% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_event_subscription}
\alias{databasemigrationservice_delete_event_subscription}
\title{Deletes an AWS DMS event subscription}
\usage{
databasemigrationservice_delete_event_subscription(SubscriptionName)
}
\arguments{
\item{SubscriptionName}{[required] The name of the DMS event notification subscription to be deleted.}
}
\description{
Deletes an AWS DMS event subscription.
}
\section{Request syntax}{
\preformatted{svc$delete_event_subscription(
  SubscriptionName = "string"
)
}
}

\keyword{internal}
