% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_replication_task}
\alias{databasemigrationservice_delete_replication_task}
\title{Deletes the specified replication task}
\usage{
databasemigrationservice_delete_replication_task(ReplicationTaskArn)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name (ARN) of the replication task to be deleted.}
}
\value{
A list with the following syntax:\preformatted{list(
  ReplicationTask = list(
    ReplicationTaskIdentifier = "string",
    SourceEndpointArn = "string",
    TargetEndpointArn = "string",
    ReplicationInstanceArn = "string",
    MigrationType = "full-load"|"cdc"|"full-load-and-cdc",
    TableMappings = "string",
    ReplicationTaskSettings = "string",
    Status = "string",
    LastFailureMessage = "string",
    StopReason = "string",
    ReplicationTaskCreationDate = as.POSIXct(
      "2015-01-01"
    ),
    ReplicationTaskStartDate = as.POSIXct(
      "2015-01-01"
    ),
    CdcStartPosition = "string",
    CdcStopPosition = "string",
    RecoveryCheckpoint = "string",
    ReplicationTaskArn = "string",
    ReplicationTaskStats = list(
      FullLoadProgressPercent = 123,
      ElapsedTimeMillis = 123,
      TablesLoaded = 123,
      TablesLoading = 123,
      TablesQueued = 123,
      TablesErrored = 123,
      FreshStartDate = as.POSIXct(
        "2015-01-01"
      ),
      StartDate = as.POSIXct(
        "2015-01-01"
      ),
      StopDate = as.POSIXct(
        "2015-01-01"
      ),
      FullLoadStartDate = as.POSIXct(
        "2015-01-01"
      ),
      FullLoadFinishDate = as.POSIXct(
        "2015-01-01"
      )
    ),
    TaskData = "string",
    TargetReplicationInstanceArn = "string"
  )
)
}
}
\description{
Deletes the specified replication task.
}
\section{Request syntax}{
\preformatted{svc$delete_replication_task(
  ReplicationTaskArn = "string"
)
}
}

\examples{
\dontrun{
# Deletes the specified replication task.
svc$delete_replication_task(
  ReplicationTaskArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJQ"
)
}

}
\keyword{internal}
