% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_replication_task_assessment_run}
\alias{databasemigrationservice_delete_replication_task_assessment_run}
\title{Deletes the record of a single premigration assessment run}
\usage{
databasemigrationservice_delete_replication_task_assessment_run(
  ReplicationTaskAssessmentRunArn)
}
\arguments{
\item{ReplicationTaskAssessmentRunArn}{[required] Amazon Resource Name (ARN) of the premigration assessment run to be
deleted.}
}
\value{
A list with the following syntax:\preformatted{list(
  ReplicationTaskAssessmentRun = list(
    ReplicationTaskAssessmentRunArn = "string",
    ReplicationTaskArn = "string",
    Status = "string",
    ReplicationTaskAssessmentRunCreationDate = as.POSIXct(
      "2015-01-01"
    ),
    AssessmentProgress = list(
      IndividualAssessmentCount = 123,
      IndividualAssessmentCompletedCount = 123
    ),
    LastFailureMessage = "string",
    ServiceAccessRoleArn = "string",
    ResultLocationBucket = "string",
    ResultLocationFolder = "string",
    ResultEncryptionMode = "string",
    ResultKmsKeyArn = "string",
    AssessmentRunName = "string"
  )
)
}
}
\description{
Deletes the record of a single premigration assessment run.

This operation removes all metadata that AWS DMS maintains about this
assessment run. However, the operation leaves untouched all information
about this assessment run that is stored in your Amazon S3 bucket.
}
\section{Request syntax}{
\preformatted{svc$delete_replication_task_assessment_run(
  ReplicationTaskAssessmentRunArn = "string"
)
}
}

\keyword{internal}
