% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_account_attributes}
\alias{databasemigrationservice_describe_account_attributes}
\title{Lists all of the AWS DMS attributes for a customer account}
\usage{
databasemigrationservice_describe_account_attributes()
}
\description{
Lists all of the AWS DMS attributes for a customer account. These
attributes include AWS DMS quotas for the account and a unique account
identifier in a particular DMS region. DMS quotas include a list of
resource quotas supported by the account, such as the number of
replication instances allowed. The description for each resource quota,
includes the quota name, current usage toward that quota, and the
quota\'s maximum value. DMS uses the unique account identifier to name
each artifact used by DMS in the given region.
}
\details{
This command does not take any parameters.
}
\section{Request syntax}{
\preformatted{svc$describe_account_attributes()
}
}

\examples{
# Lists all of the AWS DMS attributes for a customer account. The
# attributes include AWS DMS quotas for the account, such as the number of
# replication instances allowed. The description for a quota includes the
# quota name, current usage toward that quota, and the quota's maximum
# value. This operation does not take any parameters.
\dontrun{svc$describe_account_attributes()}

}
\keyword{internal}
