% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_applicable_individual_assessments}
\alias{databasemigrationservice_describe_applicable_individual_assessments}
\title{Provides a list of individual assessments that you can specify for a new
premigration assessment run, given one or more parameters}
\usage{
databasemigrationservice_describe_applicable_individual_assessments(
  ReplicationTaskArn, ReplicationInstanceArn, SourceEngineName,
  TargetEngineName, MigrationType, MaxRecords, Marker)
}
\arguments{
\item{ReplicationTaskArn}{Amazon Resource Name (ARN) of a migration task on which you want to base
the default list of individual assessments.}

\item{ReplicationInstanceArn}{ARN of a replication instance on which you want to base the default list
of individual assessments.}

\item{SourceEngineName}{Name of a database engine that the specified replication instance
supports as a source.}

\item{TargetEngineName}{Name of a database engine that the specified replication instance
supports as a target.}

\item{MigrationType}{Name of the migration type that each provided individual assessment must
support.}

\item{MaxRecords}{Maximum number of records to include in the response. If more records
exist than the specified \code{MaxRecords} value, a pagination token called a
marker is included in the response so that the remaining results can be
retrieved.}

\item{Marker}{Optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  IndividualAssessmentNames = list(
    "string"
  ),
  Marker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Provides a list of individual assessments that you can specify for a new
premigration assessment run, given one or more parameters.

If you specify an existing migration task, this operation provides the
default individual assessments you can specify for that task. Otherwise,
the specified parameters model elements of a possible migration task on
which to base a premigration assessment run.

To use these migration task modeling parameters, you must specify an
existing replication instance, a source database engine, a target
database engine, and a migration type. This combination of parameters
potentially limits the default individual assessments available for an
assessment run created for a corresponding migration task.

If you specify no parameters, this operation provides a list of all
possible individual assessments that you can specify for an assessment
run. If you specify any one of the task modeling parameters, you must
specify all of them or the operation cannot provide a list of individual
assessments. The only parameter that you can specify alone is for an
existing migration task. The specified task definition then determines
the default list of individual assessments that you can specify in an
assessment run for the task.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_applicable_individual_assessments(
  ReplicationTaskArn = "string",
  ReplicationInstanceArn = "string",
  SourceEngineName = "string",
  TargetEngineName = "string",
  MigrationType = "full-load"|"cdc"|"full-load-and-cdc",
  MaxRecords = 123,
  Marker = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
