% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_connections}
\alias{databasemigrationservice_describe_connections}
\title{Describes the status of the connections that have been made between the
replication instance and an endpoint}
\usage{
databasemigrationservice_describe_connections(Filters, MaxRecords,
  Marker)
}
\arguments{
\item{Filters}{The filters applied to the connection.

Valid filter names: endpoint-arn \\| replication-instance-arn}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Describes the status of the connections that have been made between the
replication instance and an endpoint. Connections are created when you
test an endpoint.
}
\section{Request syntax}{
\preformatted{svc$describe_connections(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}
}

\examples{
\dontrun{
# Describes the status of the connections that have been made between the
# replication instance and an endpoint. Connections are created when you
# test an endpoint.
svc$describe_connections(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string",
        "string"
      )
    )
  ),
  Marker = "",
  MaxRecords = 123L
)
}

}
\keyword{internal}
