% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_endpoint_settings}
\alias{databasemigrationservice_describe_endpoint_settings}
\title{Returns information about the possible endpoint settings available when
you create an endpoint for a specific database engine}
\usage{
databasemigrationservice_describe_endpoint_settings(EngineName,
  MaxRecords, Marker)
}
\arguments{
\item{EngineName}{[required] The databse engine used for your source or target endpoint.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\value{
A list with the following syntax:\preformatted{list(
  Marker = "string",
  EndpointSettings = list(
    list(
      Name = "string",
      Type = "string"|"boolean"|"integer"|"enum",
      EnumValues = list(
        "string"
      ),
      Sensitive = TRUE|FALSE,
      Units = "string",
      Applicability = "string",
      IntValueMin = 123,
      IntValueMax = 123
    )
  )
)
}
}
\description{
Returns information about the possible endpoint settings available when
you create an endpoint for a specific database engine.
}
\section{Request syntax}{
\preformatted{svc$describe_endpoint_settings(
  EngineName = "string",
  MaxRecords = 123,
  Marker = "string"
)
}
}

\keyword{internal}
