% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_endpoint_types}
\alias{databasemigrationservice_describe_endpoint_types}
\title{Returns information about the type of endpoints available}
\usage{
databasemigrationservice_describe_endpoint_types(Filters, MaxRecords,
  Marker)
}
\arguments{
\item{Filters}{Filters applied to the endpoint types.

Valid filter names: engine-name | endpoint-type}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Marker = "string",
  SupportedEndpointTypes = list(
    list(
      EngineName = "string",
      SupportsCDC = TRUE|FALSE,
      EndpointType = "source"|"target",
      ReplicationInstanceEngineMinimumVersion = "string",
      EngineDisplayName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about the type of endpoints available.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_endpoint_types(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Returns information about the type of endpoints available.
svc$describe_endpoint_types(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string",
        "string"
      )
    )
  ),
  Marker = "",
  MaxRecords = 123L
)
}

}
\keyword{internal}
