% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_endpoints}
\alias{databasemigrationservice_describe_endpoints}
\title{Returns information about the endpoints for your account in the current
region}
\usage{
databasemigrationservice_describe_endpoints(Filters, MaxRecords, Marker)
}
\arguments{
\item{Filters}{Filters applied to the describe action.

Valid filter names: endpoint-arn \\| endpoint-type \\| endpoint-id \\|
engine-name}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Returns information about the endpoints for your account in the current
region.
}
\section{Request syntax}{
\preformatted{svc$describe_endpoints(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}
}

\examples{
# Returns information about the endpoints for your account in the current
# region.
\donttest{svc$describe_endpoints(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string",
        "string"
      )
    )
  ),
  Marker = "",
  MaxRecords = 123L
)}

}
\keyword{internal}
