% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_event_categories}
\alias{databasemigrationservice_describe_event_categories}
\title{Lists categories for all event source types, or, if specified, for a
specified source type}
\usage{
databasemigrationservice_describe_event_categories(SourceType, Filters)
}
\arguments{
\item{SourceType}{The type of AWS DMS resource that generates events.

Valid values: replication-instance | replication-task}

\item{Filters}{Filters applied to the event categories.}
}
\value{
A list with the following syntax:\preformatted{list(
  EventCategoryGroupList = list(
    list(
      SourceType = "string",
      EventCategories = list(
        "string"
      )
    )
  )
)
}
}
\description{
Lists categories for all event source types, or, if specified, for a
specified source type. You can see a list of the event categories and
source types in \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Events.html}{Working with Events and Notifications}
in the \emph{AWS Database Migration Service User Guide.}
}
\section{Request syntax}{
\preformatted{svc$describe_event_categories(
  SourceType = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  )
)
}
}

\keyword{internal}
