% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_events}
\alias{databasemigrationservice_describe_events}
\title{Lists events for a given source identifier and source type}
\usage{
databasemigrationservice_describe_events(SourceIdentifier, SourceType,
  StartTime, EndTime, Duration, EventCategories, Filters, MaxRecords,
  Marker)
}
\arguments{
\item{SourceIdentifier}{The identifier of an event source.}

\item{SourceType}{The type of AWS DMS resource that generates events.

Valid values: replication-instance \| replication-task}

\item{StartTime}{The start time for the events to be listed.}

\item{EndTime}{The end time for the events to be listed.}

\item{Duration}{The duration of the events to be listed.}

\item{EventCategories}{A list of event categories for the source type that you've chosen.}

\item{Filters}{Filters applied to the action.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Lists events for a given source identifier and source type. You can also
specify a start and end time. For more information on AWS DMS events,
see \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Events.html}{Working with Events and Notifications}
in the \emph{AWS Database Migration User Guide.}
}
\section{Request syntax}{
\preformatted{svc$describe_events(
  SourceIdentifier = "string",
  SourceType = "replication-instance",
  StartTime = as.POSIXct(
    "2015-01-01"
  ),
  EndTime = as.POSIXct(
    "2015-01-01"
  ),
  Duration = 123,
  EventCategories = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}
}

\keyword{internal}
