% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_orderable_replication_instances}
\alias{databasemigrationservice_describe_orderable_replication_instances}
\title{Returns information about the replication instance types that can be
created in the specified region}
\usage{
databasemigrationservice_describe_orderable_replication_instances(
  MaxRecords, Marker)
}
\arguments{
\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  OrderableReplicationInstances = list(
    list(
      EngineVersion = "string",
      ReplicationInstanceClass = "string",
      StorageType = "string",
      MinAllocatedStorage = 123,
      MaxAllocatedStorage = 123,
      DefaultAllocatedStorage = 123,
      IncludedAllocatedStorage = 123,
      AvailabilityZones = list(
        "string"
      ),
      ReleaseStatus = "beta"
    )
  ),
  Marker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about the replication instance types that can be
created in the specified region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_orderable_replication_instances(
  MaxRecords = 123,
  Marker = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Returns information about the replication instance types that can be
# created in the specified region.
svc$describe_orderable_replication_instances(
  Marker = "",
  MaxRecords = 123L
)
}

}
\keyword{internal}
