% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_pending_maintenance_actions}
\alias{databasemigrationservice_describe_pending_maintenance_actions}
\title{For internal use only}
\usage{
databasemigrationservice_describe_pending_maintenance_actions(
  ReplicationInstanceArn, Filters, Marker, MaxRecords)
}
\arguments{
\item{ReplicationInstanceArn}{The Amazon Resource Name (ARN) of the replication instance.}

\item{Filters}{}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PendingMaintenanceActions = list(
    list(
      ResourceIdentifier = "string",
      PendingMaintenanceActionDetails = list(
        list(
          Action = "string",
          AutoAppliedAfterDate = as.POSIXct(
            "2015-01-01"
          ),
          ForcedApplyDate = as.POSIXct(
            "2015-01-01"
          ),
          OptInStatus = "string",
          CurrentApplyDate = as.POSIXct(
            "2015-01-01"
          ),
          Description = "string"
        )
      )
    )
  ),
  Marker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
For internal use only
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_pending_maintenance_actions(
  ReplicationInstanceArn = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  Marker = "string",
  MaxRecords = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
