% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_refresh_schemas_status}
\alias{databasemigrationservice_describe_refresh_schemas_status}
\title{Returns the status of the RefreshSchemas operation}
\usage{
databasemigrationservice_describe_refresh_schemas_status(EndpointArn)
}
\arguments{
\item{EndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the
endpoint.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RefreshSchemasStatus = list(
    EndpointArn = "string",
    ReplicationInstanceArn = "string",
    Status = "successful"|"failed"|"refreshing",
    LastRefreshDate = as.POSIXct(
      "2015-01-01"
    ),
    LastFailureMessage = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the status of the RefreshSchemas operation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_refresh_schemas_status(
  EndpointArn = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Returns the status of the refresh-schemas operation.
svc$describe_refresh_schemas_status(
  EndpointArn = ""
)
}

}
\keyword{internal}
