% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_replication_instances}
\alias{databasemigrationservice_describe_replication_instances}
\title{Returns information about replication instances for your account in the
current region}
\usage{
databasemigrationservice_describe_replication_instances(Filters,
  MaxRecords, Marker)
}
\arguments{
\item{Filters}{Filters applied to replication instances.

Valid filter names: replication-instance-arn | replication-instance-id |
replication-instance-class | engine-version}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Marker = "string",
  ReplicationInstances = list(
    list(
      ReplicationInstanceIdentifier = "string",
      ReplicationInstanceClass = "string",
      ReplicationInstanceStatus = "string",
      AllocatedStorage = 123,
      InstanceCreateTime = as.POSIXct(
        "2015-01-01"
      ),
      VpcSecurityGroups = list(
        list(
          VpcSecurityGroupId = "string",
          Status = "string"
        )
      ),
      AvailabilityZone = "string",
      ReplicationSubnetGroup = list(
        ReplicationSubnetGroupIdentifier = "string",
        ReplicationSubnetGroupDescription = "string",
        VpcId = "string",
        SubnetGroupStatus = "string",
        Subnets = list(
          list(
            SubnetIdentifier = "string",
            SubnetAvailabilityZone = list(
              Name = "string"
            ),
            SubnetStatus = "string"
          )
        )
      ),
      PreferredMaintenanceWindow = "string",
      PendingModifiedValues = list(
        ReplicationInstanceClass = "string",
        AllocatedStorage = 123,
        MultiAZ = TRUE|FALSE,
        EngineVersion = "string"
      ),
      MultiAZ = TRUE|FALSE,
      EngineVersion = "string",
      AutoMinorVersionUpgrade = TRUE|FALSE,
      KmsKeyId = "string",
      ReplicationInstanceArn = "string",
      ReplicationInstancePublicIpAddress = "string",
      ReplicationInstancePrivateIpAddress = "string",
      ReplicationInstancePublicIpAddresses = list(
        "string"
      ),
      ReplicationInstancePrivateIpAddresses = list(
        "string"
      ),
      PubliclyAccessible = TRUE|FALSE,
      SecondaryAvailabilityZone = "string",
      FreeUntil = as.POSIXct(
        "2015-01-01"
      ),
      DnsNameServers = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about replication instances for your account in the
current region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_replication_instances(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Returns the status of the refresh-schemas operation.
svc$describe_replication_instances(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string",
        "string"
      )
    )
  ),
  Marker = "",
  MaxRecords = 123L
)
}

}
\keyword{internal}
