% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_replication_task_individual_assessments}
\alias{databasemigrationservice_describe_replication_task_individual_assessments}
\title{Returns a paginated list of individual assessments based on filter
settings}
\usage{
databasemigrationservice_describe_replication_task_individual_assessments(
  Filters, MaxRecords, Marker)
}
\arguments{
\item{Filters}{Filters applied to the individual assessments described in the form of
key-value pairs.

Valid filter names: \code{replication-task-assessment-run-arn},
\code{replication-task-arn}, \code{status}}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Marker = "string",
  ReplicationTaskIndividualAssessments = list(
    list(
      ReplicationTaskIndividualAssessmentArn = "string",
      ReplicationTaskAssessmentRunArn = "string",
      IndividualAssessmentName = "string",
      Status = "string",
      ReplicationTaskIndividualAssessmentStartDate = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns a paginated list of individual assessments based on filter
settings.

These filter settings can specify a combination of premigration
assessment runs, migration tasks, and assessment status values.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_replication_task_individual_assessments(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
