% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_replication_tasks}
\alias{databasemigrationservice_describe_replication_tasks}
\title{Returns information about replication tasks for your account in the
current region}
\usage{
databasemigrationservice_describe_replication_tasks(Filters, MaxRecords,
  Marker, WithoutSettings)
}
\arguments{
\item{Filters}{Filters applied to replication tasks.

Valid filter names: replication-task-arn | replication-task-id |
migration-type | endpoint-arn | replication-instance-arn}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}

\item{WithoutSettings}{An option to set to avoid returning information about settings. Use this
to reduce overhead when setting information is too large. To use this
option, choose \code{true}; otherwise, choose \code{false} (the default).}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Marker = "string",
  ReplicationTasks = list(
    list(
      ReplicationTaskIdentifier = "string",
      SourceEndpointArn = "string",
      TargetEndpointArn = "string",
      ReplicationInstanceArn = "string",
      MigrationType = "full-load"|"cdc"|"full-load-and-cdc",
      TableMappings = "string",
      ReplicationTaskSettings = "string",
      Status = "string",
      LastFailureMessage = "string",
      StopReason = "string",
      ReplicationTaskCreationDate = as.POSIXct(
        "2015-01-01"
      ),
      ReplicationTaskStartDate = as.POSIXct(
        "2015-01-01"
      ),
      CdcStartPosition = "string",
      CdcStopPosition = "string",
      RecoveryCheckpoint = "string",
      ReplicationTaskArn = "string",
      ReplicationTaskStats = list(
        FullLoadProgressPercent = 123,
        ElapsedTimeMillis = 123,
        TablesLoaded = 123,
        TablesLoading = 123,
        TablesQueued = 123,
        TablesErrored = 123,
        FreshStartDate = as.POSIXct(
          "2015-01-01"
        ),
        StartDate = as.POSIXct(
          "2015-01-01"
        ),
        StopDate = as.POSIXct(
          "2015-01-01"
        ),
        FullLoadStartDate = as.POSIXct(
          "2015-01-01"
        ),
        FullLoadFinishDate = as.POSIXct(
          "2015-01-01"
        )
      ),
      TaskData = "string",
      TargetReplicationInstanceArn = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about replication tasks for your account in the
current region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_replication_tasks(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string",
  WithoutSettings = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Returns information about replication tasks for your account in the
# current region.
svc$describe_replication_tasks(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string",
        "string"
      )
    )
  ),
  Marker = "",
  MaxRecords = 123L
)
}

}
\keyword{internal}
