% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_import_certificate}
\alias{databasemigrationservice_import_certificate}
\title{Uploads the specified certificate}
\usage{
databasemigrationservice_import_certificate(CertificateIdentifier,
  CertificatePem, CertificateWallet, Tags)
}
\arguments{
\item{CertificateIdentifier}{[required] A customer-assigned name for the certificate. Identifiers must begin
with a letter and must contain only ASCII letters, digits, and hyphens.
They can't end with a hyphen or contain two consecutive hyphens.}

\item{CertificatePem}{The contents of a \code{.pem} file, which contains an X.509 certificate.}

\item{CertificateWallet}{The location of an imported Oracle Wallet certificate for use with SSL.}

\item{Tags}{The tags associated with the certificate.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Certificate = list(
    CertificateIdentifier = "string",
    CertificateCreationDate = as.POSIXct(
      "2015-01-01"
    ),
    CertificatePem = "string",
    CertificateWallet = raw,
    CertificateArn = "string",
    CertificateOwner = "string",
    ValidFromDate = as.POSIXct(
      "2015-01-01"
    ),
    ValidToDate = as.POSIXct(
      "2015-01-01"
    ),
    SigningAlgorithm = "string",
    KeyLength = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Uploads the specified certificate.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$import_certificate(
  CertificateIdentifier = "string",
  CertificatePem = "string",
  CertificateWallet = raw,
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Uploads the specified certificate.
svc$import_certificate(
  CertificateIdentifier = "",
  CertificatePem = ""
)
}

}
\keyword{internal}
