% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_modify_event_subscription}
\alias{databasemigrationservice_modify_event_subscription}
\title{Modifies an existing AWS DMS event notification subscription}
\usage{
databasemigrationservice_modify_event_subscription(SubscriptionName,
  SnsTopicArn, SourceType, EventCategories, Enabled)
}
\arguments{
\item{SubscriptionName}{[required] The name of the AWS DMS event notification subscription to be modified.}

\item{SnsTopicArn}{The Amazon Resource Name (ARN) of the Amazon SNS topic created for event
notification. The ARN is created by Amazon SNS when you create a topic
and subscribe to it.}

\item{SourceType}{The type of AWS DMS resource that generates the events you want to
subscribe to.

Valid values: replication-instance \| replication-task}

\item{EventCategories}{A list of event categories for a source type that you want to subscribe
to. Use the \code{DescribeEventCategories} action to see a list of event
categories.}

\item{Enabled}{A Boolean value; set to \strong{true} to activate the subscription.}
}
\description{
Modifies an existing AWS DMS event notification subscription.
}
\section{Request syntax}{
\preformatted{svc$modify_event_subscription(
  SubscriptionName = "string",
  SnsTopicArn = "string",
  SourceType = "string",
  EventCategories = list(
    "string"
  ),
  Enabled = TRUE|FALSE
)
}
}

\keyword{internal}
