% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_modify_replication_subnet_group}
\alias{databasemigrationservice_modify_replication_subnet_group}
\title{Modifies the settings for the specified replication subnet group}
\usage{
databasemigrationservice_modify_replication_subnet_group(
  ReplicationSubnetGroupIdentifier, ReplicationSubnetGroupDescription,
  SubnetIds)
}
\arguments{
\item{ReplicationSubnetGroupIdentifier}{[required] The name of the replication instance subnet group.}

\item{ReplicationSubnetGroupDescription}{A description for the replication instance subnet group.}

\item{SubnetIds}{[required] A list of subnet IDs.}
}
\value{
A list with the following syntax:\preformatted{list(
  ReplicationSubnetGroup = list(
    ReplicationSubnetGroupIdentifier = "string",
    ReplicationSubnetGroupDescription = "string",
    VpcId = "string",
    SubnetGroupStatus = "string",
    Subnets = list(
      list(
        SubnetIdentifier = "string",
        SubnetAvailabilityZone = list(
          Name = "string"
        ),
        SubnetStatus = "string"
      )
    )
  )
)
}
}
\description{
Modifies the settings for the specified replication subnet group.
}
\section{Request syntax}{
\preformatted{svc$modify_replication_subnet_group(
  ReplicationSubnetGroupIdentifier = "string",
  ReplicationSubnetGroupDescription = "string",
  SubnetIds = list(
    "string"
  )
)
}
}

\examples{
\dontrun{
# Modifies the settings for the specified replication subnet group.
svc$modify_replication_subnet_group(
  ReplicationSubnetGroupDescription = "",
  ReplicationSubnetGroupIdentifier = "",
  SubnetIds = list()
)
}

}
\keyword{internal}
