% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_reboot_replication_instance}
\alias{databasemigrationservice_reboot_replication_instance}
\title{Reboots a replication instance}
\usage{
databasemigrationservice_reboot_replication_instance(
  ReplicationInstanceArn, ForceFailover)
}
\arguments{
\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the replication instance.}

\item{ForceFailover}{If this parameter is \code{true}, the reboot is conducted through a Multi-AZ
failover. (If the instance isn't configured for Multi-AZ, then you can't
specify \code{true}.)}
}
\description{
Reboots a replication instance. Rebooting results in a momentary outage,
until the replication instance becomes available again.
}
\section{Request syntax}{
\preformatted{svc$reboot_replication_instance(
  ReplicationInstanceArn = "string",
  ForceFailover = TRUE|FALSE
)
}
}

\keyword{internal}
