% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_refresh_schemas}
\alias{databasemigrationservice_refresh_schemas}
\title{Populates the schema for the specified endpoint}
\usage{
databasemigrationservice_refresh_schemas(EndpointArn,
  ReplicationInstanceArn)
}
\arguments{
\item{EndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the
endpoint.}

\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the replication instance.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RefreshSchemasStatus = list(
    EndpointArn = "string",
    ReplicationInstanceArn = "string",
    Status = "successful"|"failed"|"refreshing",
    LastRefreshDate = as.POSIXct(
      "2015-01-01"
    ),
    LastFailureMessage = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Populates the schema for the specified endpoint. This is an asynchronous
operation and can take several minutes. You can check the status of this
operation by calling the DescribeRefreshSchemasStatus operation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$refresh_schemas(
  EndpointArn = "string",
  ReplicationInstanceArn = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Populates the schema for the specified endpoint. This is an asynchronous
# operation and can take several minutes. You can check the status of this
# operation by calling the describe-refresh-schemas-status operation.
svc$refresh_schemas(
  EndpointArn = "",
  ReplicationInstanceArn = ""
)
}

}
\keyword{internal}
