% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_start_replication_task_assessment}
\alias{databasemigrationservice_start_replication_task_assessment}
\title{Starts the replication task assessment for unsupported data types in the
source database}
\usage{
databasemigrationservice_start_replication_task_assessment(
  ReplicationTaskArn)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name (ARN) of the replication task.}
}
\value{
A list with the following syntax:\preformatted{list(
  ReplicationTask = list(
    ReplicationTaskIdentifier = "string",
    SourceEndpointArn = "string",
    TargetEndpointArn = "string",
    ReplicationInstanceArn = "string",
    MigrationType = "full-load"|"cdc"|"full-load-and-cdc",
    TableMappings = "string",
    ReplicationTaskSettings = "string",
    Status = "string",
    LastFailureMessage = "string",
    StopReason = "string",
    ReplicationTaskCreationDate = as.POSIXct(
      "2015-01-01"
    ),
    ReplicationTaskStartDate = as.POSIXct(
      "2015-01-01"
    ),
    CdcStartPosition = "string",
    CdcStopPosition = "string",
    RecoveryCheckpoint = "string",
    ReplicationTaskArn = "string",
    ReplicationTaskStats = list(
      FullLoadProgressPercent = 123,
      ElapsedTimeMillis = 123,
      TablesLoaded = 123,
      TablesLoading = 123,
      TablesQueued = 123,
      TablesErrored = 123,
      FreshStartDate = as.POSIXct(
        "2015-01-01"
      ),
      StartDate = as.POSIXct(
        "2015-01-01"
      ),
      StopDate = as.POSIXct(
        "2015-01-01"
      ),
      FullLoadStartDate = as.POSIXct(
        "2015-01-01"
      ),
      FullLoadFinishDate = as.POSIXct(
        "2015-01-01"
      )
    ),
    TaskData = "string",
    TargetReplicationInstanceArn = "string"
  )
)
}
}
\description{
Starts the replication task assessment for unsupported data types in the
source database.
}
\section{Request syntax}{
\preformatted{svc$start_replication_task_assessment(
  ReplicationTaskArn = "string"
)
}
}

\keyword{internal}
