% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_start_replication_task_assessment_run}
\alias{databasemigrationservice_start_replication_task_assessment_run}
\title{Starts a new premigration assessment run for one or more individual
assessments of a migration task}
\usage{
databasemigrationservice_start_replication_task_assessment_run(
  ReplicationTaskArn, ServiceAccessRoleArn, ResultLocationBucket,
  ResultLocationFolder, ResultEncryptionMode, ResultKmsKeyArn,
  AssessmentRunName, IncludeOnly, Exclude)
}
\arguments{
\item{ReplicationTaskArn}{[required] Amazon Resource Name (ARN) of the migration task associated with the
premigration assessment run that you want to start.}

\item{ServiceAccessRoleArn}{[required] ARN of a service role needed to start the assessment run.}

\item{ResultLocationBucket}{[required] Amazon S3 bucket where you want AWS DMS to store the results of this
assessment run.}

\item{ResultLocationFolder}{Folder within an Amazon S3 bucket where you want AWS DMS to store the
results of this assessment run.}

\item{ResultEncryptionMode}{Encryption mode that you can specify to encrypt the results of this
assessment run. If you don't specify this request parameter, AWS DMS
stores the assessment run results without encryption. You can specify
one of the options following:
\itemize{
\item \code{"SSE_S3"} – The server-side encryption provided as a default by
Amazon S3.
\item \code{"SSE_KMS"} – AWS Key Management Service (AWS KMS) encryption. This
encryption can use either a custom KMS encryption key that you
specify or the default KMS encryption key that DMS provides.
}}

\item{ResultKmsKeyArn}{ARN of a custom KMS encryption key that you specify when you set
\code{ResultEncryptionMode} to \verb{\\"SSE_KMS}".}

\item{AssessmentRunName}{[required] Unique name to identify the assessment run.}

\item{IncludeOnly}{Space-separated list of names for specific individual assessments that
you want to include. These names come from the default list of
individual assessments that AWS DMS supports for the associated
migration task. This task is specified by \code{ReplicationTaskArn}.

You can't set a value for \code{IncludeOnly} if you also set a value for
\code{Exclude} in the API operation.

To identify the names of the default individual assessments that AWS DMS
supports for the associated migration task, run the
\code{\link[=databasemigrationservice_describe_applicable_individual_assessments]{describe_applicable_individual_assessments}}
operation using its own \code{ReplicationTaskArn} request parameter.}

\item{Exclude}{Space-separated list of names for specific individual assessments that
you want to exclude. These names come from the default list of
individual assessments that AWS DMS supports for the associated
migration task. This task is specified by \code{ReplicationTaskArn}.

You can't set a value for \code{Exclude} if you also set a value for
\code{IncludeOnly} in the API operation.

To identify the names of the default individual assessments that AWS DMS
supports for the associated migration task, run the
\code{\link[=databasemigrationservice_describe_applicable_individual_assessments]{describe_applicable_individual_assessments}}
operation using its own \code{ReplicationTaskArn} request parameter.}
}
\description{
Starts a new premigration assessment run for one or more individual
assessments of a migration task.

The assessments that you can specify depend on the source and target
database engine and the migration type defined for the given task. To
run this operation, your migration task must already be created. After
you run this operation, you can review the status of each individual
assessment. You can also run the migration task manually after the
assessment run and its individual assessments complete.
}
\section{Request syntax}{
\preformatted{svc$start_replication_task_assessment_run(
  ReplicationTaskArn = "string",
  ServiceAccessRoleArn = "string",
  ResultLocationBucket = "string",
  ResultLocationFolder = "string",
  ResultEncryptionMode = "string",
  ResultKmsKeyArn = "string",
  AssessmentRunName = "string",
  IncludeOnly = list(
    "string"
  ),
  Exclude = list(
    "string"
  )
)
}
}

\keyword{internal}
