% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_stop_replication_task}
\alias{databasemigrationservice_stop_replication_task}
\title{Stops the replication task}
\usage{
databasemigrationservice_stop_replication_task(ReplicationTaskArn)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name(ARN) of the replication task to be stopped.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ReplicationTask = list(
    ReplicationTaskIdentifier = "string",
    SourceEndpointArn = "string",
    TargetEndpointArn = "string",
    ReplicationInstanceArn = "string",
    MigrationType = "full-load"|"cdc"|"full-load-and-cdc",
    TableMappings = "string",
    ReplicationTaskSettings = "string",
    Status = "string",
    LastFailureMessage = "string",
    StopReason = "string",
    ReplicationTaskCreationDate = as.POSIXct(
      "2015-01-01"
    ),
    ReplicationTaskStartDate = as.POSIXct(
      "2015-01-01"
    ),
    CdcStartPosition = "string",
    CdcStopPosition = "string",
    RecoveryCheckpoint = "string",
    ReplicationTaskArn = "string",
    ReplicationTaskStats = list(
      FullLoadProgressPercent = 123,
      ElapsedTimeMillis = 123,
      TablesLoaded = 123,
      TablesLoading = 123,
      TablesQueued = 123,
      TablesErrored = 123,
      FreshStartDate = as.POSIXct(
        "2015-01-01"
      ),
      StartDate = as.POSIXct(
        "2015-01-01"
      ),
      StopDate = as.POSIXct(
        "2015-01-01"
      ),
      FullLoadStartDate = as.POSIXct(
        "2015-01-01"
      ),
      FullLoadFinishDate = as.POSIXct(
        "2015-01-01"
      )
    ),
    TaskData = "string",
    TargetReplicationInstanceArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Stops the replication task.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_replication_task(
  ReplicationTaskArn = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Stops the replication task.
svc$stop_replication_task(
  ReplicationTaskArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ7E"
)
}

}
\keyword{internal}
