% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_service.R
\name{datasync}
\alias{datasync}
\title{AWS DataSync}
\usage{
datasync(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
AWS DataSync is a managed data transfer service that makes it simpler
for you to automate moving data between on-premises storage and Amazon
Simple Storage Service (Amazon S3) or Amazon Elastic File System (Amazon
EFS).

This API interface reference for AWS DataSync contains documentation for
a programming interface that you can use to manage AWS DataSync.
}
\section{Service syntax}{
\preformatted{svc <- datasync(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=datasync_cancel_task_execution]{cancel_task_execution} \tab Cancels execution of a task \cr
\link[=datasync_create_agent]{create_agent} \tab Activates an AWS DataSync agent that you have deployed on your host \cr
\link[=datasync_create_location_efs]{create_location_efs} \tab Creates an endpoint for an Amazon EFS file system \cr
\link[=datasync_create_location_nfs]{create_location_nfs} \tab Creates an endpoint for a Network File System (NFS) file system \cr
\link[=datasync_create_location_s3]{create_location_s3} \tab Creates an endpoint for an Amazon S3 bucket \cr
\link[=datasync_create_task]{create_task} \tab Creates a task \cr
\link[=datasync_delete_agent]{delete_agent} \tab Deletes an agent \cr
\link[=datasync_delete_location]{delete_location} \tab Deletes the configuration of a location used by AWS DataSync \cr
\link[=datasync_delete_task]{delete_task} \tab Deletes a task \cr
\link[=datasync_describe_agent]{describe_agent} \tab Returns metadata such as the name, the network interfaces, and the status (that is, whether the agent is running or not) for an agent\cr
\link[=datasync_describe_location_efs]{describe_location_efs} \tab Returns metadata, such as the path information about an Amazon EFS location \cr
\link[=datasync_describe_location_nfs]{describe_location_nfs} \tab Returns metadata, such as the path information, about a NFS location \cr
\link[=datasync_describe_location_s3]{describe_location_s3} \tab Returns metadata, such as bucket name, about an Amazon S3 bucket location \cr
\link[=datasync_describe_task]{describe_task} \tab Returns metadata about a task \cr
\link[=datasync_describe_task_execution]{describe_task_execution} \tab Returns detailed metadata about a task that is being executed \cr
\link[=datasync_list_agents]{list_agents} \tab Returns a list of agents owned by an AWS account in the AWS Region specified in the request \cr
\link[=datasync_list_locations]{list_locations} \tab Returns a lists of source and destination locations \cr
\link[=datasync_list_tags_for_resource]{list_tags_for_resource} \tab Returns all the tags associated with a specified resources \cr
\link[=datasync_list_task_executions]{list_task_executions} \tab Returns a list of executed tasks \cr
\link[=datasync_list_tasks]{list_tasks} \tab Returns a list of all the tasks \cr
\link[=datasync_start_task_execution]{start_task_execution} \tab Starts a specific invocation of a task \cr
\link[=datasync_tag_resource]{tag_resource} \tab Applies a key-value pair to an AWS resource \cr
\link[=datasync_untag_resource]{untag_resource} \tab Removes a tag from an AWS resource \cr
\link[=datasync_update_agent]{update_agent} \tab Updates the name of an agent \cr
\link[=datasync_update_task]{update_task} \tab Updates the metadata associated with a task
}
}

\examples{
\donttest{svc <- datasync()
svc$cancel_task_execution(
  Foo = 123
)}

}
