% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_create_agent}
\alias{datasync_create_agent}
\title{Activates an AWS DataSync agent that you have deployed on your host}
\usage{
datasync_create_agent(ActivationKey, AgentName, Tags, VpcEndpointId,
  SubnetArns, SecurityGroupArns)
}
\arguments{
\item{ActivationKey}{[required] Your agent activation key. You can get the activation key either by
sending an HTTP GET request with redirects that enable you to get the
agent IP address (port 80). Alternatively, you can get it from the AWS
DataSync console.

The redirect URL returned in the response provides you the activation
key for your agent in the query string parameter \code{activationKey}. It
might also include other activation-related parameters; however, these
are merely defaults. The arguments you pass to this API call determine
the actual configuration of your agent.

For more information, see Activating an Agent in the \emph{AWS DataSync User
Guide.}}

\item{AgentName}{The name you configured for your agent. This value is a text reference
that is used to identify the agent in the console.}

\item{Tags}{The key-value pair that represents the tag that you want to associate
with the agent. The value can be an empty string. This value helps you
manage, filter, and search for your agents.

Valid characters for key and value are letters, spaces, and numbers
representable in UTF-8 format, and the following special characters: + -
= . \\_ : / @.}

\item{VpcEndpointId}{The ID of the VPC (Virtual Private Cloud) endpoint that the agent has
access to. This is the client-side VPC endpoint, also called a
PrivateLink. If you don't have a PrivateLink VPC endpoint, see \href{https://docs.aws.amazon.com/vpc/latest/userguide/endpoint-service.html#create-endpoint-service}{Creating a VPC Endpoint Service Configuration}
in the AWS VPC User Guide.

VPC endpoint ID looks like this: \verb{vpce-01234d5aff67890e1}.}

\item{SubnetArns}{The Amazon Resource Names (ARNs) of the subnets in which DataSync will
create elastic network interfaces for each data transfer task. The agent
that runs a task must be private. When you start a task that is
associated with an agent created in a VPC, or one that has access to an
IP address in a VPC, then the task is also private. In this case,
DataSync creates four network interfaces for each task in your subnet.
For a data transfer to work, the agent must be able to route to all
these four network interfaces.}

\item{SecurityGroupArns}{The ARNs of the security groups used to protect your data transfer task
subnets. See CreateAgentRequest$SubnetArns.}
}
\description{
Activates an AWS DataSync agent that you have deployed on your host. The
activation process associates your agent with your account. In the
activation process, you specify information such as the AWS Region that
you want to activate the agent in. You activate the agent in the AWS
Region where your target locations (in Amazon S3 or Amazon EFS) reside.
Your tasks are created in this AWS Region.
}
\details{
You can activate the agent in a VPC (Virtual private Cloud) or provide
the agent access to a VPC endpoint so you can run tasks without going
over the public Internet.

You can use an agent for more than one location. If a task uses multiple
agents, all of them need to have status AVAILABLE for the task to run.
If you use multiple agents for a source location, the status of all the
agents must be AVAILABLE for the task to run.

Agents are automatically updated by AWS on a regular basis, using a
mechanism that ensures minimal interruption to your tasks.
}
\section{Request syntax}{
\preformatted{svc$create_agent(
  ActivationKey = "string",
  AgentName = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  VpcEndpointId = "string",
  SubnetArns = list(
    "string"
  ),
  SecurityGroupArns = list(
    "string"
  )
)
}
}

\keyword{internal}
