% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_create_location_efs}
\alias{datasync_create_location_efs}
\title{Creates an endpoint for an Amazon EFS file system}
\usage{
datasync_create_location_efs(Subdirectory, EfsFilesystemArn, Ec2Config,
  Tags)
}
\arguments{
\item{Subdirectory}{A subdirectory in the location's path. This subdirectory in the EFS file
system is used to read data from the EFS source location or write data
to the EFS destination. By default, AWS DataSync uses the root
directory.}

\item{EfsFilesystemArn}{[required] The Amazon Resource Name (ARN) for the Amazon EFS file system.}

\item{Ec2Config}{[required] The subnet and security group that the Amazon EFS file system uses. The
security group that you provide needs to be able to communicate with the
security group on the mount target in the subnet specified.

The exact relationship between security group M (of the mount target)
and security group S (which you provide for DataSync to use at this
stage) is as follows:
\itemize{
\item Security group M (which you associate with the mount target) must
allow inbound access for the Transmission Control Protocol (TCP) on
the NFS port (2049) from security group S. You can enable inbound
connections either by IP address (CIDR range) or security group.
\item Security group S (provided to DataSync to access EFS) should have a
rule that enables outbound connections to the NFS port on one of the
file system's mount targets. You can enable outbound connections
either by IP address (CIDR range) or security group.

For information about security groups and mount targets, see
"https://docs.aws.amazon.com/efs/latest/ug/security-considerations.html\#network-access"
(Security Groups for Amazon EC2 Instances and Mount Targets) in the
\emph{Amazon EFS User Guide}.
}}

\item{Tags}{The key-value pair that represents a tag that you want to add to the
resource. The value can be an empty string. This value helps you manage,
filter, and search for your resources. We recommend that you create a
name tag for your location.}
}
\description{
Creates an endpoint for an Amazon EFS file system.
}
\section{Request syntax}{
\preformatted{svc$create_location_efs(
  Subdirectory = "string",
  EfsFilesystemArn = "string",
  Ec2Config = list(
    SubnetArn = "string",
    SecurityGroupArns = list(
      "string"
    )
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
