% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_create_location_fsx_windows}
\alias{datasync_create_location_fsx_windows}
\title{Creates an endpoint for an Amazon FSx for Windows file system}
\usage{
datasync_create_location_fsx_windows(Subdirectory, FsxFilesystemArn,
  SecurityGroupArns, Tags, User, Domain, Password)
}
\arguments{
\item{Subdirectory}{A subdirectory in the location’s path. This subdirectory in the Amazon
FSx for Windows file system is used to read data from the Amazon FSx for
Windows source location or write data to the FSx for Windows
destination.}

\item{FsxFilesystemArn}{[required] The Amazon Resource Name (ARN) for the FSx for Windows file system.}

\item{SecurityGroupArns}{[required] The Amazon Resource Names (ARNs) of the security groups that are to use
to configure the FSx for Windows file system.}

\item{Tags}{The key-value pair that represents a tag that you want to add to the
resource. The value can be an empty string. This value helps you manage,
filter, and search for your resources. We recommend that you create a
name tag for your location.}

\item{User}{[required] The user who has the permissions to access files and folders in the FSx
for Windows file system.}

\item{Domain}{The name of the Windows domain that the FSx for Windows server belongs
to.}

\item{Password}{[required] The password of the user who has the permissions to access files and
folders in the FSx for Windows file system.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  LocationArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates an endpoint for an Amazon FSx for Windows file system.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_location_fsx_windows(
  Subdirectory = "string",
  FsxFilesystemArn = "string",
  SecurityGroupArns = list(
    "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  User = "string",
  Domain = "string",
  Password = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
