% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_create_location_object_storage}
\alias{datasync_create_location_object_storage}
\title{Creates an endpoint for a self-managed object storage bucket}
\usage{
datasync_create_location_object_storage(ServerHostname, ServerPort,
  ServerProtocol, Subdirectory, BucketName, AccessKey, SecretKey,
  AgentArns, Tags)
}
\arguments{
\item{ServerHostname}{[required] The name of the self-managed object storage server. This value is the IP
address or Domain Name Service (DNS) name of the object storage server.
An agent uses this host name to mount the object storage server in a
network.}

\item{ServerPort}{The port that your self-managed object storage server accepts inbound
network traffic on. The server port is set by default to TCP 80 (HTTP)
or TCP 443 (HTTPS). You can specify a custom port if your self-managed
object storage server requires one.}

\item{ServerProtocol}{The protocol that the object storage server uses to communicate. Valid
values are HTTP or HTTPS.}

\item{Subdirectory}{The subdirectory in the self-managed object storage server that is used
to read data from.}

\item{BucketName}{[required] The bucket on the self-managed object storage server that is used to
read data from.}

\item{AccessKey}{Optional. The access key is used if credentials are required to access
the self-managed object storage server. If your object storage requires
a user name and password to authenticate, use \code{AccessKey} and
\code{SecretKey} to provide the user name and password, respectively.}

\item{SecretKey}{Optional. The secret key is used if credentials are required to access
the self-managed object storage server. If your object storage requires
a user name and password to authenticate, use \code{AccessKey} and
\code{SecretKey} to provide the user name and password, respectively.}

\item{AgentArns}{[required] The Amazon Resource Name (ARN) of the agents associated with the
self-managed object storage server location.}

\item{Tags}{The key-value pair that represents the tag that you want to add to the
location. The value can be an empty string. We recommend using tags to
name your resources.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  LocationArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates an endpoint for a self-managed object storage bucket. For more
information about self-managed object storage locations, see
create-object-location.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_location_object_storage(
  ServerHostname = "string",
  ServerPort = 123,
  ServerProtocol = "HTTPS"|"HTTP",
  Subdirectory = "string",
  BucketName = "string",
  AccessKey = "string",
  SecretKey = "string",
  AgentArns = list(
    "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
