% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_create_location_s3}
\alias{datasync_create_location_s3}
\title{Creates an endpoint for an Amazon S3 bucket}
\usage{
datasync_create_location_s3(Subdirectory, S3BucketArn, S3Config, Tags)
}
\arguments{
\item{Subdirectory}{A subdirectory in the Amazon S3 bucket. This subdirectory in Amazon S3
is used to read data from the S3 source location or write data to the S3
destination.}

\item{S3BucketArn}{[required] The Amazon Resource Name (ARN) of the Amazon S3 bucket.}

\item{S3Config}{[required]}

\item{Tags}{The key-value pair that represents the tag that you want to add to the
location. The value can be an empty string. We recommend using tags to
name your resources.}
}
\description{
Creates an endpoint for an Amazon S3 bucket.
}
\details{
For AWS DataSync to access a destination S3 bucket, it needs an AWS
Identity and Access Management (IAM) role that has the required
permissions. You can set up the required permissions by creating an IAM
policy that grants the required permissions and attaching the policy to
the role. An example of such a policy is shown in the examples section.

For more information, see
"https://docs.aws.amazon.com/datasync/latest/userguide/working-with-locations.html\#create-s3-location"
(Configuring Amazon S3 Location Settings) in the \emph{AWS DataSync User
Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_location_s3(
  Subdirectory = "string",
  S3BucketArn = "string",
  S3Config = list(
    BucketAccessRoleArn = "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
