% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_create_task}
\alias{datasync_create_task}
\title{Creates a task}
\usage{
datasync_create_task(SourceLocationArn, DestinationLocationArn,
  CloudWatchLogGroupArn, Name, Options, Excludes, Schedule, Tags)
}
\arguments{
\item{SourceLocationArn}{[required] The Amazon Resource Name (ARN) of the source location for the task.}

\item{DestinationLocationArn}{[required] The Amazon Resource Name (ARN) of an AWS storage resource's location.}

\item{CloudWatchLogGroupArn}{The Amazon Resource Name (ARN) of the Amazon CloudWatch log group that
is used to monitor and log events in the task.}

\item{Name}{The name of a task. This value is a text reference that is used to
identify the task in the console.}

\item{Options}{The set of configuration options that control the behavior of a single
execution of the task that occurs when you call
\code{\link[=datasync_start_task_execution]{start_task_execution}}. You can
configure these options to preserve metadata such as user ID (UID) and
group ID (GID), file permissions, data integrity verification, and so
on.

For each individual task execution, you can override these options by
specifying the \code{OverrideOptions} before starting the task execution. For
more information, see the operation.}

\item{Excludes}{A list of filter rules that determines which files to exclude from a
task. The list should contain a single filter string that consists of
the patterns to exclude. The patterns are delimited by "|" (that is, a
pipe), for example, \code{"/folder1|/folder2"}.}

\item{Schedule}{Specifies a schedule used to periodically transfer files from a source
to a destination location. The schedule should be specified in UTC time.
For more information, see task-scheduling.}

\item{Tags}{The key-value pair that represents the tag that you want to add to the
resource. The value can be an empty string.}
}
\value{
A list with the following syntax:\preformatted{list(
  TaskArn = "string"
)
}
}
\description{
Creates a task.

A task includes a source location and a destination location, and a
configuration that specifies how data is transferred. A task always
transfers data from the source location to the destination location. The
configuration specifies options such as task scheduling, bandwidth
limits, etc. A task is the complete definition of a data transfer.

When you create a task that transfers data between AWS services in
different AWS Regions, one of the two locations that you specify must
reside in the Region where DataSync is being used. The other location
must be specified in a different Region.

You can transfer data between commercial AWS Regions except for China,
or between AWS GovCloud (US-East and US-West) Regions.

When you use DataSync to copy files or objects between AWS Regions, you
pay for data transfer between Regions. This is billed as data transfer
OUT from your source Region to your destination Region. For more
information, see \href{https://aws.amazon.com/ec2/pricing/on-demand/#Data_Transfer}{Data Transfer pricing}.
}
\section{Request syntax}{
\preformatted{svc$create_task(
  SourceLocationArn = "string",
  DestinationLocationArn = "string",
  CloudWatchLogGroupArn = "string",
  Name = "string",
  Options = list(
    VerifyMode = "POINT_IN_TIME_CONSISTENT"|"ONLY_FILES_TRANSFERRED"|"NONE",
    OverwriteMode = "ALWAYS"|"NEVER",
    Atime = "NONE"|"BEST_EFFORT",
    Mtime = "NONE"|"PRESERVE",
    Uid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    Gid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    PreserveDeletedFiles = "PRESERVE"|"REMOVE",
    PreserveDevices = "NONE"|"PRESERVE",
    PosixPermissions = "NONE"|"PRESERVE",
    BytesPerSecond = 123,
    TaskQueueing = "ENABLED"|"DISABLED",
    LogLevel = "OFF"|"BASIC"|"TRANSFER",
    TransferMode = "CHANGED"|"ALL"
  ),
  Excludes = list(
    list(
      FilterType = "SIMPLE_PATTERN",
      Value = "string"
    )
  ),
  Schedule = list(
    ScheduleExpression = "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
