% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_delete_agent}
\alias{datasync_delete_agent}
\title{Deletes an agent}
\usage{
datasync_delete_agent(AgentArn)
}
\arguments{
\item{AgentArn}{[required] The Amazon Resource Name (ARN) of the agent to delete. Use the
\code{\link[=datasync_list_agents]{list_agents}} operation to return a list of
agents for your account and AWS Region.}
}
\value{
An empty list.
}
\description{
Deletes an agent. To specify which agent to delete, use the Amazon
Resource Name (ARN) of the agent in your request. The operation
disassociates the agent from your AWS account. However, it doesn't
delete the agent virtual machine (VM) from your on-premises environment.
}
\section{Request syntax}{
\preformatted{svc$delete_agent(
  AgentArn = "string"
)
}
}

\keyword{internal}
