% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_delete_location}
\alias{datasync_delete_location}
\title{Deletes the configuration of a location used by AWS DataSync}
\usage{
datasync_delete_location(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the location to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes the configuration of a location used by AWS DataSync.
}
\section{Request syntax}{
\preformatted{svc$delete_location(
  LocationArn = "string"
)
}
}

\keyword{internal}
