% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_delete_task}
\alias{datasync_delete_task}
\title{Deletes a task}
\usage{
datasync_delete_task(TaskArn)
}
\arguments{
\item{TaskArn}{[required] The Amazon Resource Name (ARN) of the task to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a task.
}
\section{Request syntax}{
\preformatted{svc$delete_task(
  TaskArn = "string"
)
}
}

\keyword{internal}
