% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_agent}
\alias{datasync_describe_agent}
\title{Returns metadata such as the name, the network interfaces, and the
status (that is, whether the agent is running or not) for an agent}
\usage{
datasync_describe_agent(AgentArn)
}
\arguments{
\item{AgentArn}{[required] The Amazon Resource Name (ARN) of the agent to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AgentArn = "string",
  Name = "string",
  Status = "ONLINE"|"OFFLINE",
  LastConnectionTime = as.POSIXct(
    "2015-01-01"
  ),
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  EndpointType = "PUBLIC"|"PRIVATE_LINK"|"FIPS",
  PrivateLinkConfig = list(
    VpcEndpointId = "string",
    PrivateLinkEndpoint = "string",
    SubnetArns = list(
      "string"
    ),
    SecurityGroupArns = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns metadata such as the name, the network interfaces, and the
status (that is, whether the agent is running or not) for an agent. To
specify which agent to describe, use the Amazon Resource Name (ARN) of
the agent in your request.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_agent(
  AgentArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
