% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_location_efs}
\alias{datasync_describe_location_efs}
\title{Returns metadata, such as the path information about an Amazon EFS
location}
\usage{
datasync_describe_location_efs(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the EFS location to describe.}
}
\description{
Returns metadata, such as the path information about an Amazon EFS
location.
}
\section{Request syntax}{
\preformatted{svc$describe_location_efs(
  LocationArn = "string"
)
}
}

\keyword{internal}
