% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_location_fsx_windows}
\alias{datasync_describe_location_fsx_windows}
\title{Returns metadata, such as the path information about an Amazon FSx for
Windows location}
\usage{
datasync_describe_location_fsx_windows(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the FSx for Windows location to
describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  LocationArn = "string",
  LocationUri = "string",
  SecurityGroupArns = list(
    "string"
  ),
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  User = "string",
  Domain = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns metadata, such as the path information about an Amazon FSx for
Windows location.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_location_fsx_windows(
  LocationArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
