% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_location_s3}
\alias{datasync_describe_location_s3}
\title{Returns metadata, such as bucket name, about an Amazon S3 bucket
location}
\usage{
datasync_describe_location_s3(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the Amazon S3 bucket location to
describe.}
}
\description{
Returns metadata, such as bucket name, about an Amazon S3 bucket
location.
}
\section{Request syntax}{
\preformatted{svc$describe_location_s3(
  LocationArn = "string"
)
}
}

\keyword{internal}
