% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_location_smb}
\alias{datasync_describe_location_smb}
\title{Returns metadata, such as the path and user information about an SMB
location}
\usage{
datasync_describe_location_smb(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the SMB location to describe.}
}
\value{
A list with the following syntax:\preformatted{list(
  LocationArn = "string",
  LocationUri = "string",
  AgentArns = list(
    "string"
  ),
  User = "string",
  Domain = "string",
  MountOptions = list(
    Version = "AUTOMATIC"|"SMB2"|"SMB3"
  ),
  CreationTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Returns metadata, such as the path and user information about an SMB
location.
}
\section{Request syntax}{
\preformatted{svc$describe_location_smb(
  LocationArn = "string"
)
}
}

\keyword{internal}
