% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_list_agents}
\alias{datasync_list_agents}
\title{Returns a list of agents owned by an AWS account in the AWS Region
specified in the request}
\usage{
datasync_list_agents(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of agents to list.}

\item{NextToken}{An opaque string that indicates the position at which to begin the next
list of agents.}
}
\value{
A list with the following syntax:\preformatted{list(
  Agents = list(
    list(
      AgentArn = "string",
      Name = "string",
      Status = "ONLINE"|"OFFLINE"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns a list of agents owned by an AWS account in the AWS Region
specified in the request. The returned list is ordered by agent Amazon
Resource Name (ARN).

By default, this operation returns a maximum of 100 agents. This
operation supports pagination that enables you to optionally reduce the
number of agents returned in a response.

If you have more agents than are returned in a response (that is, the
response returns only a truncated list of your agents), the response
contains a marker that you can specify in your next request to fetch the
next page of agents.
}
\section{Request syntax}{
\preformatted{svc$list_agents(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
