% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_list_locations}
\alias{datasync_list_locations}
\title{Returns a list of source and destination locations}
\usage{
datasync_list_locations(MaxResults, NextToken, Filters)
}
\arguments{
\item{MaxResults}{The maximum number of locations to return.}

\item{NextToken}{An opaque string that indicates the position at which to begin the next
list of locations.}

\item{Filters}{You can use API filters to narrow down the list of resources returned by
\code{\link[=datasync_list_locations]{list_locations}}. For example, to retrieve
all tasks on a specific source location, you can use
\code{\link[=datasync_list_locations]{list_locations}} with filter name
\verb{LocationType S3} and \verb{Operator Equals}.}
}
\value{
A list with the following syntax:\preformatted{list(
  Locations = list(
    list(
      LocationArn = "string",
      LocationUri = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns a list of source and destination locations.

If you have more locations than are returned in a response (that is, the
response returns only a truncated list of your agents), the response
contains a token that you can specify in your next request to fetch the
next page of locations.
}
\section{Request syntax}{
\preformatted{svc$list_locations(
  MaxResults = 123,
  NextToken = "string",
  Filters = list(
    list(
      Name = "LocationUri"|"LocationType"|"CreationTime",
      Values = list(
        "string"
      ),
      Operator = "Equals"|"NotEquals"|"In"|"LessThanOrEqual"|"LessThan"|"GreaterThanOrEqual"|"GreaterThan"|"Contains"|"NotContains"|"BeginsWith"
    )
  )
)
}
}

\keyword{internal}
