% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_list_tags_for_resource}
\alias{datasync_list_tags_for_resource}
\title{Returns all the tags associated with a specified resource}
\usage{
datasync_list_tags_for_resource(ResourceArn, MaxResults, NextToken)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource whose tags to list.}

\item{MaxResults}{The maximum number of locations to return.}

\item{NextToken}{An opaque string that indicates the position at which to begin the next
list of locations.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns all the tags associated with a specified resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
