% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_list_task_executions}
\alias{datasync_list_task_executions}
\title{Returns a list of executed tasks}
\usage{
datasync_list_task_executions(TaskArn, MaxResults, NextToken)
}
\arguments{
\item{TaskArn}{The Amazon Resource Name (ARN) of the task whose tasks you want to list.}

\item{MaxResults}{The maximum number of executed tasks to list.}

\item{NextToken}{An opaque string that indicates the position at which to begin the next
list of the executed tasks.}
}
\description{
Returns a list of executed tasks.
}
\section{Request syntax}{
\preformatted{svc$list_task_executions(
  TaskArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
