% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_untag_resource}
\alias{datasync_untag_resource}
\title{Removes a tag from an AWS resource}
\usage{
datasync_untag_resource(ResourceArn, Keys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to remove the tag from.}

\item{Keys}{[required] The keys in the key-value pair in the tag to remove.}
}
\value{
An empty list.
}
\description{
Removes a tag from an AWS resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$untag_resource(
  ResourceArn = "string",
  Keys = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
