% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_update_agent}
\alias{datasync_update_agent}
\title{Updates the name of an agent}
\usage{
datasync_update_agent(AgentArn, Name)
}
\arguments{
\item{AgentArn}{[required] The Amazon Resource Name (ARN) of the agent to update.}

\item{Name}{The name that you want to use to configure the agent.}
}
\value{
An empty list.
}
\description{
Updates the name of an agent.
}
\section{Request syntax}{
\preformatted{svc$update_agent(
  AgentArn = "string",
  Name = "string"
)
}
}

\keyword{internal}
