% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_update_location_nfs}
\alias{datasync_update_location_nfs}
\title{Updates some of the parameters of a previously created location for
Network File System (NFS) access}
\usage{
datasync_update_location_nfs(LocationArn, Subdirectory, OnPremConfig,
  MountOptions)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the NFS location to update.}

\item{Subdirectory}{The subdirectory in the NFS file system that is used to read data from
the NFS source location or write data to the NFS destination. The NFS
path should be a path that's exported by the NFS server, or a
subdirectory of that path. The path should be such that it can be
mounted by other NFS clients in your network.

To see all the paths exported by your NFS server, run
"\verb{showmount -e nfs-server-name}" from an NFS client that has access to
your server. You can specify any directory that appears in the results,
and any subdirectory of that directory. Ensure that the NFS export is
accessible without Kerberos authentication.

To transfer all the data in the folder that you specified, DataSync must
have permissions to read all the data. To ensure this, either configure
the NFS export with \code{no_root_squash}, or ensure that the files you want
DataSync to access have permissions that allow read access for all
users. Doing either option enables the agent to read the files. For the
agent to access directories, you must additionally enable all execute
access.

If you are copying data to or from your AWS Snowcone device, see \href{https://docs.aws.amazon.com/datasync/latest/userguide/create-nfs-location.html#nfs-on-snowcone}{NFS Server on AWS Snowcone}
for more information.

For information about NFS export configuration, see 18.7. The
/etc/exports Configuration File in the Red Hat Enterprise Linux
documentation.}

\item{OnPremConfig}{}

\item{MountOptions}{}
}
\value{
An empty list.
}
\description{
Updates some of the parameters of a previously created location for
Network File System (NFS) access. For information about creating an NFS
location, see create-nfs-location.
}
\section{Request syntax}{
\preformatted{svc$update_location_nfs(
  LocationArn = "string",
  Subdirectory = "string",
  OnPremConfig = list(
    AgentArns = list(
      "string"
    )
  ),
  MountOptions = list(
    Version = "AUTOMATIC"|"NFS3"|"NFS4_0"|"NFS4_1"
  )
)
}
}

\keyword{internal}
