% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_update_location_object_storage}
\alias{datasync_update_location_object_storage}
\title{Updates some of the parameters of a previously created location for
self-managed object storage server access}
\usage{
datasync_update_location_object_storage(LocationArn, ServerPort,
  ServerProtocol, Subdirectory, AccessKey, SecretKey, AgentArns)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the self-managed object storage server
location to be updated.}

\item{ServerPort}{The port that your self-managed object storage server accepts inbound
network traffic on. The server port is set by default to TCP 80 (HTTP)
or TCP 443 (HTTPS). You can specify a custom port if your self-managed
object storage server requires one.}

\item{ServerProtocol}{The protocol that the object storage server uses to communicate. Valid
values are \code{HTTP} or \code{HTTPS}.}

\item{Subdirectory}{The subdirectory in the self-managed object storage server that is used
to read data from.}

\item{AccessKey}{Optional. The access key is used if credentials are required to access
the self-managed object storage server. If your object storage requires
a user name and password to authenticate, use \code{AccessKey} and
\code{SecretKey} to provide the user name and password, respectively.}

\item{SecretKey}{Optional. The secret key is used if credentials are required to access
the self-managed object storage server. If your object storage requires
a user name and password to authenticate, use \code{AccessKey} and
\code{SecretKey} to provide the user name and password, respectively.}

\item{AgentArns}{The Amazon Resource Name (ARN) of the agents associated with the
self-managed object storage server location.}
}
\value{
An empty list.
}
\description{
Updates some of the parameters of a previously created location for
self-managed object storage server access. For information about
creating a self-managed object storage location, see
create-object-location.
}
\section{Request syntax}{
\preformatted{svc$update_location_object_storage(
  LocationArn = "string",
  ServerPort = 123,
  ServerProtocol = "HTTPS"|"HTTP",
  Subdirectory = "string",
  AccessKey = "string",
  SecretKey = "string",
  AgentArns = list(
    "string"
  )
)
}
}

\keyword{internal}
