% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_update_location_smb}
\alias{datasync_update_location_smb}
\title{Updates some of the parameters of a previously created location for
Server Message Block (SMB) file system access}
\usage{
datasync_update_location_smb(LocationArn, Subdirectory, User, Domain,
  Password, AgentArns, MountOptions)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the SMB location to update.}

\item{Subdirectory}{The subdirectory in the SMB file system that is used to read data from
the SMB source location or write data to the SMB destination. The SMB
path should be a path that's exported by the SMB server, or a
subdirectory of that path. The path should be such that it can be
mounted by other SMB clients in your network.

\code{Subdirectory} must be specified with forward slashes. For example,
\verb{/path/to/folder}.

To transfer all the data in the folder that you specified, DataSync must
have permissions to mount the SMB share and to access all the data in
that share. To ensure this, do either of the following:
\itemize{
\item Ensure that the user/password specified belongs to the user who can
mount the share and who has the appropriate permissions for all of
the files and directories that you want DataSync to access.
\item Use credentials of a member of the Backup Operators group to mount
the share.
}

Doing either of these options enables the agent to access the data. For
the agent to access directories, you must also enable all execute
access.}

\item{User}{The user who can mount the share has the permissions to access files and
folders in the SMB share.}

\item{Domain}{The name of the Windows domain that the SMB server belongs to.}

\item{Password}{The password of the user who can mount the share has the permissions to
access files and folders in the SMB share.}

\item{AgentArns}{The Amazon Resource Names (ARNs) of agents to use for a Simple Message
Block (SMB) location.}

\item{MountOptions}{}
}
\value{
An empty list.
}
\description{
Updates some of the parameters of a previously created location for
Server Message Block (SMB) file system access. For information about
creating an SMB location, see create-smb-location.
}
\section{Request syntax}{
\preformatted{svc$update_location_smb(
  LocationArn = "string",
  Subdirectory = "string",
  User = "string",
  Domain = "string",
  Password = "string",
  AgentArns = list(
    "string"
  ),
  MountOptions = list(
    Version = "AUTOMATIC"|"SMB2"|"SMB3"
  )
)
}
}

\keyword{internal}
