% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_update_task}
\alias{datasync_update_task}
\title{Updates the metadata associated with a task}
\usage{
datasync_update_task(TaskArn, Options, Excludes, Schedule, Name,
  CloudWatchLogGroupArn)
}
\arguments{
\item{TaskArn}{[required] The Amazon Resource Name (ARN) of the resource name of the task to
update.}

\item{Options}{}

\item{Excludes}{A list of filter rules that determines which files to exclude from a
task. The list should contain a single filter string that consists of
the patterns to exclude. The patterns are delimited by "|" (that is, a
pipe), for example: \code{"/folder1|/folder2"}}

\item{Schedule}{Specifies a schedule used to periodically transfer files from a source
to a destination location. You can configure your task to execute
hourly, daily, weekly or on specific days of the week. You control when
in the day or hour you want the task to execute. The time you specify is
UTC time. For more information, see task-scheduling.}

\item{Name}{The name of the task to update.}

\item{CloudWatchLogGroupArn}{The Amazon Resource Name (ARN) of the resource name of the CloudWatch
LogGroup.}
}
\description{
Updates the metadata associated with a task.
}
\section{Request syntax}{
\preformatted{svc$update_task(
  TaskArn = "string",
  Options = list(
    VerifyMode = "POINT_IN_TIME_CONSISTENT"|"ONLY_FILES_TRANSFERRED"|"NONE",
    OverwriteMode = "ALWAYS"|"NEVER",
    Atime = "NONE"|"BEST_EFFORT",
    Mtime = "NONE"|"PRESERVE",
    Uid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    Gid = "NONE"|"INT_VALUE"|"NAME"|"BOTH",
    PreserveDeletedFiles = "PRESERVE"|"REMOVE",
    PreserveDevices = "NONE"|"PRESERVE",
    PosixPermissions = "NONE"|"PRESERVE",
    BytesPerSecond = 123,
    TaskQueueing = "ENABLED"|"DISABLED",
    LogLevel = "OFF"|"BASIC"|"TRANSFER"
  ),
  Excludes = list(
    list(
      FilterType = "SIMPLE_PATTERN",
      Value = "string"
    )
  ),
  Schedule = list(
    ScheduleExpression = "string"
  ),
  Name = "string",
  CloudWatchLogGroupArn = "string"
)
}
}

\keyword{internal}
