% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport_service.R
\name{importexport}
\alias{importexport}
\title{AWS Import/Export}
\usage{
importexport(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
AWS Import/Export Service AWS Import/Export accelerates transferring
large amounts of data between the AWS cloud and portable storage devices
that you mail to us. AWS Import/Export transfers data directly onto and
off of your storage devices using Amazon's high-speed internal network
and bypassing the Internet. For large data sets, AWS Import/Export is
often faster than Internet transfer and more cost effective than
upgrading your connectivity.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- importexport(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=importexport_cancel_job]{cancel_job} \tab This operation cancels a specified job\cr
\link[=importexport_create_job]{create_job} \tab This operation initiates the process of scheduling an upload or download of your data\cr
\link[=importexport_get_shipping_label]{get_shipping_label} \tab This operation generates a pre-paid UPS shipping label that you will use to ship your device to AWS for processing\cr
\link[=importexport_get_status]{get_status} \tab This operation returns information about a job, including where the job is in the processing pipeline, the status of the results, and the signature value associated with the job\cr
\link[=importexport_list_jobs]{list_jobs} \tab This operation returns the jobs associated with the requester\cr
\link[=importexport_update_job]{update_job} \tab You use this operation to change the parameters specified in the original manifest file by supplying a new manifest file
}
}

\examples{
\dontrun{
svc <- importexport()
svc$cancel_job(
  Foo = 123
)
}

}
