% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport_operations.R
\name{importexport_cancel_job}
\alias{importexport_cancel_job}
\title{This operation cancels a specified job}
\usage{
importexport_cancel_job(JobId, APIVersion)
}
\arguments{
\item{JobId}{[required]}

\item{APIVersion}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Success = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
This operation cancels a specified job. Only the job owner can cancel
it. The operation fails if the job has already started or is complete.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$cancel_job(
  JobId = "string",
  APIVersion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
