% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport_operations.R
\name{importexport_create_job}
\alias{importexport_create_job}
\title{This operation initiates the process of scheduling an upload or download
of your data}
\usage{
importexport_create_job(JobType, Manifest, ManifestAddendum,
  ValidateOnly, APIVersion)
}
\arguments{
\item{JobType}{[required]}

\item{Manifest}{[required]}

\item{ManifestAddendum}{}

\item{ValidateOnly}{[required]}

\item{APIVersion}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  JobId = "string",
  JobType = "Import"|"Export",
  Signature = "string",
  SignatureFileContents = "string",
  WarningMessage = "string",
  ArtifactList = list(
    list(
      Description = "string",
      URL = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
This operation initiates the process of scheduling an upload or download
of your data. You include in the request a manifest that describes the
data transfer specifics. The response to the request includes a job ID,
which you can use in other operations, a signature that you use to
identify your storage device, and the address where you should ship your
storage device.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_job(
  JobType = "Import"|"Export",
  Manifest = "string",
  ManifestAddendum = "string",
  ValidateOnly = TRUE|FALSE,
  APIVersion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
