% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport_operations.R
\name{importexport_get_shipping_label}
\alias{importexport_get_shipping_label}
\title{This operation generates a pre-paid UPS shipping label that you will use
to ship your device to AWS for processing}
\usage{
importexport_get_shipping_label(jobIds, name, company, phoneNumber,
  country, stateOrProvince, city, postalCode, street1, street2, street3,
  APIVersion)
}
\arguments{
\item{jobIds}{[required]}

\item{name}{}

\item{company}{}

\item{phoneNumber}{}

\item{country}{}

\item{stateOrProvince}{}

\item{city}{}

\item{postalCode}{}

\item{street1}{}

\item{street2}{}

\item{street3}{}

\item{APIVersion}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ShippingLabelURL = "string",
  Warning = "string"
)
}\if{html}{\out{</div>}}
}
\description{
This operation generates a pre-paid UPS shipping label that you will use
to ship your device to AWS for processing.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_shipping_label(
  jobIds = list(
    "string"
  ),
  name = "string",
  company = "string",
  phoneNumber = "string",
  country = "string",
  stateOrProvince = "string",
  city = "string",
  postalCode = "string",
  street1 = "string",
  street2 = "string",
  street3 = "string",
  APIVersion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
