% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport_operations.R
\name{importexport_get_status}
\alias{importexport_get_status}
\title{This operation returns information about a job, including where the job
is in the processing pipeline, the status of the results, and the
signature value associated with the job}
\usage{
importexport_get_status(JobId, APIVersion)
}
\arguments{
\item{JobId}{[required]}

\item{APIVersion}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  JobId = "string",
  JobType = "Import"|"Export",
  LocationCode = "string",
  LocationMessage = "string",
  ProgressCode = "string",
  ProgressMessage = "string",
  Carrier = "string",
  TrackingNumber = "string",
  LogBucket = "string",
  LogKey = "string",
  ErrorCount = 123,
  Signature = "string",
  SignatureFileContents = "string",
  CurrentManifest = "string",
  CreationDate = as.POSIXct(
    "2015-01-01"
  ),
  ArtifactList = list(
    list(
      Description = "string",
      URL = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
This operation returns information about a job, including where the job
is in the processing pipeline, the status of the results, and the
signature value associated with the job. You can only return information
about jobs you own.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_status(
  JobId = "string",
  APIVersion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
