% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport_operations.R
\name{importexport_update_job}
\alias{importexport_update_job}
\title{You use this operation to change the parameters specified in the
original manifest file by supplying a new manifest file}
\usage{
importexport_update_job(JobId, Manifest, JobType, ValidateOnly,
  APIVersion)
}
\arguments{
\item{JobId}{[required]}

\item{Manifest}{[required]}

\item{JobType}{[required]}

\item{ValidateOnly}{[required]}

\item{APIVersion}{}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Success = TRUE|FALSE,
  WarningMessage = "string",
  ArtifactList = list(
    list(
      Description = "string",
      URL = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
You use this operation to change the parameters specified in the
original manifest file by supplying a new manifest file. The manifest
file attached to this request replaces the original manifest file. You
can only use the operation after a CreateJob request but before the data
transfer starts and you can only use it on jobs you own.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_job(
  JobId = "string",
  Manifest = "string",
  JobType = "Import"|"Export",
  ValidateOnly = TRUE|FALSE,
  APIVersion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
