% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_service.R
\name{migrationhub}
\alias{migrationhub}
\title{AWS Migration Hub}
\usage{
migrationhub(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
The AWS Migration Hub API methods help to obtain server and application
migration status and integrate your resource-specific migration tool by
providing a programmatic interface to Migration Hub.

Remember that you must set your AWS Migration Hub home region before you
call any of these APIs, or a \code{HomeRegionNotSetException} error will be
returned. Also, you must make the API calls while in your home region.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- migrationhub(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=migrationhub_associate_created_artifact]{associate_created_artifact} \tab Associates a created artifact of an AWS cloud resource, the target receiving the migration, with the migration task performed by a migration tool\cr
\link[=migrationhub_associate_discovered_resource]{associate_discovered_resource} \tab Associates a discovered resource ID from Application Discovery Service with a migration task\cr
\link[=migrationhub_create_progress_update_stream]{create_progress_update_stream} \tab Creates a progress update stream which is an AWS resource used for access control as well as a namespace for migration task names that is implicitly linked to your AWS account\cr
\link[=migrationhub_delete_progress_update_stream]{delete_progress_update_stream} \tab Deletes a progress update stream, including all of its tasks, which was previously created as an AWS resource used for access control\cr
\link[=migrationhub_describe_application_state]{describe_application_state} \tab Gets the migration status of an application\cr
\link[=migrationhub_describe_migration_task]{describe_migration_task} \tab Retrieves a list of all attributes associated with a specific migration task\cr
\link[=migrationhub_disassociate_created_artifact]{disassociate_created_artifact} \tab Disassociates a created artifact of an AWS resource with a migration task performed by a migration tool that was previously associated\cr
\link[=migrationhub_disassociate_discovered_resource]{disassociate_discovered_resource} \tab Disassociate an Application Discovery Service discovered resource from a migration task\cr
\link[=migrationhub_import_migration_task]{import_migration_task} \tab Registers a new migration task which represents a server, database, etc\cr
\link[=migrationhub_list_application_states]{list_application_states} \tab Lists all the migration statuses for your applications\cr
\link[=migrationhub_list_created_artifacts]{list_created_artifacts} \tab Lists the created artifacts attached to a given migration task in an update stream\cr
\link[=migrationhub_list_discovered_resources]{list_discovered_resources} \tab Lists discovered resources associated with the given MigrationTask\cr
\link[=migrationhub_list_migration_tasks]{list_migration_tasks} \tab Lists all, or filtered by resource name, migration tasks associated with the user account making this call\cr
\link[=migrationhub_list_progress_update_streams]{list_progress_update_streams} \tab Lists progress update streams associated with the user account making this call\cr
\link[=migrationhub_notify_application_state]{notify_application_state} \tab Sets the migration state of an application\cr
\link[=migrationhub_notify_migration_task_state]{notify_migration_task_state} \tab Notifies Migration Hub of the current status, progress, or other detail regarding a migration task\cr
\link[=migrationhub_put_resource_attributes]{put_resource_attributes} \tab Provides identifying details of the resource being migrated so that it can be associated in the Application Discovery Service repository
}
}

\examples{
\dontrun{
svc <- migrationhub()
svc$associate_created_artifact(
  Foo = 123
)
}

}
