% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_create_progress_update_stream}
\alias{migrationhub_create_progress_update_stream}
\title{Creates a progress update stream which is an AWS resource used for
access control as well as a namespace for migration task names that is
implicitly linked to your AWS account}
\usage{
migrationhub_create_progress_update_stream(ProgressUpdateStreamName,
  DryRun)
}
\arguments{
\item{ProgressUpdateStreamName}{[required] The name of the ProgressUpdateStream. \emph{Do not store personal data in
this field.}}

\item{DryRun}{Optional boolean flag to indicate whether any effect should take place.
Used to test if the caller has permission to make the call.}
}
\value{
An empty list.
}
\description{
Creates a progress update stream which is an AWS resource used for
access control as well as a namespace for migration task names that is
implicitly linked to your AWS account. It must uniquely identify the
migration tool as it is used for all updates made by the tool; however,
it does not need to be unique for each AWS account because it is scoped
to the AWS account.
}
\section{Request syntax}{
\preformatted{svc$create_progress_update_stream(
  ProgressUpdateStreamName = "string",
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
