% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_import_migration_task}
\alias{migrationhub_import_migration_task}
\title{Registers a new migration task which represents a server, database, etc}
\usage{
migrationhub_import_migration_task(ProgressUpdateStream,
  MigrationTaskName, DryRun)
}
\arguments{
\item{ProgressUpdateStream}{[required] The name of the ProgressUpdateStream. >}

\item{MigrationTaskName}{[required] Unique identifier that references the migration task. \emph{Do not store
personal data in this field.}}

\item{DryRun}{Optional boolean flag to indicate whether any effect should take place.
Used to test if the caller has permission to make the call.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Registers a new migration task which represents a server, database,
etc., being migrated to AWS by a migration tool.

This API is a prerequisite to calling the
\code{\link[=migrationhub_notify_migration_task_state]{notify_migration_task_state}}
API as the migration tool must first register the migration task with
Migration Hub.
}
\section{Request syntax}{
\preformatted{svc$import_migration_task(
  ProgressUpdateStream = "string",
  MigrationTaskName = "string",
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
