% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_list_application_states}
\alias{migrationhub_list_application_states}
\title{Lists all the migration statuses for your applications}
\usage{
migrationhub_list_application_states(ApplicationIds, NextToken,
  MaxResults)
}
\arguments{
\item{ApplicationIds}{The configurationIds from the Application Discovery Service that
uniquely identifies your applications.}

\item{NextToken}{If a \code{NextToken} was returned by a previous call, there are more results
available. To retrieve the next page of results, make the call again
using the returned token in \code{NextToken}.}

\item{MaxResults}{Maximum number of results to be returned per page.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ApplicationStateList = list(
    list(
      ApplicationId = "string",
      ApplicationStatus = "NOT_STARTED"|"IN_PROGRESS"|"COMPLETED",
      LastUpdatedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all the migration statuses for your applications. If you use the
optional \code{ApplicationIds} parameter, only the migration statuses for
those applications will be returned.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_application_states(
  ApplicationIds = list(
    "string"
  ),
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
