% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_list_created_artifacts}
\alias{migrationhub_list_created_artifacts}
\title{Lists the created artifacts attached to a given migration task in an
update stream}
\usage{
migrationhub_list_created_artifacts(ProgressUpdateStream,
  MigrationTaskName, NextToken, MaxResults)
}
\arguments{
\item{ProgressUpdateStream}{[required] The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{[required] Unique identifier that references the migration task. \emph{Do not store
personal data in this field.}}

\item{NextToken}{If a \code{NextToken} was returned by a previous call, there are more results
available. To retrieve the next page of results, make the call again
using the returned token in \code{NextToken}.}

\item{MaxResults}{Maximum number of results to be returned per page.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  CreatedArtifactList = list(
    list(
      Name = "string",
      Description = "string"
    )
  )
)
}
}
\description{
Lists the created artifacts attached to a given migration task in an
update stream. This API has the following traits:
\itemize{
\item Gets the list of the created artifacts while migration is taking
place.
\item Shows the artifacts created by the migration tool that was
associated by the
\code{\link[=migrationhub_associate_created_artifact]{associate_created_artifact}}
API.
\item Lists created artifacts in a paginated interface.
}
}
\section{Request syntax}{
\preformatted{svc$list_created_artifacts(
  ProgressUpdateStream = "string",
  MigrationTaskName = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
