% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_list_discovered_resources}
\alias{migrationhub_list_discovered_resources}
\title{Lists discovered resources associated with the given MigrationTask}
\usage{
migrationhub_list_discovered_resources(ProgressUpdateStream,
  MigrationTaskName, NextToken, MaxResults)
}
\arguments{
\item{ProgressUpdateStream}{[required] The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{[required] The name of the MigrationTask. \emph{Do not store personal data in this
field.}}

\item{NextToken}{If a \code{NextToken} was returned by a previous call, there are more results
available. To retrieve the next page of results, make the call again
using the returned token in \code{NextToken}.}

\item{MaxResults}{The maximum number of results returned per page.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  DiscoveredResourceList = list(
    list(
      ConfigurationId = "string",
      Description = "string"
    )
  )
)
}
}
\description{
Lists discovered resources associated with the given \code{MigrationTask}.
}
\section{Request syntax}{
\preformatted{svc$list_discovered_resources(
  ProgressUpdateStream = "string",
  MigrationTaskName = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
