% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_list_migration_tasks}
\alias{migrationhub_list_migration_tasks}
\title{Lists all, or filtered by resource name, migration tasks associated with
the user account making this call}
\usage{
migrationhub_list_migration_tasks(NextToken, MaxResults, ResourceName)
}
\arguments{
\item{NextToken}{If a \code{NextToken} was returned by a previous call, there are more results
available. To retrieve the next page of results, make the call again
using the returned token in \code{NextToken}.}

\item{MaxResults}{Value to specify how many results are returned per page.}

\item{ResourceName}{Filter migration tasks by discovered resource name.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  MigrationTaskSummaryList = list(
    list(
      ProgressUpdateStream = "string",
      MigrationTaskName = "string",
      Status = "NOT_STARTED"|"IN_PROGRESS"|"FAILED"|"COMPLETED",
      ProgressPercent = 123,
      StatusDetail = "string",
      UpdateDateTime = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Lists all, or filtered by resource name, migration tasks associated with
the user account making this call. This API has the following traits:
\itemize{
\item Can show a summary list of the most recent migration tasks.
\item Can show a summary list of migration tasks associated with a given
discovered resource.
\item Lists migration tasks in a paginated interface.
}
}
\section{Request syntax}{
\preformatted{svc$list_migration_tasks(
  NextToken = "string",
  MaxResults = 123,
  ResourceName = "string"
)
}
}

\keyword{internal}
