% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_list_progress_update_streams}
\alias{migrationhub_list_progress_update_streams}
\title{Lists progress update streams associated with the user account making
this call}
\usage{
migrationhub_list_progress_update_streams(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{If a \code{NextToken} was returned by a previous call, there are more results
available. To retrieve the next page of results, make the call again
using the returned token in \code{NextToken}.}

\item{MaxResults}{Filter to limit the maximum number of results to list per page.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ProgressUpdateStreamSummaryList = list(
    list(
      ProgressUpdateStreamName = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists progress update streams associated with the user account making
this call.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_progress_update_streams(
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
