% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_notify_application_state}
\alias{migrationhub_notify_application_state}
\title{Sets the migration state of an application}
\usage{
migrationhub_notify_application_state(ApplicationId, Status,
  UpdateDateTime, DryRun)
}
\arguments{
\item{ApplicationId}{[required] The configurationId in Application Discovery Service that uniquely
identifies the grouped application.}

\item{Status}{[required] Status of the application - Not Started, In-Progress, Complete.}

\item{UpdateDateTime}{The timestamp when the application state changed.}

\item{DryRun}{Optional boolean flag to indicate whether any effect should take place.
Used to test if the caller has permission to make the call.}
}
\value{
An empty list.
}
\description{
Sets the migration state of an application. For a given application
identified by the value passed to \code{ApplicationId}, its status is set or
updated by passing one of three values to \code{Status}:
\code{NOT_STARTED | IN_PROGRESS | COMPLETED}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$notify_application_state(
  ApplicationId = "string",
  Status = "NOT_STARTED"|"IN_PROGRESS"|"COMPLETED",
  UpdateDateTime = as.POSIXct(
    "2015-01-01"
  ),
  DryRun = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
