% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_notify_migration_task_state}
\alias{migrationhub_notify_migration_task_state}
\title{Notifies Migration Hub of the current status, progress, or other detail
regarding a migration task}
\usage{
migrationhub_notify_migration_task_state(ProgressUpdateStream,
  MigrationTaskName, Task, UpdateDateTime, NextUpdateSeconds, DryRun)
}
\arguments{
\item{ProgressUpdateStream}{[required] The name of the ProgressUpdateStream.}

\item{MigrationTaskName}{[required] Unique identifier that references the migration task. \emph{Do not store
personal data in this field.}}

\item{Task}{[required] Information about the task's progress and status.}

\item{UpdateDateTime}{[required] The timestamp when the task was gathered.}

\item{NextUpdateSeconds}{[required] Number of seconds after the UpdateDateTime within which the Migration
Hub can expect an update. If Migration Hub does not receive an update
within the specified interval, then the migration task will be
considered stale.}

\item{DryRun}{Optional boolean flag to indicate whether any effect should take place.
Used to test if the caller has permission to make the call.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Notifies Migration Hub of the current status, progress, or other detail
regarding a migration task. This API has the following traits:
\itemize{
\item Migration tools will call the
\code{\link[=migrationhub_notify_migration_task_state]{notify_migration_task_state}}
API to share the latest progress and status.
\item \code{MigrationTaskName} is used for addressing updates to the correct
target.
\item \code{ProgressUpdateStream} is used for access control and to provide a
namespace for each migration tool.
}
}
\section{Request syntax}{
\preformatted{svc$notify_migration_task_state(
  ProgressUpdateStream = "string",
  MigrationTaskName = "string",
  Task = list(
    Status = "NOT_STARTED"|"IN_PROGRESS"|"FAILED"|"COMPLETED",
    StatusDetail = "string",
    ProgressPercent = 123
  ),
  UpdateDateTime = as.POSIXct(
    "2015-01-01"
  ),
  NextUpdateSeconds = 123,
  DryRun = TRUE|FALSE
)
}
}

\keyword{internal}
