% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_create_app}
\alias{sms_create_app}
\title{Creates an application}
\usage{
sms_create_app(name, description, roleName, clientToken, serverGroups,
  tags)
}
\arguments{
\item{name}{The name of the new application.}

\item{description}{The description of the new application}

\item{roleName}{The name of the service role in the customer's account to be used by AWS
SMS.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of application creation.}

\item{serverGroups}{The server groups to include in the application.}

\item{tags}{The tags to be associated with the application.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  appSummary = list(
    appId = "string",
    importedAppId = "string",
    name = "string",
    description = "string",
    status = "CREATING"|"ACTIVE"|"UPDATING"|"DELETING"|"DELETED"|"DELETE_FAILED",
    statusMessage = "string",
    replicationConfigurationStatus = "NOT_CONFIGURED"|"CONFIGURED",
    replicationStatus = "READY_FOR_CONFIGURATION"|"CONFIGURATION_IN_PROGRESS"|"CONFIGURATION_INVALID"|"READY_FOR_REPLICATION"|"VALIDATION_IN_PROGRESS"|"REPLICATION_PENDING"|"REPLICATION_IN_PROGRESS"|"REPLICATED"|"PARTIALLY_REPLICATED"|"DELTA_REPLICATION_IN_PROGRESS"|"DELTA_REPLICATED"|"DELTA_REPLICATION_FAILED"|"REPLICATION_FAILED"|"REPLICATION_STOPPING"|"REPLICATION_STOP_FAILED"|"REPLICATION_STOPPED",
    replicationStatusMessage = "string",
    latestReplicationTime = as.POSIXct(
      "2015-01-01"
    ),
    launchConfigurationStatus = "NOT_CONFIGURED"|"CONFIGURED",
    launchStatus = "READY_FOR_CONFIGURATION"|"CONFIGURATION_IN_PROGRESS"|"CONFIGURATION_INVALID"|"READY_FOR_LAUNCH"|"VALIDATION_IN_PROGRESS"|"LAUNCH_PENDING"|"LAUNCH_IN_PROGRESS"|"LAUNCHED"|"PARTIALLY_LAUNCHED"|"DELTA_LAUNCH_IN_PROGRESS"|"DELTA_LAUNCH_FAILED"|"LAUNCH_FAILED"|"TERMINATE_IN_PROGRESS"|"TERMINATE_FAILED"|"TERMINATED",
    launchStatusMessage = "string",
    launchDetails = list(
      latestLaunchTime = as.POSIXct(
        "2015-01-01"
      ),
      stackName = "string",
      stackId = "string"
    ),
    creationTime = as.POSIXct(
      "2015-01-01"
    ),
    lastModified = as.POSIXct(
      "2015-01-01"
    ),
    roleName = "string",
    totalServerGroups = 123,
    totalServers = 123
  ),
  serverGroups = list(
    list(
      serverGroupId = "string",
      name = "string",
      serverList = list(
        list(
          serverId = "string",
          serverType = "VIRTUAL_MACHINE",
          vmServer = list(
            vmServerAddress = list(
              vmManagerId = "string",
              vmId = "string"
            ),
            vmName = "string",
            vmManagerName = "string",
            vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
            vmPath = "string"
          ),
          replicationJobId = "string",
          replicationJobTerminated = TRUE|FALSE
        )
      )
    )
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates an application. An application consists of one or more server
groups. Each server group contain one or more servers.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_app(
  name = "string",
  description = "string",
  roleName = "string",
  clientToken = "string",
  serverGroups = list(
    list(
      serverGroupId = "string",
      name = "string",
      serverList = list(
        list(
          serverId = "string",
          serverType = "VIRTUAL_MACHINE",
          vmServer = list(
            vmServerAddress = list(
              vmManagerId = "string",
              vmId = "string"
            ),
            vmName = "string",
            vmManagerName = "string",
            vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
            vmPath = "string"
          ),
          replicationJobId = "string",
          replicationJobTerminated = TRUE|FALSE
        )
      )
    )
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
