% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_create_replication_job}
\alias{sms_create_replication_job}
\title{Creates a replication job}
\usage{
sms_create_replication_job(serverId, seedReplicationTime, frequency,
  runOnce, licenseType, roleName, description, numberOfRecentAmisToKeep,
  encrypted, kmsKeyId)
}
\arguments{
\item{serverId}{[required] The ID of the server.}

\item{seedReplicationTime}{[required] The seed replication time.}

\item{frequency}{The time between consecutive replication runs, in hours.}

\item{runOnce}{Indicates whether to run the replication job one time.}

\item{licenseType}{The license type to be used for the AMI created by a successful
replication run.}

\item{roleName}{The name of the IAM role to be used by the AWS SMS.}

\item{description}{The description of the replication job.}

\item{numberOfRecentAmisToKeep}{The maximum number of SMS-created AMIs to retain. The oldest is deleted
after the maximum number is reached and a new AMI is created.}

\item{encrypted}{Indicates whether the replication job produces encrypted AMIs.}

\item{kmsKeyId}{The ID of the KMS key for replication jobs that produce encrypted AMIs.
This value can be any of the following:
\itemize{
\item KMS key ID
\item KMS key alias
\item ARN referring to the KMS key ID
\item ARN referring to the KMS key alias
}

If encrypted is \emph{true} but a KMS key ID is not specified, the customer's
default KMS key for Amazon EBS is used.}
}
\value{
A list with the following syntax:\preformatted{list(
  replicationJobId = "string"
)
}
}
\description{
Creates a replication job. The replication job schedules periodic
replication runs to replicate your server to AWS. Each replication run
creates an Amazon Machine Image (AMI).
}
\section{Request syntax}{
\preformatted{svc$create_replication_job(
  serverId = "string",
  seedReplicationTime = as.POSIXct(
    "2015-01-01"
  ),
  frequency = 123,
  runOnce = TRUE|FALSE,
  licenseType = "AWS"|"BYOL",
  roleName = "string",
  description = "string",
  numberOfRecentAmisToKeep = 123,
  encrypted = TRUE|FALSE,
  kmsKeyId = "string"
)
}
}

\keyword{internal}
