% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_delete_app}
\alias{sms_delete_app}
\title{Deletes the specified application}
\usage{
sms_delete_app(appId, forceStopAppReplication, forceTerminateApp)
}
\arguments{
\item{appId}{The ID of the application.}

\item{forceStopAppReplication}{Indicates whether to stop all replication jobs corresponding to the
servers in the application while deleting the application.}

\item{forceTerminateApp}{Indicates whether to terminate the stack corresponding to the
application while deleting the application.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes the specified application. Optionally deletes the launched stack
associated with the application and all AWS SMS replication jobs for
servers in the application.
}
\section{Request syntax}{
\preformatted{svc$delete_app(
  appId = "string",
  forceStopAppReplication = TRUE|FALSE,
  forceTerminateApp = TRUE|FALSE
)
}
}

\keyword{internal}
