% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_delete_app_launch_configuration}
\alias{sms_delete_app_launch_configuration}
\title{Deletes the launch configuration for the specified application}
\usage{
sms_delete_app_launch_configuration(appId)
}
\arguments{
\item{appId}{The ID of the application.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes the launch configuration for the specified application.
}
\section{Request syntax}{
\preformatted{svc$delete_app_launch_configuration(
  appId = "string"
)
}
}

\keyword{internal}
