% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_delete_server_catalog}
\alias{sms_delete_server_catalog}
\title{Deletes all servers from your server catalog}
\usage{
sms_delete_server_catalog()
}
\value{
An empty list.
}
\description{
Deletes all servers from your server catalog.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_server_catalog()
}\if{html}{\out{</div>}}
}

\keyword{internal}
