% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_disassociate_connector}
\alias{sms_disassociate_connector}
\title{Disassociates the specified connector from AWS SMS}
\usage{
sms_disassociate_connector(connectorId)
}
\arguments{
\item{connectorId}{[required] The identifier of the connector.}
}
\description{
Disassociates the specified connector from AWS SMS.
}
\details{
After you disassociate a connector, it is no longer available to support
replication jobs.
}
\section{Request syntax}{
\preformatted{svc$disassociate_connector(
  connectorId = "string"
)
}
}

\keyword{internal}
