% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_generate_change_set}
\alias{sms_generate_change_set}
\title{Generates a target change set for a currently launched stack and writes
it to an Amazon S3 object in the customer’s Amazon S3 bucket}
\usage{
sms_generate_change_set(appId, changesetFormat)
}
\arguments{
\item{appId}{The ID of the application associated with the change set.}

\item{changesetFormat}{The format for the change set.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  s3Location = list(
    bucket = "string",
    key = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Generates a target change set for a currently launched stack and writes
it to an Amazon S3 object in the customer’s Amazon S3 bucket.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$generate_change_set(
  appId = "string",
  changesetFormat = "JSON"|"YAML"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
