% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_generate_template}
\alias{sms_generate_template}
\title{Generates an Amazon CloudFormation template based on the current launch
configuration and writes it to an Amazon S3 object in the customer’s
Amazon S3 bucket}
\usage{
sms_generate_template(appId, templateFormat)
}
\arguments{
\item{appId}{ID of the application associated with the Amazon CloudFormation
template.}

\item{templateFormat}{Format for generating the Amazon CloudFormation template.}
}
\description{
Generates an Amazon CloudFormation template based on the current launch
configuration and writes it to an Amazon S3 object in the customer's
Amazon S3 bucket.
}
\section{Request syntax}{
\preformatted{svc$generate_template(
  appId = "string",
  templateFormat = "JSON"|"YAML"
)
}
}

\keyword{internal}
