% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_app}
\alias{sms_get_app}
\title{Retrieve information about the specified application}
\usage{
sms_get_app(appId)
}
\arguments{
\item{appId}{The ID of the application.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  appSummary = list(
    appId = "string",
    importedAppId = "string",
    name = "string",
    description = "string",
    status = "CREATING"|"ACTIVE"|"UPDATING"|"DELETING"|"DELETED"|"DELETE_FAILED",
    statusMessage = "string",
    replicationConfigurationStatus = "NOT_CONFIGURED"|"CONFIGURED",
    replicationStatus = "READY_FOR_CONFIGURATION"|"CONFIGURATION_IN_PROGRESS"|"CONFIGURATION_INVALID"|"READY_FOR_REPLICATION"|"VALIDATION_IN_PROGRESS"|"REPLICATION_PENDING"|"REPLICATION_IN_PROGRESS"|"REPLICATED"|"PARTIALLY_REPLICATED"|"DELTA_REPLICATION_IN_PROGRESS"|"DELTA_REPLICATED"|"DELTA_REPLICATION_FAILED"|"REPLICATION_FAILED"|"REPLICATION_STOPPING"|"REPLICATION_STOP_FAILED"|"REPLICATION_STOPPED",
    replicationStatusMessage = "string",
    latestReplicationTime = as.POSIXct(
      "2015-01-01"
    ),
    launchConfigurationStatus = "NOT_CONFIGURED"|"CONFIGURED",
    launchStatus = "READY_FOR_CONFIGURATION"|"CONFIGURATION_IN_PROGRESS"|"CONFIGURATION_INVALID"|"READY_FOR_LAUNCH"|"VALIDATION_IN_PROGRESS"|"LAUNCH_PENDING"|"LAUNCH_IN_PROGRESS"|"LAUNCHED"|"PARTIALLY_LAUNCHED"|"DELTA_LAUNCH_IN_PROGRESS"|"DELTA_LAUNCH_FAILED"|"LAUNCH_FAILED"|"TERMINATE_IN_PROGRESS"|"TERMINATE_FAILED"|"TERMINATED",
    launchStatusMessage = "string",
    launchDetails = list(
      latestLaunchTime = as.POSIXct(
        "2015-01-01"
      ),
      stackName = "string",
      stackId = "string"
    ),
    creationTime = as.POSIXct(
      "2015-01-01"
    ),
    lastModified = as.POSIXct(
      "2015-01-01"
    ),
    roleName = "string",
    totalServerGroups = 123,
    totalServers = 123
  ),
  serverGroups = list(
    list(
      serverGroupId = "string",
      name = "string",
      serverList = list(
        list(
          serverId = "string",
          serverType = "VIRTUAL_MACHINE",
          vmServer = list(
            vmServerAddress = list(
              vmManagerId = "string",
              vmId = "string"
            ),
            vmName = "string",
            vmManagerName = "string",
            vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
            vmPath = "string"
          ),
          replicationJobId = "string",
          replicationJobTerminated = TRUE|FALSE
        )
      )
    )
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieve information about the specified application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_app(
  appId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
