% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_app_validation_configuration}
\alias{sms_get_app_validation_configuration}
\title{Retrieves information about a configuration for validating an
application}
\usage{
sms_get_app_validation_configuration(appId)
}
\arguments{
\item{appId}{[required] The ID of the application.}
}
\description{
Retrieves information about a configuration for validating an
application.
}
\section{Request syntax}{
\preformatted{svc$get_app_validation_configuration(
  appId = "string"
)
}
}

\keyword{internal}
