% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_app_validation_output}
\alias{sms_get_app_validation_output}
\title{Retrieves output from validating an application}
\usage{
sms_get_app_validation_output(appId)
}
\arguments{
\item{appId}{[required] The ID of the application.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  validationOutputList = list(
    list(
      validationId = "string",
      name = "string",
      status = "READY_FOR_VALIDATION"|"PENDING"|"IN_PROGRESS"|"SUCCEEDED"|"FAILED",
      statusMessage = "string",
      latestValidationTime = as.POSIXct(
        "2015-01-01"
      ),
      appValidationOutput = list(
        ssmOutput = list(
          s3Location = list(
            bucket = "string",
            key = "string"
          )
        )
      ),
      serverValidationOutput = list(
        server = list(
          serverId = "string",
          serverType = "VIRTUAL_MACHINE",
          vmServer = list(
            vmServerAddress = list(
              vmManagerId = "string",
              vmId = "string"
            ),
            vmName = "string",
            vmManagerName = "string",
            vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
            vmPath = "string"
          ),
          replicationJobId = "string",
          replicationJobTerminated = TRUE|FALSE
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves output from validating an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_app_validation_output(
  appId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
