% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_replication_runs}
\alias{sms_get_replication_runs}
\title{Describes the replication runs for the specified replication job}
\usage{
sms_get_replication_runs(replicationJobId, nextToken, maxResults)
}
\arguments{
\item{replicationJobId}{[required] The ID of the replication job.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in a single call. The default
value is 50. To retrieve the remaining results, make another call with
the returned \code{NextToken} value.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  replicationJob = list(
    replicationJobId = "string",
    serverId = "string",
    serverType = "VIRTUAL_MACHINE",
    vmServer = list(
      vmServerAddress = list(
        vmManagerId = "string",
        vmId = "string"
      ),
      vmName = "string",
      vmManagerName = "string",
      vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
      vmPath = "string"
    ),
    seedReplicationTime = as.POSIXct(
      "2015-01-01"
    ),
    frequency = 123,
    runOnce = TRUE|FALSE,
    nextReplicationRunStartTime = as.POSIXct(
      "2015-01-01"
    ),
    licenseType = "AWS"|"BYOL",
    roleName = "string",
    latestAmiId = "string",
    state = "PENDING"|"ACTIVE"|"FAILED"|"DELETING"|"DELETED"|"COMPLETED"|"PAUSED_ON_FAILURE"|"FAILING",
    statusMessage = "string",
    description = "string",
    numberOfRecentAmisToKeep = 123,
    encrypted = TRUE|FALSE,
    kmsKeyId = "string",
    replicationRunList = list(
      list(
        replicationRunId = "string",
        state = "PENDING"|"MISSED"|"ACTIVE"|"FAILED"|"COMPLETED"|"DELETING"|"DELETED",
        type = "ON_DEMAND"|"AUTOMATIC",
        stageDetails = list(
          stage = "string",
          stageProgress = "string"
        ),
        statusMessage = "string",
        amiId = "string",
        scheduledStartTime = as.POSIXct(
          "2015-01-01"
        ),
        completedTime = as.POSIXct(
          "2015-01-01"
        ),
        description = "string",
        encrypted = TRUE|FALSE,
        kmsKeyId = "string"
      )
    )
  ),
  replicationRunList = list(
    list(
      replicationRunId = "string",
      state = "PENDING"|"MISSED"|"ACTIVE"|"FAILED"|"COMPLETED"|"DELETING"|"DELETED",
      type = "ON_DEMAND"|"AUTOMATIC",
      stageDetails = list(
        stage = "string",
        stageProgress = "string"
      ),
      statusMessage = "string",
      amiId = "string",
      scheduledStartTime = as.POSIXct(
        "2015-01-01"
      ),
      completedTime = as.POSIXct(
        "2015-01-01"
      ),
      description = "string",
      encrypted = TRUE|FALSE,
      kmsKeyId = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes the replication runs for the specified replication job.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_replication_runs(
  replicationJobId = "string",
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
