% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_servers}
\alias{sms_get_servers}
\title{Describes the servers in your server catalog}
\usage{
sms_get_servers(nextToken, maxResults, vmServerAddressList)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in a single call. The default
value is 50. To retrieve the remaining results, make another call with
the returned \code{NextToken} value.}

\item{vmServerAddressList}{The server addresses.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  lastModifiedOn = as.POSIXct(
    "2015-01-01"
  ),
  serverCatalogStatus = "NOT_IMPORTED"|"IMPORTING"|"AVAILABLE"|"DELETED"|"EXPIRED",
  serverList = list(
    list(
      serverId = "string",
      serverType = "VIRTUAL_MACHINE",
      vmServer = list(
        vmServerAddress = list(
          vmManagerId = "string",
          vmId = "string"
        ),
        vmName = "string",
        vmManagerName = "string",
        vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
        vmPath = "string"
      ),
      replicationJobId = "string",
      replicationJobTerminated = TRUE|FALSE
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes the servers in your server catalog.

Before you can describe your servers, you must import them using
\code{\link[=sms_import_server_catalog]{import_server_catalog}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_servers(
  nextToken = "string",
  maxResults = 123,
  vmServerAddressList = list(
    list(
      vmManagerId = "string",
      vmId = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
