% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_import_app_catalog}
\alias{sms_import_app_catalog}
\title{Allows application import from AWS Migration Hub}
\usage{
sms_import_app_catalog(roleName)
}
\arguments{
\item{roleName}{The name of the service role. If you omit this parameter, we create a
service-linked role for AWS Migration Hub in your account. Otherwise,
the role that you provide must have the \href{https://docs.aws.amazon.com/migrationhub/latest/ug/new-customer-setup.html#sms-managed}{policy and trust policy}
described in the \emph{AWS Migration Hub User Guide}.}
}
\value{
An empty list.
}
\description{
Allows application import from AWS Migration Hub.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$import_app_catalog(
  roleName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
