% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_import_server_catalog}
\alias{sms_import_server_catalog}
\title{Gathers a complete list of on-premises servers}
\usage{
sms_import_server_catalog()
}
\value{
An empty list.
}
\description{
Gathers a complete list of on-premises servers. Connectors must be
installed and monitoring all servers to import.

This call returns immediately, but might take additional time to
retrieve all the servers.
}
\section{Request syntax}{
\preformatted{svc$import_server_catalog()
}
}

\keyword{internal}
