% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_launch_app}
\alias{sms_launch_app}
\title{Launches the specified application as a stack in AWS CloudFormation}
\usage{
sms_launch_app(appId)
}
\arguments{
\item{appId}{The ID of the application.}
}
\value{
An empty list.
}
\description{
Launches the specified application as a stack in AWS CloudFormation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$launch_app(
  appId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
