% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_put_app_launch_configuration}
\alias{sms_put_app_launch_configuration}
\title{Creates or updates the launch configuration for the specified
application}
\usage{
sms_put_app_launch_configuration(appId, roleName, autoLaunch,
  serverGroupLaunchConfigurations)
}
\arguments{
\item{appId}{The ID of the application.}

\item{roleName}{The name of service role in the customer's account that AWS
CloudFormation uses to launch the application.}

\item{autoLaunch}{Indicates whether the application is configured to launch automatically
after replication is complete.}

\item{serverGroupLaunchConfigurations}{Information about the launch configurations for server groups in the
application.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Creates or updates the launch configuration for the specified
application.
}
\section{Request syntax}{
\preformatted{svc$put_app_launch_configuration(
  appId = "string",
  roleName = "string",
  autoLaunch = TRUE|FALSE,
  serverGroupLaunchConfigurations = list(
    list(
      serverGroupId = "string",
      launchOrder = 123,
      serverLaunchConfigurations = list(
        list(
          server = list(
            serverId = "string",
            serverType = "VIRTUAL_MACHINE",
            vmServer = list(
              vmServerAddress = list(
                vmManagerId = "string",
                vmId = "string"
              ),
              vmName = "string",
              vmManagerName = "string",
              vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
              vmPath = "string"
            ),
            replicationJobId = "string",
            replicationJobTerminated = TRUE|FALSE
          ),
          logicalId = "string",
          vpc = "string",
          subnet = "string",
          securityGroup = "string",
          ec2KeyName = "string",
          userData = list(
            s3Location = list(
              bucket = "string",
              key = "string"
            )
          ),
          instanceType = "string",
          associatePublicIpAddress = TRUE|FALSE,
          iamInstanceProfileName = "string",
          configureScript = list(
            bucket = "string",
            key = "string"
          ),
          configureScriptType = "SHELL_SCRIPT"|"POWERSHELL_SCRIPT"
        )
      )
    )
  )
)
}
}

\keyword{internal}
