% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_put_app_validation_configuration}
\alias{sms_put_app_validation_configuration}
\title{Creates or updates a validation configuration for the specified
application}
\usage{
sms_put_app_validation_configuration(appId, appValidationConfigurations,
  serverGroupValidationConfigurations)
}
\arguments{
\item{appId}{[required] The ID of the application.}

\item{appValidationConfigurations}{The configuration for application validation.}

\item{serverGroupValidationConfigurations}{The configuration for instance validation.}
}
\description{
Creates or updates a validation configuration for the specified
application.
}
\section{Request syntax}{
\preformatted{svc$put_app_validation_configuration(
  appId = "string",
  appValidationConfigurations = list(
    list(
      validationId = "string",
      name = "string",
      appValidationStrategy = "SSM",
      ssmValidationParameters = list(
        source = list(
          s3Location = list(
            bucket = "string",
            key = "string"
          )
        ),
        instanceId = "string",
        scriptType = "SHELL_SCRIPT"|"POWERSHELL_SCRIPT",
        command = "string",
        executionTimeoutSeconds = 123,
        outputS3BucketName = "string"
      )
    )
  ),
  serverGroupValidationConfigurations = list(
    list(
      serverGroupId = "string",
      serverValidationConfigurations = list(
        list(
          server = list(
            serverId = "string",
            serverType = "VIRTUAL_MACHINE",
            vmServer = list(
              vmServerAddress = list(
                vmManagerId = "string",
                vmId = "string"
              ),
              vmName = "string",
              vmManagerName = "string",
              vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
              vmPath = "string"
            ),
            replicationJobId = "string",
            replicationJobTerminated = TRUE|FALSE
          ),
          validationId = "string",
          name = "string",
          serverValidationStrategy = "USERDATA",
          userDataValidationParameters = list(
            source = list(
              s3Location = list(
                bucket = "string",
                key = "string"
              )
            ),
            scriptType = "SHELL_SCRIPT"|"POWERSHELL_SCRIPT"
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
