% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_start_on_demand_replication_run}
\alias{sms_start_on_demand_replication_run}
\title{Starts an on-demand replication run for the specified replication job}
\usage{
sms_start_on_demand_replication_run(replicationJobId, description)
}
\arguments{
\item{replicationJobId}{[required] The ID of the replication job.}

\item{description}{The description of the replication run.}
}
\value{
A list with the following syntax:\preformatted{list(
  replicationRunId = "string"
)
}
}
\description{
Starts an on-demand replication run for the specified replication job.
This replication run starts immediately. This replication run is in
addition to the ones already scheduled.

There is a limit on the number of on-demand replications runs that you
can request in a 24-hour period.
}
\section{Request syntax}{
\preformatted{svc$start_on_demand_replication_run(
  replicationJobId = "string",
  description = "string"
)
}
}

\keyword{internal}
