% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_stop_app_replication}
\alias{sms_stop_app_replication}
\title{Stops replicating the specified application by deleting the replication
job for each server in the application}
\usage{
sms_stop_app_replication(appId)
}
\arguments{
\item{appId}{The ID of the application.}
}
\value{
An empty list.
}
\description{
Stops replicating the specified application by deleting the replication
job for each server in the application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_app_replication(
  appId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
