% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_terminate_app}
\alias{sms_terminate_app}
\title{Terminates the stack for the specified application}
\usage{
sms_terminate_app(appId)
}
\arguments{
\item{appId}{The ID of the application.}
}
\value{
An empty list.
}
\description{
Terminates the stack for the specified application.
}
\section{Request syntax}{
\preformatted{svc$terminate_app(
  appId = "string"
)
}
}

\keyword{internal}
