% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_cancel_cluster}
\alias{snowball_cancel_cluster}
\title{Cancels a cluster job}
\usage{
snowball_cancel_cluster(ClusterId)
}
\arguments{
\item{ClusterId}{[required] The 39-character ID for the cluster that you want to cancel, for example
\verb{CID123e4567-e89b-12d3-a456-426655440000}.}
}
\value{
An empty list.
}
\description{
Cancels a cluster job. You can only cancel a cluster job while it's in
the \code{AwaitingQuorum} status. You'll have at least an hour after creating
a cluster job to cancel it.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$cancel_cluster(
  ClusterId = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# This operation cancels a cluster job. You can only cancel a cluster job
# while it's in the AwaitingQuorum status.
svc$cancel_cluster(
  ClusterId = "CID123e4567-e89b-12d3-a456-426655440000"
)
}

}
\keyword{internal}
