% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_cancel_job}
\alias{snowball_cancel_job}
\title{Cancels the specified job}
\usage{
snowball_cancel_job(JobId)
}
\arguments{
\item{JobId}{[required] The 39-character job ID for the job that you want to cancel, for example
\verb{JID123e4567-e89b-12d3-a456-426655440000}.}
}
\value{
An empty list.
}
\description{
Cancels the specified job. You can only cancel a job before its
\code{JobState} value changes to \code{PreparingAppliance}. Requesting the
\code{\link[=snowball_list_jobs]{list_jobs}} or
\code{\link[=snowball_describe_job]{describe_job}} action returns a job's
\code{JobState} as part of the response element data returned.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$cancel_job(
  JobId = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# This operation cancels a job. You can only cancel a job before its
# JobState value changes to PreparingAppliance.
svc$cancel_job(
  JobId = "JID123e4567-e89b-12d3-a456-426655440000"
)
}

}
\keyword{internal}
