% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_create_job}
\alias{snowball_create_job}
\title{Creates a job to import or export data between Amazon S3 and your
on-premises data center}
\usage{
snowball_create_job(JobType, Resources, Description, AddressId,
  KmsKeyARN, RoleARN, SnowballCapacityPreference, ShippingOption,
  Notification, ClusterId, SnowballType, ForwardingAddressId,
  TaxDocuments, DeviceConfiguration)
}
\arguments{
\item{JobType}{Defines the type of job that you're creating.}

\item{Resources}{Defines the Amazon S3 buckets associated with this job.

With \code{IMPORT} jobs, you specify the bucket or buckets that your
transferred data will be imported into.

With \code{EXPORT} jobs, you specify the bucket or buckets that your
transferred data will be exported from. Optionally, you can also specify
a \code{KeyRange} value. If you choose to export a range, you define the
length of the range by providing either an inclusive \code{BeginMarker}
value, an inclusive \code{EndMarker} value, or both. Ranges are UTF-8 binary
sorted.}

\item{Description}{Defines an optional description of this specific job, for example
\verb{Important Photos 2016-08-11}.}

\item{AddressId}{The ID for the address that you want the Snow device shipped to.}

\item{KmsKeyARN}{The \code{KmsKeyARN} that you want to associate with this job. \code{KmsKeyARN}s
are created using the
\href{https://docs.aws.amazon.com/kms/latest/APIReference/API_CreateKey.html}{CreateKey}
AWS Key Management Service (KMS) API action.}

\item{RoleARN}{The \code{RoleARN} that you want to associate with this job. \code{RoleArn}s are
created using the
\href{https://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateRole.html}{CreateRole}
AWS Identity and Access Management (IAM) API action.}

\item{SnowballCapacityPreference}{If your job is being created in one of the US regions, you have the
option of specifying what size Snow device you'd like for this job. In
all other regions, Snowballs come with 80 TB in storage capacity.}

\item{ShippingOption}{The shipping speed for this job. This speed doesn't dictate how soon
you'll get the Snow device, rather it represents how quickly the Snow
device moves to its destination while in transit. Regional shipping
speeds are as follows:
\itemize{
\item In Australia, you have access to express shipping. Typically, Snow
devices shipped express are delivered in about a day.
\item In the European Union (EU), you have access to express shipping.
Typically, Snow devices shipped express are delivered in about a
day. In addition, most countries in the EU have access to standard
shipping, which typically takes less than a week, one way.
\item In India, Snow devices are delivered in one to seven days.
\item In the US, you have access to one-day shipping and two-day shipping.
}}

\item{Notification}{Defines the Amazon Simple Notification Service (Amazon SNS) notification
settings for this job.}

\item{ClusterId}{The ID of a cluster. If you're creating a job for a node in a cluster,
you need to provide only this \code{clusterId} value. The other job
attributes are inherited from the cluster.}

\item{SnowballType}{The type of AWS Snow Family device to use for this job.

For cluster jobs, AWS Snow Family currently supports only the \code{EDGE}
device type.

The type of AWS Snow device to use for this job. Currently, the only
supported device type for cluster jobs is \code{EDGE}.

For more information, see \href{https://docs.aws.amazon.com/snowball/latest/developer-guide/device-differences.html}{Snowball Edge Device Options}
in the Snowball Edge Developer Guide.}

\item{ForwardingAddressId}{The forwarding address ID for a job. This field is not supported in most
regions.}

\item{TaxDocuments}{The tax documents required in your AWS Region.}

\item{DeviceConfiguration}{Defines the device configuration for an AWS Snowcone job.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  JobId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a job to import or export data between Amazon S3 and your
on-premises data center. Your AWS account must have the right trust
policies and permissions in place to create a job for a Snow device. If
you're creating a job for a node in a cluster, you only need to provide
the \code{clusterId} value; the other job attributes are inherited from the
cluster.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_job(
  JobType = "IMPORT"|"EXPORT"|"LOCAL_USE",
  Resources = list(
    S3Resources = list(
      list(
        BucketArn = "string",
        KeyRange = list(
          BeginMarker = "string",
          EndMarker = "string"
        )
      )
    ),
    LambdaResources = list(
      list(
        LambdaArn = "string",
        EventTriggers = list(
          list(
            EventResourceARN = "string"
          )
        )
      )
    ),
    Ec2AmiResources = list(
      list(
        AmiId = "string",
        SnowballAmiId = "string"
      )
    )
  ),
  Description = "string",
  AddressId = "string",
  KmsKeyARN = "string",
  RoleARN = "string",
  SnowballCapacityPreference = "T50"|"T80"|"T100"|"T42"|"T98"|"T8"|"NoPreference",
  ShippingOption = "SECOND_DAY"|"NEXT_DAY"|"EXPRESS"|"STANDARD",
  Notification = list(
    SnsTopicARN = "string",
    JobStatesToNotify = list(
      "New"|"PreparingAppliance"|"PreparingShipment"|"InTransitToCustomer"|"WithCustomer"|"InTransitToAWS"|"WithAWSSortingFacility"|"WithAWS"|"InProgress"|"Complete"|"Cancelled"|"Listing"|"Pending"
    ),
    NotifyAll = TRUE|FALSE
  ),
  ClusterId = "string",
  SnowballType = "STANDARD"|"EDGE"|"EDGE_C"|"EDGE_CG"|"EDGE_S"|"SNC1_HDD",
  ForwardingAddressId = "string",
  TaxDocuments = list(
    IND = list(
      GSTIN = "string"
    )
  ),
  DeviceConfiguration = list(
    SnowconeDeviceConfiguration = list(
      WirelessConnection = list(
        IsWifiEnabled = TRUE|FALSE
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Creates a job to import or export data between Amazon S3 and your
# on-premises data center. Your AWS account must have the right trust
# policies and permissions in place to create a job for Snowball. If
# you're creating a job for a node in a cluster, you only need to provide
# the clusterId value; the other job attributes are inherited from the
# cluster.
svc$create_job(
  AddressId = "ADID1234ab12-3eec-4eb3-9be6-9374c10eb51b",
  Description = "My Job",
  JobType = "IMPORT",
  KmsKeyARN = "arn:aws:kms:us-east-1:123456789012:key/abcd1234-12ab-34cd-56ef-123456123456",
  Notification = list(
    JobStatesToNotify = list(),
    NotifyAll = FALSE
  ),
  Resources = list(
    S3Resources = list(
      list(
        BucketArn = "arn:aws:s3:::MyBucket",
        KeyRange = structure(
          list(),
          names = character(
            0
          )
        )
      )
    )
  ),
  RoleARN = "arn:aws:iam::123456789012:role/snowball-import-S3-role",
  ShippingOption = "SECOND_DAY",
  SnowballCapacityPreference = "T80",
  SnowballType = "STANDARD"
)
}

}
\keyword{internal}
