% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_create_long_term_pricing}
\alias{snowball_create_long_term_pricing}
\title{Creates a job with long term usage option for a device}
\usage{
snowball_create_long_term_pricing(LongTermPricingType,
  IsLongTermPricingAutoRenew, SnowballType)
}
\arguments{
\item{LongTermPricingType}{[required] The type of long term pricing option you want for the device - one year
or three year long term pricing.}

\item{IsLongTermPricingAutoRenew}{Specifies whether the current long term pricing type for the device
should be renewed.}

\item{SnowballType}{The type of AWS Snow Family device to use for the long term pricing job.}
}
\value{
A list with the following syntax:\preformatted{list(
  LongTermPricingId = "string"
)
}
}
\description{
Creates a job with long term usage option for a device. The long term
usage is a one year or three year long term pricing type for the device.
You are billed upfront and AWS give discounts for long term pricing. For
detailed information see XXXXXXXX
}
\section{Request syntax}{
\preformatted{svc$create_long_term_pricing(
  LongTermPricingType = "OneYear"|"ThreeYear",
  IsLongTermPricingAutoRenew = TRUE|FALSE,
  SnowballType = "STANDARD"|"EDGE"|"EDGE_C"|"EDGE_CG"|"EDGE_S"|"SNC1_HDD"|"SNC1_SSD"
)
}
}

\keyword{internal}
