% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_create_return_shipping_label}
\alias{snowball_create_return_shipping_label}
\title{Creates a shipping label that will be used to return the Snow device to
AWS}
\usage{
snowball_create_return_shipping_label(JobId, ShippingOption)
}
\arguments{
\item{JobId}{[required] The ID for a job that you want to create the return shipping label for.
For example \verb{JID123e4567-e89b-12d3-a456-426655440000}.}

\item{ShippingOption}{The shipping speed for a particular job. This speed doesn't dictate how
soon the device is returned to AWS. This speed represents how quickly it
moves to its destination while in transit. Regional shipping speeds are
as follows:}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Status = "InProgress"|"TimedOut"|"Succeeded"|"Failed"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a shipping label that will be used to return the Snow device to
AWS.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_return_shipping_label(
  JobId = "string",
  ShippingOption = "SECOND_DAY"|"NEXT_DAY"|"EXPRESS"|"STANDARD"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
