% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_describe_address}
\alias{snowball_describe_address}
\title{Takes an AddressId and returns specific details about that address in
the form of an Address object}
\usage{
snowball_describe_address(AddressId)
}
\arguments{
\item{AddressId}{[required] The automatically generated ID for a specific address.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Address = list(
    AddressId = "string",
    Name = "string",
    Company = "string",
    Street1 = "string",
    Street2 = "string",
    Street3 = "string",
    City = "string",
    StateOrProvince = "string",
    PrefectureOrDistrict = "string",
    Landmark = "string",
    Country = "string",
    PostalCode = "string",
    PhoneNumber = "string",
    IsRestricted = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Takes an \code{AddressId} and returns specific details about that address in
the form of an \code{Address} object.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_address(
  AddressId = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# This operation describes an address for a job.
svc$describe_address(
  AddressId = "ADID1234ab12-3eec-4eb3-9be6-9374c10eb51b"
)
}

}
\keyword{internal}
