% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_get_job_unlock_code}
\alias{snowball_get_job_unlock_code}
\title{Returns the UnlockCode code value for the specified job}
\usage{
snowball_get_job_unlock_code(JobId)
}
\arguments{
\item{JobId}{[required] The ID for the job that you want to get the \code{UnlockCode} value for, for
example \verb{JID123e4567-e89b-12d3-a456-426655440000}.}
}
\description{
Returns the \code{UnlockCode} code value for the specified job. A particular
\code{UnlockCode} value can be accessed for up to 90 days after the
associated job has been created.
}
\details{
The \code{UnlockCode} value is a 29-character code with 25 alphanumeric
characters and 4 hyphens. This code is used to decrypt the manifest file
when it is passed along with the manifest to the Snowball through the
Snowball client when the client is started for the first time.

As a best practice, we recommend that you don\'t save a copy of the
\code{UnlockCode} in the same location as the manifest file for that job.
Saving these separately helps prevent unauthorized parties from gaining
access to the Snowball associated with that job.
}
\section{Request syntax}{
\preformatted{svc$get_job_unlock_code(
  JobId = "string"
)
}
}

\section{Examples}{
\preformatted{# Returns the UnlockCode code value for the specified job. A particular
# UnlockCode value can be accessed for up to 90 days after the associated
# job has been created.
# 
# The UnlockCode value is a 29-character code with
# 25 alphanumeric characters and 4 hyphens. This code is used to decrypt
# the manifest file when it is passed along with the manifest to the
# Snowball through the Snowball client when the client is started for the
# first time.
# 
# As a best practice, we recommend that you don't save a copy
# of the UnlockCode in the same location as the manifest file for that
# job. Saving these separately helps prevent unauthorized parties from
# gaining access to the Snowball associated with that job.
svc$get_job_unlock_code(
  JobId = "JID123e4567-e89b-12d3-a456-426655440000"
)
}
}

\keyword{internal}
