% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_get_software_updates}
\alias{snowball_get_software_updates}
\title{Returns an Amazon S3 presigned URL for an update file associated with a
specified JobId}
\usage{
snowball_get_software_updates(JobId)
}
\arguments{
\item{JobId}{[required] The ID for a job that you want to get the software update file for, for
example \verb{JID123e4567-e89b-12d3-a456-426655440000}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UpdatesURI = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns an Amazon S3 presigned URL for an update file associated with a
specified \code{JobId}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_software_updates(
  JobId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
