% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_list_clusters}
\alias{snowball_list_clusters}
\title{Returns an array of ClusterListEntry objects of the specified length}
\usage{
snowball_list_clusters(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The number of \code{ClusterListEntry} objects to return.}

\item{NextToken}{HTTP requests are stateless. To identify what object comes "next" in the
list of \code{ClusterListEntry} objects, you have the option of specifying
\code{NextToken} as the starting point for your returned list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ClusterListEntries = list(
    list(
      ClusterId = "string",
      ClusterState = "AwaitingQuorum"|"Pending"|"InUse"|"Complete"|"Cancelled",
      CreationDate = as.POSIXct(
        "2015-01-01"
      ),
      Description = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns an array of \code{ClusterListEntry} objects of the specified length.
Each \code{ClusterListEntry} object contains a cluster's state, a cluster's
ID, and other important status information.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_clusters(
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Returns an array of ClusterListEntry objects of the specified length.
# Each ClusterListEntry object contains a cluster's state, a cluster's ID,
# and other important status information.
svc$list_clusters()
}

}
\keyword{internal}
