% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_list_jobs}
\alias{snowball_list_jobs}
\title{Returns an array of JobListEntry objects of the specified length}
\usage{
snowball_list_jobs(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The number of \code{JobListEntry} objects to return.}

\item{NextToken}{HTTP requests are stateless. To identify what object comes "next" in the
list of \code{JobListEntry} objects, you have the option of specifying
\code{NextToken} as the starting point for your returned list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  JobListEntries = list(
    list(
      JobId = "string",
      JobState = "New"|"PreparingAppliance"|"PreparingShipment"|"InTransitToCustomer"|"WithCustomer"|"InTransitToAWS"|"WithAWSSortingFacility"|"WithAWS"|"InProgress"|"Complete"|"Cancelled"|"Listing"|"Pending",
      IsMaster = TRUE|FALSE,
      JobType = "IMPORT"|"EXPORT"|"LOCAL_USE",
      SnowballType = "STANDARD"|"EDGE"|"EDGE_C"|"EDGE_CG"|"EDGE_S"|"SNC1_HDD",
      CreationDate = as.POSIXct(
        "2015-01-01"
      ),
      Description = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns an array of \code{JobListEntry} objects of the specified length. Each
\code{JobListEntry} object contains a job's state, a job's ID, and a value
that indicates whether the job is a job part, in the case of export
jobs. Calling this API action in one of the US regions will return jobs
from the list of all jobs associated with this account in all US
regions.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_jobs(
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Returns an array of JobListEntry objects of the specified length. Each
# JobListEntry object contains a job's state, a job's ID, and a value that
# indicates whether the job is a job part, in the case of export jobs.
# Calling this API action in one of the US regions will return jobs from
# the list of all jobs associated with this account in all US regions.
svc$list_jobs()
}

}
\keyword{internal}
