% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_list_long_term_pricing}
\alias{snowball_list_long_term_pricing}
\title{Lists all long term pricing types}
\usage{
snowball_list_long_term_pricing(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of
\code{\link[=snowball_list_long_term_pricing]{list_long_term_pricing}} objects to
return.}

\item{NextToken}{Because HTTP requests are stateless, this is the starting point for your
next list of \code{\link[=snowball_list_long_term_pricing]{list_long_term_pricing}}
to return.}
}
\value{
A list with the following syntax:\preformatted{list(
  LongTermPricingEntries = list(
    list(
      LongTermPricingId = "string",
      LongTermPricingEndDate = as.POSIXct(
        "2015-01-01"
      ),
      LongTermPricingStartDate = as.POSIXct(
        "2015-01-01"
      ),
      LongTermPricingType = "OneYear"|"ThreeYear",
      CurrentActiveJob = "string",
      ReplacementJob = "string",
      IsLongTermPricingAutoRenew = TRUE|FALSE,
      LongTermPricingStatus = "string",
      SnowballType = "STANDARD"|"EDGE"|"EDGE_C"|"EDGE_CG"|"EDGE_S"|"SNC1_HDD"|"SNC1_SSD",
      JobIds = list(
        "string"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists all long term pricing types.
}
\section{Request syntax}{
\preformatted{svc$list_long_term_pricing(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
