% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_update_long_term_pricing}
\alias{snowball_update_long_term_pricing}
\title{Updates the long term pricing type}
\usage{
snowball_update_long_term_pricing(LongTermPricingId, ReplacementJob,
  IsLongTermPricingAutoRenew)
}
\arguments{
\item{LongTermPricingId}{[required] The ID of the long term pricing type for the device.}

\item{ReplacementJob}{Specifies that a device that is ordered with long term pricing should be
replaced with a new device.}

\item{IsLongTermPricingAutoRenew}{If set to \code{true}, specifies that the current long term pricing type for
the device should be automatically renewed before the long term pricing
contract expires.}
}
\value{
An empty list.
}
\description{
Updates the long term pricing type.
}
\section{Request syntax}{
\preformatted{svc$update_long_term_pricing(
  LongTermPricingId = "string",
  ReplacementJob = "string",
  IsLongTermPricingAutoRenew = TRUE|FALSE
)
}
}

\keyword{internal}
