% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_create_server}
\alias{transfer_create_server}
\title{Instantiates an autoscaling virtual server based on the selected file
transfer protocol in AWS}
\usage{
transfer_create_server(Certificate, Domain, EndpointDetails,
  EndpointType, HostKey, IdentityProviderDetails, IdentityProviderType,
  LoggingRole, Protocols, SecurityPolicyName, Tags)
}
\arguments{
\item{Certificate}{The Amazon Resource Name (ARN) of the AWS Certificate Manager (ACM)
certificate. Required when \code{Protocols} is set to \code{FTPS}.

To request a new public certificate, see \href{https://docs.aws.amazon.com/acm/latest/userguide/gs-acm-request-public.html}{Request a public certificate}
in the \emph{AWS Certificate Manager User Guide}.

To import an existing certificate into ACM, see \href{https://docs.aws.amazon.com/acm/latest/userguide/import-certificate.html}{Importing certificates into ACM}
in the \emph{AWS Certificate Manager User Guide}.

To request a private certificate to use FTPS through private IP
addresses, see \href{https://docs.aws.amazon.com/acm/latest/userguide/gs-acm-request-private.html}{Request a private certificate}
in the \emph{AWS Certificate Manager User Guide}.

Certificates with the following cryptographic algorithms and key sizes
are supported:
\itemize{
\item 2048-bit RSA (RSA_2048)
\item 4096-bit RSA (RSA_4096)
\item Elliptic Prime Curve 256 bit (EC_prime256v1)
\item Elliptic Prime Curve 384 bit (EC_secp384r1)
\item Elliptic Prime Curve 521 bit (EC_secp521r1)
}

The certificate must be a valid SSL/TLS X.509 version 3 certificate with
FQDN or IP address specified and information about the issuer.}

\item{Domain}{}

\item{EndpointDetails}{The virtual private cloud (VPC) endpoint settings that are configured
for your server. When you host your endpoint within your VPC, you can
make it accessible only to resources within your VPC, or you can attach
Elastic IPs and make it accessible to clients over the internet. Your
VPC's default security groups are automatically assigned to your
endpoint.}

\item{EndpointType}{The type of VPC endpoint that you want your server to connect to. You
can choose to connect to the public internet or a VPC endpoint. With a
VPC endpoint, you can restrict access to your server and resources only
within your VPC.

It is recommended that you use \code{VPC} as the \code{EndpointType}. With this
endpoint type, you have the option to directly associate up to three
Elastic IPv4 addresses (BYO IP included) with your server's endpoint and
use VPC security groups to restrict traffic by the client's public IP
address. This is not possible with \code{EndpointType} set to \code{VPC_ENDPOINT}.}

\item{HostKey}{The RSA private key as generated by the
\verb{ssh-keygen -N "" -m PEM -f my-new-server-key} command.

If you aren't planning to migrate existing users from an existing
SFTP-enabled server to a new server, don't update the host key.
Accidentally changing a server's host key can be disruptive.

For more information, see \href{https://docs.aws.amazon.com/transfer/latest/userguide/edit-server-config.html#configuring-servers-change-host-key}{Change the host key for your SFTP-enabled server}
in the \emph{AWS Transfer Family User Guide}.}

\item{IdentityProviderDetails}{Required when \code{IdentityProviderType} is set to \code{API_GATEWAY}. Accepts an
array containing all of the information required to call a
customer-supplied authentication API, including the API Gateway URL. Not
required when \code{IdentityProviderType} is set to \code{SERVICE_MANAGED}.}

\item{IdentityProviderType}{Specifies the mode of authentication for a server. The default value is
\code{SERVICE_MANAGED}, which allows you to store and access user credentials
within the AWS Transfer Family service. Use the \code{API_GATEWAY} value to
integrate with an identity provider of your choosing. The \code{API_GATEWAY}
setting requires you to provide an API Gateway endpoint URL to call for
authentication using the \code{IdentityProviderDetails} parameter.}

\item{LoggingRole}{Allows the service to write your users' activity to your Amazon
CloudWatch logs for monitoring and auditing purposes.}

\item{Protocols}{Specifies the file transfer protocol or protocols over which your file
transfer protocol client can connect to your server's endpoint. The
available protocols are:
\itemize{
\item \code{SFTP} (Secure Shell (SSH) File Transfer Protocol): File transfer
over SSH
\item \code{FTPS} (File Transfer Protocol Secure): File transfer with TLS
encryption
\item \code{FTP} (File Transfer Protocol): Unencrypted file transfer
}

If you select \code{FTPS}, you must choose a certificate stored in AWS
Certificate Manager (ACM) which will be used to identify your server
when clients connect to it over FTPS.

If \code{Protocol} includes either \code{FTP} or \code{FTPS}, then the \code{EndpointType}
must be \code{VPC} and the \code{IdentityProviderType} must be \code{API_GATEWAY}.

If \code{Protocol} includes \code{FTP}, then \code{AddressAllocationIds} cannot be
associated.

If \code{Protocol} is set only to \code{SFTP}, the \code{EndpointType} can be set to
\code{PUBLIC} and the \code{IdentityProviderType} can be set to \code{SERVICE_MANAGED}.}

\item{SecurityPolicyName}{Specifies the name of the security policy that is attached to the
server.}

\item{Tags}{Key-value pairs that can be used to group and search for servers.}
}
\value{
A list with the following syntax:\preformatted{list(
  ServerId = "string"
)
}
}
\description{
Instantiates an autoscaling virtual server based on the selected file
transfer protocol in AWS. When you make updates to your file transfer
protocol-enabled server or when you work with users, use the
service-generated \code{ServerId} property that is assigned to the newly
created server.
}
\section{Request syntax}{
\preformatted{svc$create_server(
  Certificate = "string",
  Domain = "S3"|"EFS",
  EndpointDetails = list(
    AddressAllocationIds = list(
      "string"
    ),
    SubnetIds = list(
      "string"
    ),
    VpcEndpointId = "string",
    VpcId = "string",
    SecurityGroupIds = list(
      "string"
    )
  ),
  EndpointType = "PUBLIC"|"VPC"|"VPC_ENDPOINT",
  HostKey = "string",
  IdentityProviderDetails = list(
    Url = "string",
    InvocationRole = "string"
  ),
  IdentityProviderType = "SERVICE_MANAGED"|"API_GATEWAY",
  LoggingRole = "string",
  Protocols = list(
    "SFTP"|"FTP"|"FTPS"
  ),
  SecurityPolicyName = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
