% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_delete_server}
\alias{transfer_delete_server}
\title{Deletes the Secure File Transfer Protocol (SFTP) server that you specify}
\usage{
transfer_delete_server(ServerId)
}
\arguments{
\item{ServerId}{[required] A unique system-assigned identifier for an SFTP server instance.}
}
\description{
Deletes the Secure File Transfer Protocol (SFTP) server that you
specify. If you used \code{SERVICE_MANAGED} as your \code{IdentityProviderType},
you need to delete all users associated with this server before deleting
the server itself
}
\details{
No response returns from this call.
}
\section{Request syntax}{
\preformatted{svc$delete_server(
  ServerId = "string"
)
}
}

\keyword{internal}
