% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_delete_user}
\alias{transfer_delete_user}
\title{Deletes the user belonging to the server you specify}
\usage{
transfer_delete_user(ServerId, UserName)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server instance that has
the user assigned to it.}

\item{UserName}{[required] A unique string that identifies a user that is being deleted from the
server.}
}
\description{
Deletes the user belonging to the server you specify.
}
\details{
No response returns from this operation.

When you delete a user from a server, the user\'s information is lost.
}
\section{Request syntax}{
\preformatted{svc$delete_user(
  ServerId = "string",
  UserName = "string"
)
}
}

\keyword{internal}
