% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_describe_security_policy}
\alias{transfer_describe_security_policy}
\title{Describes the security policy that is attached to your file transfer
protocol-enabled server}
\usage{
transfer_describe_security_policy(SecurityPolicyName)
}
\arguments{
\item{SecurityPolicyName}{[required] Specifies the name of the security policy that is attached to the
server.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SecurityPolicy = list(
    Fips = TRUE|FALSE,
    SecurityPolicyName = "string",
    SshCiphers = list(
      "string"
    ),
    SshKexs = list(
      "string"
    ),
    SshMacs = list(
      "string"
    ),
    TlsCiphers = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the security policy that is attached to your file transfer
protocol-enabled server. The response contains a description of the
security policy's properties. For more information about security
policies, see \href{https://docs.aws.amazon.com/transfer/latest/userguide/security-policies.html}{Working with security policies}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_security_policy(
  SecurityPolicyName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
