% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_describe_server}
\alias{transfer_describe_server}
\title{Describes a file transfer protocol-enabled server that you specify by
passing the ServerId parameter}
\usage{
transfer_describe_server(ServerId)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for a server.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Server = list(
    Arn = "string",
    Certificate = "string",
    Domain = "S3"|"EFS",
    EndpointDetails = list(
      AddressAllocationIds = list(
        "string"
      ),
      SubnetIds = list(
        "string"
      ),
      VpcEndpointId = "string",
      VpcId = "string",
      SecurityGroupIds = list(
        "string"
      )
    ),
    EndpointType = "PUBLIC"|"VPC"|"VPC_ENDPOINT",
    HostKeyFingerprint = "string",
    IdentityProviderDetails = list(
      Url = "string",
      InvocationRole = "string"
    ),
    IdentityProviderType = "SERVICE_MANAGED"|"API_GATEWAY",
    LoggingRole = "string",
    Protocols = list(
      "SFTP"|"FTP"|"FTPS"
    ),
    SecurityPolicyName = "string",
    ServerId = "string",
    State = "OFFLINE"|"ONLINE"|"STARTING"|"STOPPING"|"START_FAILED"|"STOP_FAILED",
    Tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    UserCount = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a file transfer protocol-enabled server that you specify by
passing the \code{ServerId} parameter.

The response contains a description of a server's properties. When you
set \code{EndpointType} to VPC, the response will contain the
\code{EndpointDetails}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_server(
  ServerId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
