% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_describe_user}
\alias{transfer_describe_user}
\title{Describes the user assigned to the specific file transfer
protocol-enabled server, as identified by its ServerId property}
\usage{
transfer_describe_user(ServerId, UserName)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for a server that has this user
assigned.}

\item{UserName}{[required] The name of the user assigned to one or more servers. User names are
part of the sign-in credentials to use the AWS Transfer Family service
and perform file transfer tasks.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ServerId = "string",
  User = list(
    Arn = "string",
    HomeDirectory = "string",
    HomeDirectoryMappings = list(
      list(
        Entry = "string",
        Target = "string"
      )
    ),
    HomeDirectoryType = "PATH"|"LOGICAL",
    Policy = "string",
    PosixProfile = list(
      Uid = 123,
      Gid = 123,
      SecondaryGids = list(
        123
      )
    ),
    Role = "string",
    SshPublicKeys = list(
      list(
        DateImported = as.POSIXct(
          "2015-01-01"
        ),
        SshPublicKeyBody = "string",
        SshPublicKeyId = "string"
      )
    ),
    Tags = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    UserName = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the user assigned to the specific file transfer
protocol-enabled server, as identified by its \code{ServerId} property.

The response from this call returns the properties of the user
associated with the \code{ServerId} value that was specified.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_user(
  ServerId = "string",
  UserName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
