% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_import_ssh_public_key}
\alias{transfer_import_ssh_public_key}
\title{Adds a Secure Shell (SSH) public key to a user account identified by a
UserName value assigned to the specific file transfer protocol-enabled
server, identified by ServerId}
\usage{
transfer_import_ssh_public_key(ServerId, SshPublicKeyBody, UserName)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for a server.}

\item{SshPublicKeyBody}{[required] The public key portion of an SSH key pair.}

\item{UserName}{[required] The name of the user account that is assigned to one or more servers.}
}
\description{
Adds a Secure Shell (SSH) public key to a user account identified by a
\code{UserName} value assigned to the specific file transfer protocol-enabled
server, identified by \code{ServerId}.

The response returns the \code{UserName} value, the \code{ServerId} value, and the
name of the \code{SshPublicKeyId}.
}
\section{Request syntax}{
\preformatted{svc$import_ssh_public_key(
  ServerId = "string",
  SshPublicKeyBody = "string",
  UserName = "string"
)
}
}

\keyword{internal}
