% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_list_security_policies}
\alias{transfer_list_security_policies}
\title{Lists the security policies that are attached to your file transfer
protocol-enabled servers}
\usage{
transfer_list_security_policies(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{Specifies the number of security policies to return as a response to the
\code{\link[=transfer_list_security_policies]{list_security_policies}} query.}

\item{NextToken}{When additional results are obtained from the
\code{\link[=transfer_list_security_policies]{list_security_policies}} command, a
\code{NextToken} parameter is returned in the output. You can then pass the
\code{NextToken} parameter in a subsequent command to continue listing
additional security policies.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NextToken = "string",
  SecurityPolicyNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the security policies that are attached to your file transfer
protocol-enabled servers.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_security_policies(
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
