% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_list_users}
\alias{transfer_list_users}
\title{Lists the users for a file transfer protocol-enabled server that you
specify by passing the ServerId parameter}
\usage{
transfer_list_users(MaxResults, NextToken, ServerId)
}
\arguments{
\item{MaxResults}{Specifies the number of users to return as a response to the
\code{\link[=transfer_list_users]{list_users}} request.}

\item{NextToken}{When you can get additional results from the
\code{\link[=transfer_list_users]{list_users}} call, a \code{NextToken} parameter is
returned in the output. You can then pass in a subsequent command to the
\code{NextToken} parameter to continue listing additional users.}

\item{ServerId}{[required] A system-assigned unique identifier for a server that has users assigned
to it.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NextToken = "string",
  ServerId = "string",
  Users = list(
    list(
      Arn = "string",
      HomeDirectory = "string",
      HomeDirectoryType = "PATH"|"LOGICAL",
      Role = "string",
      SshPublicKeyCount = 123,
      UserName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the users for a file transfer protocol-enabled server that you
specify by passing the \code{ServerId} parameter.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_users(
  MaxResults = 123,
  NextToken = "string",
  ServerId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
