% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_untag_resource}
\alias{transfer_untag_resource}
\title{Detaches a key-value pair from a resource, as identified by its Amazon
Resource Name (ARN)}
\usage{
transfer_untag_resource(Arn, TagKeys)
}
\arguments{
\item{Arn}{[required] The value of the resource that will have the tag removed. An Amazon
Resource Name (ARN) is an identifier for a specific AWS resource, such
as a server, user, or role.}

\item{TagKeys}{[required] TagKeys are key-value pairs assigned to ARNs that can be used to group
and search for resources by type. This metadata can be attached to
resources for any purpose.}
}
\description{
Detaches a key-value pair from a resource, as identified by its Amazon
Resource Name (ARN). Resources are users, servers, roles, and other
entities.
}
\details{
No response is returned from this call.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  Arn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
