% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_update_server}
\alias{transfer_update_server}
\title{Updates the server properties after that server has been created}
\usage{
transfer_update_server(EndpointDetails, EndpointType, HostKey,
  IdentityProviderDetails, LoggingRole, ServerId)
}
\arguments{
\item{EndpointDetails}{The virtual private cloud (VPC) endpoint settings that are configured
for your SFTP server. With a VPC endpoint, your SFTP server isn\'t
accessible over the public internet.}

\item{EndpointType}{The type of endpoint that you want your SFTP server to connect to. You
can choose to connect to the public internet or a virtual private cloud
(VPC) endpoint. With a VPC endpoint, your SFTP server isn\'t accessible
over the public internet.}

\item{HostKey}{The RSA private key as generated by
\verb{ssh-keygen -N "" -f my-new-server-key}.

If you aren\'t planning to migrate existing users from an existing SFTP
server to a new AWS SFTP server, don\'t update the host key.
Accidentally changing a server\'s host key can be disruptive.

For more information, see
\"https://docs.aws.amazon.com/transfer/latest/userguide/configuring-servers.html\\#change-host-key\"
in the \emph{AWS SFTP User Guide.}}

\item{IdentityProviderDetails}{This response parameter is an array containing all of the information
required to call a customer\'s authentication API method.}

\item{LoggingRole}{A value that changes the AWS Identity and Access Management (IAM) role
that allows Amazon S3 events to be logged in Amazon CloudWatch, turning
logging on or off.}

\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server instance that the
user account is assigned to.}
}
\description{
Updates the server properties after that server has been created.
}
\details{
The \code{UpdateServer} call returns the \code{ServerId} of the Secure File
Transfer Protocol (SFTP) server you updated.
}
\section{Request syntax}{
\preformatted{svc$update_server(
  EndpointDetails = list(
    VpcEndpointId = "string"
  ),
  EndpointType = "PUBLIC"|"VPC_ENDPOINT",
  HostKey = "string",
  IdentityProviderDetails = list(
    Url = "string",
    InvocationRole = "string"
  ),
  LoggingRole = "string",
  ServerId = "string"
)
}
}

\keyword{internal}
