% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_batch_delete_import_data}
\alias{applicationdiscoveryservice_batch_delete_import_data}
\title{Deletes one or more import tasks, each identified by their import ID}
\usage{
applicationdiscoveryservice_batch_delete_import_data(importTaskIds)
}
\arguments{
\item{importTaskIds}{[required] The IDs for the import tasks that you want to delete.}
}
\value{
A list with the following syntax:\preformatted{list(
  errors = list(
    list(
      importTaskId = "string",
      errorCode = "NOT_FOUND"|"INTERNAL_SERVER_ERROR"|"OVER_LIMIT",
      errorDescription = "string"
    )
  )
)
}
}
\description{
Deletes one or more import tasks, each identified by their import ID.
Each import task has a number of records that can identify servers or
applications.

AWS Application Discovery Service has built-in matching logic that will
identify when discovered servers match existing entries that you've
previously discovered, the information for the already-existing
discovered server is updated. When you delete an import task that
contains records that were used to match, the information in those
matched records that comes from the deleted records will also be
deleted.
}
\section{Request syntax}{
\preformatted{svc$batch_delete_import_data(
  importTaskIds = list(
    "string"
  )
)
}
}

\keyword{internal}
