% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_create_application}
\alias{applicationdiscoveryservice_create_application}
\title{Creates an application with the given name and description}
\usage{
applicationdiscoveryservice_create_application(name, description)
}
\arguments{
\item{name}{[required] Name of the application to be created.}

\item{description}{Description of the application to be created.}
}
\value{
A list with the following syntax:\preformatted{list(
  configurationId = "string"
)
}
}
\description{
Creates an application with the given name and description.
}
\section{Request syntax}{
\preformatted{svc$create_application(
  name = "string",
  description = "string"
)
}
}

\keyword{internal}
