% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_delete_applications}
\alias{applicationdiscoveryservice_delete_applications}
\title{Deletes a list of applications and their associations with configuration
items}
\usage{
applicationdiscoveryservice_delete_applications(configurationIds)
}
\arguments{
\item{configurationIds}{[required] Configuration ID of an application to be deleted.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a list of applications and their associations with configuration
items.
}
\section{Request syntax}{
\preformatted{svc$delete_applications(
  configurationIds = list(
    "string"
  )
)
}
}

\keyword{internal}
