% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_describe_continuous_exports}
\alias{applicationdiscoveryservice_describe_continuous_exports}
\title{Lists exports as specified by ID}
\usage{
applicationdiscoveryservice_describe_continuous_exports(exportIds,
  maxResults, nextToken)
}
\arguments{
\item{exportIds}{The unique IDs assigned to the exports.}

\item{maxResults}{A number between 1 and 100 specifying the maximum number of continuous
export descriptions returned.}

\item{nextToken}{The token from the previous call to
\code{\link[=applicationdiscoveryservice_describe_export_tasks]{describe_export_tasks}}.}
}
\description{
Lists exports as specified by ID. All continuous exports associated with
your user account can be listed if you call
\code{\link[=applicationdiscoveryservice_describe_continuous_exports]{describe_continuous_exports}}
as is without passing any parameters.
}
\section{Request syntax}{
\preformatted{svc$describe_continuous_exports(
  exportIds = list(
    "string"
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
