% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationdiscoveryservice_operations.R
\name{applicationdiscoveryservice_get_discovery_summary}
\alias{applicationdiscoveryservice_get_discovery_summary}
\title{Retrieves a short summary of discovered assets}
\usage{
applicationdiscoveryservice_get_discovery_summary()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  servers = 123,
  applications = 123,
  serversMappedToApplications = 123,
  serversMappedtoTags = 123,
  agentSummary = list(
    activeAgents = 123,
    healthyAgents = 123,
    blackListedAgents = 123,
    shutdownAgents = 123,
    unhealthyAgents = 123,
    totalAgents = 123,
    unknownAgents = 123
  ),
  connectorSummary = list(
    activeConnectors = 123,
    healthyConnectors = 123,
    blackListedConnectors = 123,
    shutdownConnectors = 123,
    unhealthyConnectors = 123,
    totalConnectors = 123,
    unknownConnectors = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves a short summary of discovered assets.

This API operation takes no request parameters and is called as is at
the command prompt as shown in the example.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_discovery_summary()
}\if{html}{\out{</div>}}
}

\keyword{internal}
