% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_apply_pending_maintenance_action}
\alias{databasemigrationservice_apply_pending_maintenance_action}
\title{Applies a pending maintenance action to a resource (for example, to a
replication instance)}
\usage{
databasemigrationservice_apply_pending_maintenance_action(
  ReplicationInstanceArn, ApplyAction, OptInType)
}
\arguments{
\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the AWS DMS resource that the pending
maintenance action applies to.}

\item{ApplyAction}{[required] The pending maintenance action to apply to this resource.}

\item{OptInType}{[required] A value that specifies the type of opt-in request, or undoes an opt-in
request. You can't undo an opt-in request of type \code{immediate}.

Valid values:
\itemize{
\item \code{immediate} - Apply the maintenance action immediately.
\item \code{next-maintenance} - Apply the maintenance action during the next
maintenance window for the resource.
\item \code{undo-opt-in} - Cancel any existing \code{next-maintenance} opt-in
requests.
}}
}
\description{
Applies a pending maintenance action to a resource (for example, to a
replication instance).
}
\section{Request syntax}{
\preformatted{svc$apply_pending_maintenance_action(
  ReplicationInstanceArn = "string",
  ApplyAction = "string",
  OptInType = "string"
)
}
}

\keyword{internal}
