% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_create_replication_task}
\alias{databasemigrationservice_create_replication_task}
\title{Creates a replication task using the specified parameters}
\usage{
databasemigrationservice_create_replication_task(
  ReplicationTaskIdentifier, SourceEndpointArn, TargetEndpointArn,
  ReplicationInstanceArn, MigrationType, TableMappings,
  ReplicationTaskSettings, CdcStartTime, CdcStartPosition,
  CdcStopPosition, Tags)
}
\arguments{
\item{ReplicationTaskIdentifier}{[required] An identifier for the replication task.

Constraints:
\itemize{
\item Must contain from 1 to 255 alphanumeric characters or hyphens.
\item First character must be a letter.
\item Cannot end with a hyphen or contain two consecutive hyphens.
}}

\item{SourceEndpointArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies the source
endpoint.}

\item{TargetEndpointArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies the target
endpoint.}

\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of a replication instance.}

\item{MigrationType}{[required] The migration type. Valid values: \code{full-load} \\| \code{cdc} \\|
\code{full-load-and-cdc}}

\item{TableMappings}{[required] The table mappings for the task, in JSON format. For more information,
see \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Tasks.CustomizingTasks.TableMapping.html}{Table Mapping}
in the \emph{AWS Database Migration User Guide.}}

\item{ReplicationTaskSettings}{Overall settings for the task, in JSON format. For more information, see
\href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Tasks.CustomizingTasks.TaskSettings.html}{Task Settings}
in the \emph{AWS Database Migration User Guide.}}

\item{CdcStartTime}{Indicates the start time for a change data capture (CDC) operation. Use
either CdcStartTime or CdcStartPosition to specify when you want a CDC
operation to start. Specifying both values results in an error.

Timestamp Example: \\--cdc-start-time "2018-03-08T12:12:12"}

\item{CdcStartPosition}{Indicates when you want a change data capture (CDC) operation to start.
Use either CdcStartPosition or CdcStartTime to specify when you want a
CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: \\--cdc-start-position "2018-03-08T12:12:12"

Checkpoint Example: \\--cdc-start-position
\"checkpoint:V1\\#27\\#mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:1876\\#0\\#0\\#\*\\#0\\#93\"

LSN Example: \\--cdc-start-position "mysql-bin-changelog.000024:373"

When you use this task setting with a source PostgreSQL database, a
logical replication slot should already be created and associated with
the source endpoint. You can verify this by setting the \code{slotName} extra
connection attribute to the name of this logical replication slot. For
more information, see \href{https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Source.PostgreSQL.html#CHAP_Source.PostgreSQL.ConnectionAttrib}{Extra Connection Attributes When Using PostgreSQL as a Source for AWS DMS}.}

\item{CdcStopPosition}{Indicates when you want a change data capture (CDC) operation to stop.
The value can be either server time or commit time.

Server time example: \\--cdc-stop-position
"server\\_time:3018-02-09T12:12:12"

Commit time example: \\--cdc-stop-position "commit\\_time:
3018-02-09T12:12:12 "}

\item{Tags}{One or more tags to be assigned to the replication task.}
}
\description{
Creates a replication task using the specified parameters.
}
\section{Request syntax}{
\preformatted{svc$create_replication_task(
  ReplicationTaskIdentifier = "string",
  SourceEndpointArn = "string",
  TargetEndpointArn = "string",
  ReplicationInstanceArn = "string",
  MigrationType = "full-load"|"cdc"|"full-load-and-cdc",
  TableMappings = "string",
  ReplicationTaskSettings = "string",
  CdcStartTime = as.POSIXct(
    "2015-01-01"
  ),
  CdcStartPosition = "string",
  CdcStopPosition = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\examples{
# Creates a replication task using the specified parameters.
\donttest{svc$create_replication_task(
  CdcStartTime = "2016-12-14T18:25:43Z",
  MigrationType = "full-load",
  ReplicationInstanceArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJ...",
  ReplicationTaskIdentifier = "task1",
  ReplicationTaskSettings = "",
  SourceEndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ZW5UAN6P4E77EC7YWHK4RZZ3B...",
  TableMappings = "file://mappingfile.json",
  Tags = list(
    list(
      Key = "Acount",
      Value = "24352226"
    )
  ),
  TargetEndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ7..."
)}

}
\keyword{internal}
