% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_certificate}
\alias{databasemigrationservice_delete_certificate}
\title{Deletes the specified certificate}
\usage{
databasemigrationservice_delete_certificate(CertificateArn)
}
\arguments{
\item{CertificateArn}{[required] The Amazon Resource Name (ARN) of the deleted certificate.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Certificate = list(
    CertificateIdentifier = "string",
    CertificateCreationDate = as.POSIXct(
      "2015-01-01"
    ),
    CertificatePem = "string",
    CertificateWallet = raw,
    CertificateArn = "string",
    CertificateOwner = "string",
    ValidFromDate = as.POSIXct(
      "2015-01-01"
    ),
    ValidToDate = as.POSIXct(
      "2015-01-01"
    ),
    SigningAlgorithm = "string",
    KeyLength = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified certificate.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_certificate(
  CertificateArn = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Deletes the specified certificate.
svc$delete_certificate(
  CertificateArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUSM457DE6XFJCJQ"
)
}

}
\keyword{internal}
