% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_event_subscription}
\alias{databasemigrationservice_delete_event_subscription}
\title{Deletes an AWS DMS event subscription}
\usage{
databasemigrationservice_delete_event_subscription(SubscriptionName)
}
\arguments{
\item{SubscriptionName}{[required] The name of the DMS event notification subscription to be deleted.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EventSubscription = list(
    CustomerAwsId = "string",
    CustSubscriptionId = "string",
    SnsTopicArn = "string",
    Status = "string",
    SubscriptionCreationTime = "string",
    SourceType = "string",
    SourceIdsList = list(
      "string"
    ),
    EventCategoriesList = list(
      "string"
    ),
    Enabled = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes an AWS DMS event subscription.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_event_subscription(
  SubscriptionName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
