% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_replication_instance}
\alias{databasemigrationservice_delete_replication_instance}
\title{Deletes the specified replication instance}
\usage{
databasemigrationservice_delete_replication_instance(
  ReplicationInstanceArn)
}
\arguments{
\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the replication instance to be
deleted.}
}
\description{
Deletes the specified replication instance.
}
\details{
You must delete any migration tasks that are associated with the
replication instance before you can delete it.
}
\section{Request syntax}{
\preformatted{svc$delete_replication_instance(
  ReplicationInstanceArn = "string"
)
}
}

\section{Examples}{
\preformatted{# Deletes the specified replication instance. You must delete any
# migration tasks that are associated with the replication instance before
# you can delete it.
# 
# 
svc$delete_replication_instance(
  ReplicationInstanceArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJ..."
)
}
}

\keyword{internal}
