% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_replication_task}
\alias{databasemigrationservice_delete_replication_task}
\title{Deletes the specified replication task}
\usage{
databasemigrationservice_delete_replication_task(ReplicationTaskArn)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name (ARN) of the replication task to be deleted.}
}
\description{
Deletes the specified replication task.
}
\section{Request syntax}{
\preformatted{svc$delete_replication_task(
  ReplicationTaskArn = "string"
)
}
}

\section{Examples}{
\preformatted{# Deletes the specified replication task.
svc$delete_replication_task(
  ReplicationTaskArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJQ"
)
}
}

\keyword{internal}
