% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_refresh_schemas_status}
\alias{databasemigrationservice_describe_refresh_schemas_status}
\title{Returns the status of the RefreshSchemas operation}
\usage{
databasemigrationservice_describe_refresh_schemas_status(EndpointArn)
}
\arguments{
\item{EndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the
endpoint.}
}
\description{
Returns the status of the RefreshSchemas operation.
}
\section{Request syntax}{
\preformatted{svc$describe_refresh_schemas_status(
  EndpointArn = "string"
)
}
}

\examples{
# Returns the status of the refresh-schemas operation.
\donttest{svc$describe_refresh_schemas_status(
  EndpointArn = ""
)}

}
\keyword{internal}
