% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_reboot_replication_instance}
\alias{databasemigrationservice_reboot_replication_instance}
\title{Reboots a replication instance}
\usage{
databasemigrationservice_reboot_replication_instance(
  ReplicationInstanceArn, ForceFailover)
}
\arguments{
\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the replication instance.}

\item{ForceFailover}{If this parameter is \code{true}, the reboot is conducted through a Multi-AZ
failover. (If the instance isn't configured for Multi-AZ, then you can't
specify \code{true}.)}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ReplicationInstance = list(
    ReplicationInstanceIdentifier = "string",
    ReplicationInstanceClass = "string",
    ReplicationInstanceStatus = "string",
    AllocatedStorage = 123,
    InstanceCreateTime = as.POSIXct(
      "2015-01-01"
    ),
    VpcSecurityGroups = list(
      list(
        VpcSecurityGroupId = "string",
        Status = "string"
      )
    ),
    AvailabilityZone = "string",
    ReplicationSubnetGroup = list(
      ReplicationSubnetGroupIdentifier = "string",
      ReplicationSubnetGroupDescription = "string",
      VpcId = "string",
      SubnetGroupStatus = "string",
      Subnets = list(
        list(
          SubnetIdentifier = "string",
          SubnetAvailabilityZone = list(
            Name = "string"
          ),
          SubnetStatus = "string"
        )
      )
    ),
    PreferredMaintenanceWindow = "string",
    PendingModifiedValues = list(
      ReplicationInstanceClass = "string",
      AllocatedStorage = 123,
      MultiAZ = TRUE|FALSE,
      EngineVersion = "string"
    ),
    MultiAZ = TRUE|FALSE,
    EngineVersion = "string",
    AutoMinorVersionUpgrade = TRUE|FALSE,
    KmsKeyId = "string",
    ReplicationInstanceArn = "string",
    ReplicationInstancePublicIpAddress = "string",
    ReplicationInstancePrivateIpAddress = "string",
    ReplicationInstancePublicIpAddresses = list(
      "string"
    ),
    ReplicationInstancePrivateIpAddresses = list(
      "string"
    ),
    PubliclyAccessible = TRUE|FALSE,
    SecondaryAvailabilityZone = "string",
    FreeUntil = as.POSIXct(
      "2015-01-01"
    ),
    DnsNameServers = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Reboots a replication instance. Rebooting results in a momentary outage,
until the replication instance becomes available again.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$reboot_replication_instance(
  ReplicationInstanceArn = "string",
  ForceFailover = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
