% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_test_connection}
\alias{databasemigrationservice_test_connection}
\title{Tests the connection between the replication instance and the endpoint}
\usage{
databasemigrationservice_test_connection(ReplicationInstanceArn,
  EndpointArn)
}
\arguments{
\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the replication instance.}

\item{EndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the
endpoint.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Connection = list(
    ReplicationInstanceArn = "string",
    EndpointArn = "string",
    Status = "string",
    LastFailureMessage = "string",
    EndpointIdentifier = "string",
    ReplicationInstanceIdentifier = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Tests the connection between the replication instance and the endpoint.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$test_connection(
  ReplicationInstanceArn = "string",
  EndpointArn = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# Tests the connection between the replication instance and the endpoint.
svc$test_connection(
  EndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:RAAR3R22XSH46S3PWLC3NJAWKM",
  ReplicationInstanceArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJQ"
)
}

}
\keyword{internal}
