% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_create_location_s3}
\alias{datasync_create_location_s3}
\title{Creates an endpoint for an Amazon S3 bucket}
\usage{
datasync_create_location_s3(Subdirectory, S3BucketArn, S3StorageClass,
  S3Config, AgentArns, Tags)
}
\arguments{
\item{Subdirectory}{A subdirectory in the Amazon S3 bucket. This subdirectory in Amazon S3
is used to read data from the S3 source location or write data to the S3
destination.}

\item{S3BucketArn}{[required] The ARN of the Amazon S3 bucket. If the bucket is on an AWS Outpost,
this must be an access point ARN.}

\item{S3StorageClass}{The Amazon S3 storage class that you want to store your files in when
this location is used as a task destination. For buckets in AWS Regions,
the storage class defaults to Standard. For buckets on AWS Outposts, the
storage class defaults to AWS S3 Outposts.

For more information about S3 storage classes, see \href{https://aws.amazon.com/s3/storage-classes/}{Amazon S3 Storage Classes}. Some storage
classes have behaviors that can affect your S3 storage cost. For
detailed information, see using-storage-classes.}

\item{S3Config}{[required]}

\item{AgentArns}{If you are using DataSync on an AWS Outpost, specify the Amazon Resource
Names (ARNs) of the DataSync agents deployed on your Outpost. For more
information about launching a DataSync agent on an AWS Outpost, see
outposts-agent.}

\item{Tags}{The key-value pair that represents the tag that you want to add to the
location. The value can be an empty string. We recommend using tags to
name your resources.}
}
\description{
Creates an endpoint for an Amazon S3 bucket.
}
\details{
For more information, see
https://docs.aws.amazon.com/datasync/latest/userguide/create-locations-cli.html\\#create-location-s3-cli
in the \emph{AWS DataSync User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_location_s3(
  Subdirectory = "string",
  S3BucketArn = "string",
  S3StorageClass = "STANDARD"|"STANDARD_IA"|"ONEZONE_IA"|"INTELLIGENT_TIERING"|"GLACIER"|"DEEP_ARCHIVE"|"OUTPOSTS",
  S3Config = list(
    BucketAccessRoleArn = "string"
  ),
  AgentArns = list(
    "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
