% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_location_fsx_windows}
\alias{datasync_describe_location_fsx_windows}
\title{Returns metadata, such as the path information about an Amazon FSx for
Windows File Server location}
\usage{
datasync_describe_location_fsx_windows(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the FSx for Windows File Server
location to describe.}
}
\value{
A list with the following syntax:\preformatted{list(
  LocationArn = "string",
  LocationUri = "string",
  SecurityGroupArns = list(
    "string"
  ),
  CreationTime = as.POSIXct(
    "2015-01-01"
  ),
  User = "string",
  Domain = "string"
)
}
}
\description{
Returns metadata, such as the path information about an Amazon FSx for
Windows File Server location.
}
\section{Request syntax}{
\preformatted{svc$describe_location_fsx_windows(
  LocationArn = "string"
)
}
}

\keyword{internal}
