% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_location_object_storage}
\alias{datasync_describe_location_object_storage}
\title{Returns metadata about a self-managed object storage server location}
\usage{
datasync_describe_location_object_storage(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the self-managed object storage server
location that was described.}
}
\description{
Returns metadata about a self-managed object storage server location.
For more information about self-managed object storage locations, see
create-object-location.
}
\section{Request syntax}{
\preformatted{svc$describe_location_object_storage(
  LocationArn = "string"
)
}
}

\keyword{internal}
