% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_location_s3}
\alias{datasync_describe_location_s3}
\title{Returns metadata, such as bucket name, about an Amazon S3 bucket
location}
\usage{
datasync_describe_location_s3(LocationArn)
}
\arguments{
\item{LocationArn}{[required] The Amazon Resource Name (ARN) of the Amazon S3 bucket location to
describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  LocationArn = "string",
  LocationUri = "string",
  S3StorageClass = "STANDARD"|"STANDARD_IA"|"ONEZONE_IA"|"INTELLIGENT_TIERING"|"GLACIER"|"DEEP_ARCHIVE"|"OUTPOSTS",
  S3Config = list(
    BucketAccessRoleArn = "string"
  ),
  AgentArns = list(
    "string"
  ),
  CreationTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns metadata, such as bucket name, about an Amazon S3 bucket
location.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_location_s3(
  LocationArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
