% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrationhub_operations.R
\name{migrationhub_describe_application_state}
\alias{migrationhub_describe_application_state}
\title{Gets the migration status of an application}
\usage{
migrationhub_describe_application_state(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The configurationId in ADS that uniquely identifies the grouped
application.}
}
\description{
Gets the migration status of an application.
}
\section{Request syntax}{
\preformatted{svc$describe_application_state(
  ApplicationId = "string"
)
}
}

\keyword{internal}
