% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_delete_server_catalog}
\alias{sms_delete_server_catalog}
\title{Deletes all servers from your server catalog}
\usage{
sms_delete_server_catalog()
}
\description{
Deletes all servers from your server catalog.
}
\section{Request syntax}{
\preformatted{svc$delete_server_catalog()
}
}

\keyword{internal}
