% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_generate_template}
\alias{sms_generate_template}
\title{Generates an AWS CloudFormation template based on the current launch
configuration and writes it to an Amazon S3 object in the customer’s
Amazon S3 bucket}
\usage{
sms_generate_template(appId, templateFormat)
}
\arguments{
\item{appId}{The ID of the application associated with the AWS CloudFormation
template.}

\item{templateFormat}{The format for generating the AWS CloudFormation template.}
}
\value{
A list with the following syntax:\preformatted{list(
  s3Location = list(
    bucket = "string",
    key = "string"
  )
)
}
}
\description{
Generates an AWS CloudFormation template based on the current launch
configuration and writes it to an Amazon S3 object in the customer’s
Amazon S3 bucket.
}
\section{Request syntax}{
\preformatted{svc$generate_template(
  appId = "string",
  templateFormat = "JSON"|"YAML"
)
}
}

\keyword{internal}
