% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_app_replication_configuration}
\alias{sms_get_app_replication_configuration}
\title{Retrieves the application replication configuration associated with the
specified application}
\usage{
sms_get_app_replication_configuration(appId)
}
\arguments{
\item{appId}{The ID of the application.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  serverGroupReplicationConfigurations = list(
    list(
      serverGroupId = "string",
      serverReplicationConfigurations = list(
        list(
          server = list(
            serverId = "string",
            serverType = "VIRTUAL_MACHINE",
            vmServer = list(
              vmServerAddress = list(
                vmManagerId = "string",
                vmId = "string"
              ),
              vmName = "string",
              vmManagerName = "string",
              vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
              vmPath = "string"
            ),
            replicationJobId = "string",
            replicationJobTerminated = TRUE|FALSE
          ),
          serverReplicationParameters = list(
            seedTime = as.POSIXct(
              "2015-01-01"
            ),
            frequency = 123,
            runOnce = TRUE|FALSE,
            licenseType = "AWS"|"BYOL",
            numberOfRecentAmisToKeep = 123,
            encrypted = TRUE|FALSE,
            kmsKeyId = "string"
          )
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the application replication configuration associated with the
specified application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_app_replication_configuration(
  appId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
