% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_start_on_demand_app_replication}
\alias{sms_start_on_demand_app_replication}
\title{Starts an on-demand replication run for the specified application}
\usage{
sms_start_on_demand_app_replication(appId, description)
}
\arguments{
\item{appId}{[required] The ID of the application.}

\item{description}{The description of the replication run.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Starts an on-demand replication run for the specified application.
}
\section{Request syntax}{
\preformatted{svc$start_on_demand_app_replication(
  appId = "string",
  description = "string"
)
}
}

\keyword{internal}
