% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_cancel_job}
\alias{snowball_cancel_job}
\title{Cancels the specified job}
\usage{
snowball_cancel_job(JobId)
}
\arguments{
\item{JobId}{[required] The 39-character job ID for the job that you want to cancel, for example
\verb{JID123e4567-e89b-12d3-a456-426655440000}.}
}
\description{
Cancels the specified job. You can only cancel a job before its
\code{JobState} value changes to \code{PreparingAppliance}. Requesting the
\code{ListJobs} or \code{DescribeJob} action returns a job\'s \code{JobState} as part
of the response element data returned.
}
\section{Request syntax}{
\preformatted{svc$cancel_job(
  JobId = "string"
)
}
}

\section{Examples}{
\preformatted{# This operation cancels a job. You can only cancel a job before its
# JobState value changes to PreparingAppliance.
svc$cancel_job(
  JobId = "JID123e4567-e89b-12d3-a456-426655440000"
)
}
}

\keyword{internal}
