% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowball_operations.R
\name{snowball_describe_return_shipping_label}
\alias{snowball_describe_return_shipping_label}
\title{Information on the shipping label of a Snow device that is being
returned to AWS}
\usage{
snowball_describe_return_shipping_label(JobId)
}
\arguments{
\item{JobId}{The automatically generated ID for a job, for example
\verb{JID123e4567-e89b-12d3-a456-426655440000}.}
}
\value{
A list with the following syntax:\preformatted{list(
  Status = "InProgress"|"TimedOut"|"Succeeded"|"Failed",
  ExpirationDate = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Information on the shipping label of a Snow device that is being
returned to AWS.
}
\section{Request syntax}{
\preformatted{svc$describe_return_shipping_label(
  JobId = "string"
)
}
}

\keyword{internal}
