% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_create_server}
\alias{transfer_create_server}
\title{Instantiates an autoscaling virtual server based on Secure File Transfer
Protocol (SFTP) in AWS}
\usage{
transfer_create_server(EndpointDetails, EndpointType, HostKey,
  IdentityProviderDetails, IdentityProviderType, LoggingRole, Tags)
}
\arguments{
\item{EndpointDetails}{The virtual private cloud (VPC) endpoint settings that you want to
configure for your SFTP server.}

\item{EndpointType}{The type of VPC endpoint that you want your SFTP server connect to. If
you connect to a VPC endpoint, your SFTP server isn't accessible over
the public internet.}

\item{HostKey}{The RSA private key as generated by
\code{ssh-keygen -N "" -f my-new-server-key} command.

If you aren't planning to migrate existing users from an existing SFTP
server to a new AWS SFTP server, don't update the host key.
Accidentally changing a server's host key can be disruptive. For more
information, see change-host-key in the \emph{AWS SFTP User Guide.}}

\item{IdentityProviderDetails}{An array containing all of the information required to call a
customer-supplied authentication API. This parameter is not required
when the \code{IdentityProviderType} value of server that is created uses the
\code{SERVICE_MANAGED} authentication method.}

\item{IdentityProviderType}{The mode of authentication enabled for this service. The default value
is \code{SERVICE_MANAGED}, which allows you to store and access SFTP user
credentials within the service. An \code{IdentityProviderType} value of
\code{API_GATEWAY} indicates that user authentication requires a call to an
API Gateway endpoint URL provided by you to integrate an identity
provider of your choice.}

\item{LoggingRole}{A value that allows the service to write your SFTP users' activity to
your Amazon CloudWatch logs for monitoring and auditing purposes.}

\item{Tags}{Key-value pairs that can be used to group and search for servers.}
}
\description{
Instantiates an autoscaling virtual server based on Secure File Transfer
Protocol (SFTP) in AWS. The call returns the \code{ServerId} property
assigned by the service to the newly created server. Reference this
\code{ServerId} property when you make updates to your server, or work with
users.
}
\details{
The response returns the \code{ServerId} value for the newly created server.
}
\section{Request syntax}{
\preformatted{svc$create_server(
  EndpointDetails = list(
    VpcEndpointId = "string"
  ),
  EndpointType = "PUBLIC"|"VPC_ENDPOINT",
  HostKey = "string",
  IdentityProviderDetails = list(
    Url = "string",
    InvocationRole = "string"
  ),
  IdentityProviderType = "SERVICE_MANAGED"|"API_GATEWAY",
  LoggingRole = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
