% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_delete_server}
\alias{transfer_delete_server}
\title{Deletes the file transfer protocol-enabled server that you specify}
\usage{
transfer_delete_server(ServerId)
}
\arguments{
\item{ServerId}{[required] A unique system-assigned identifier for a server instance.}
}
\value{
An empty list.
}
\description{
Deletes the file transfer protocol-enabled server that you specify.

No response returns from this operation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_server(
  ServerId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
