% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_describe_server}
\alias{transfer_describe_server}
\title{Describes a file transfer protocol-enabled server that you specify by
passing the ServerId parameter}
\usage{
transfer_describe_server(ServerId)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for a server.}
}
\description{
Describes a file transfer protocol-enabled server that you specify by
passing the \code{ServerId} parameter.

The response contains a description of a server's properties. When you
set \code{EndpointType} to VPC, the response will contain the
\code{EndpointDetails}.
}
\section{Request syntax}{
\preformatted{svc$describe_server(
  ServerId = "string"
)
}
}

\keyword{internal}
