% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_start_server}
\alias{transfer_start_server}
\title{Changes the state of a file transfer protocol-enabled server from
OFFLINE to ONLINE}
\usage{
transfer_start_server(ServerId)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for a server that you start.}
}
\value{
An empty list.
}
\description{
Changes the state of a file transfer protocol-enabled server from
\code{OFFLINE} to \code{ONLINE}. It has no impact on a server that is already
\code{ONLINE}. An \code{ONLINE} server can accept and process file transfer jobs.

The state of \code{STARTING} indicates that the server is in an intermediate
state, either not fully able to respond, or not fully online. The values
of \code{START_FAILED} can indicate an error condition.

No response is returned from this call.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_server(
  ServerId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
