% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_update_server}
\alias{transfer_update_server}
\title{Updates the file transfer protocol-enabled server's properties after
that server has been created}
\usage{
transfer_update_server(Certificate, EndpointDetails, EndpointType,
  HostKey, IdentityProviderDetails, LoggingRole, Protocols,
  SecurityPolicyName, ServerId)
}
\arguments{
\item{Certificate}{The Amazon Resource Name (ARN) of the AWS Certificate Manager (ACM)
certificate. Required when \code{Protocols} is set to \code{FTPS}.

To request a new public certificate, see \href{https://docs.aws.amazon.com/acm/latest/userguide/gs-acm-request-public.html}{Request a public certificate}
in the \emph{AWS Certificate Manager User Guide}.

To import an existing certificate into ACM, see \href{https://docs.aws.amazon.com/acm/latest/userguide/import-certificate.html}{Importing certificates into ACM}
in the \emph{AWS Certificate Manager User Guide}.

To request a private certificate to use FTPS through private IP
addresses, see \href{https://docs.aws.amazon.com/acm/latest/userguide/gs-acm-request-private.html}{Request a private certificate}
in the \emph{AWS Certificate Manager User Guide}.

Certificates with the following cryptographic algorithms and key sizes
are supported:
\itemize{
\item 2048-bit RSA (RSA\\_2048)
\item 4096-bit RSA (RSA\\_4096)
\item Elliptic Prime Curve 256 bit (EC\\_prime256v1)
\item Elliptic Prime Curve 384 bit (EC\\_secp384r1)
\item Elliptic Prime Curve 521 bit (EC\\_secp521r1)
}

The certificate must be a valid SSL/TLS X.509 version 3 certificate with
FQDN or IP address specified and information about the issuer.}

\item{EndpointDetails}{The virtual private cloud (VPC) endpoint settings that are configured
for your server. With a VPC endpoint, you can restrict access to your
server to resources only within your VPC. To control incoming internet
traffic, you will need to associate one or more Elastic IP addresses
with your server's endpoint.}

\item{EndpointType}{The type of endpoint that you want your server to connect to. You can
choose to connect to the public internet or a VPC endpoint. With a VPC
endpoint, you can restrict access to your server and resources only
within your VPC.

It is recommended that you use \code{VPC} as the \code{EndpointType}. With this
endpoint type, you have the option to directly associate up to three
Elastic IPv4 addresses (BYO IP included) with your server's endpoint and
use VPC security groups to restrict traffic by the client's public IP
address. This is not possible with \code{EndpointType} set to \code{VPC_ENDPOINT}.}

\item{HostKey}{The RSA private key as generated by
\verb{ssh-keygen -N "" -m PEM -f my-new-server-key}.

If you aren't planning to migrate existing users from an existing server
to a new server, don't update the host key. Accidentally changing a
server's host key can be disruptive.

For more information, see \href{https://docs.aws.amazon.com/transfer/latest/userguide/edit-server-config.html#configuring-servers-change-host-key}{Change the host key for your SFTP-enabled server}
in the \emph{AWS Transfer Family User Guide}.}

\item{IdentityProviderDetails}{An array containing all of the information required to call a customer's
authentication API method.}

\item{LoggingRole}{Changes the AWS Identity and Access Management (IAM) role that allows
Amazon S3 events to be logged in Amazon CloudWatch, turning logging on
or off.}

\item{Protocols}{Specifies the file transfer protocol or protocols over which your file
transfer protocol client can connect to your server's endpoint. The
available protocols are:
\itemize{
\item Secure Shell (SSH) File Transfer Protocol (SFTP): File transfer over
SSH
\item File Transfer Protocol Secure (FTPS): File transfer with TLS
encryption
\item File Transfer Protocol (FTP): Unencrypted file transfer
}

If you select \code{FTPS}, you must choose a certificate stored in AWS
Certificate Manager (ACM) which will be used to identify your server
when clients connect to it over FTPS.

If \code{Protocol} includes either \code{FTP} or \code{FTPS}, then the \code{EndpointType}
must be \code{VPC} and the \code{IdentityProviderType} must be \code{API_GATEWAY}.

If \code{Protocol} includes \code{FTP}, then \code{AddressAllocationIds} cannot be
associated.

If \code{Protocol} is set only to \code{SFTP}, the \code{EndpointType} can be set to
\code{PUBLIC} and the \code{IdentityProviderType} can be set to \code{SERVICE_MANAGED}.}

\item{SecurityPolicyName}{Specifies the name of the security policy that is attached to the
server.}

\item{ServerId}{[required] A system-assigned unique identifier for a server instance that the user
account is assigned to.}
}
\description{
Updates the file transfer protocol-enabled server's properties after
that server has been created.
}
\details{
The \code{UpdateServer} call returns the \code{ServerId} of the server you
updated.
}
\section{Request syntax}{
\preformatted{svc$update_server(
  Certificate = "string",
  EndpointDetails = list(
    AddressAllocationIds = list(
      "string"
    ),
    SubnetIds = list(
      "string"
    ),
    VpcEndpointId = "string",
    VpcId = "string",
    SecurityGroupIds = list(
      "string"
    )
  ),
  EndpointType = "PUBLIC"|"VPC"|"VPC_ENDPOINT",
  HostKey = "string",
  IdentityProviderDetails = list(
    Url = "string",
    InvocationRole = "string"
  ),
  LoggingRole = "string",
  Protocols = list(
    "SFTP"|"FTP"|"FTPS"
  ),
  SecurityPolicyName = "string",
  ServerId = "string"
)
}
}

\keyword{internal}
